% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnd.R
\name{abort}
\alias{abort}
\alias{warn}
\alias{inform}
\title{Signal an error, warning, or message}
\usage{
abort(.msg, .type = NULL, ..., .call = FALSE, msg, type, call)

warn(.msg, .type = NULL, ..., .call = FALSE, msg, type, call)

inform(.msg, .type = NULL, ..., .call = FALSE, msg, type, call)
}
\arguments{
\item{.msg}{The message to display.}

\item{.type}{Subclass of the condition.}

\item{...}{Additional data to be stored in the condition object.}

\item{.call}{Whether to display the call. If a number \code{n}, the call
is taken from the nth frame on the \link[=call_stack]{call stack}.}

\item{msg, type, call}{These arguments were renamed to be prefixed
with a dot and are deprecated as of rlang 0.3.0.}
}
\description{
These functions are equivalent to base functions \code{\link[base:stop]{base::stop()}},
\code{\link[base:warning]{base::warning()}} and \code{\link[base:message]{base::message()}}, but make it easy to supply
condition metadata:
\itemize{
\item Supply \code{type} to create a classed condition. Typed conditions can
be captured or handled selectively, allowing for finer-grained
error handling.
\item Supply metadata with named \code{...} arguments. This data will be
stored in the condition object and can be examined by handlers.
}
}
\section{Call trace}{


Unlike \code{stop()} and \code{warning()}, these functions don't include call
information by default. This saves you from typing \code{call. = FALSE}
and produces cleaner error messages.
}

\section{Mufflable conditions}{


Signalling a condition with \code{inform()} or \code{warn()} causes a message
to be displayed in the console. These messages can be muffled with
\code{\link[base:suppressMessages]{base::suppressMessages()}} or \code{\link[base:suppressWarnings]{base::suppressWarnings()}}.
}

\section{Lifecycle}{


These functions were changed in rlang 0.3.0 to take condition
metadata with \code{...}. Consequently:
\itemize{
\item All arguments were renamed to be prefixed with a dot.
\item \code{.call} (previously \code{call}) can no longer be passed positionally.
}
}

\examples{
# These examples are guarded to avoid throwing errors
if (FALSE) {

# Signal an error with a message just like stop():
abort("Something bad happened")

# Give a class to the error:
abort("Something bad happened", "somepkg_bad_error")

# This will allow your users to handle the error selectively
tryCatch(
  somepkg_function(),
  somepkg_bad_error = function(err) {
    cnd_warn(err) # Demote the error to a warning
    NA            # Return an alternative value
  }
)

# You can also specify metadata that will be stored in the condition:
abort("Something bad happened", "somepkg_bad_error", data = 1:10)

# This data can then be consulted by user handlers:
tryCatch(
  somepkg_function(),
  somepkg_bad_error = function(err) {
    # Compute an alternative return value with the data:
    recover_error(err$data)
  }
)

}
}
