% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg.R
\name{arg_exclusive}
\alias{arg_exclusive}
\title{Check that arguments are mutually exclusive}
\usage{
arg_exclusive(
  ...,
  .require = TRUE,
  .frame = caller_env(),
  .error_call = .frame
)
}
\arguments{
\item{...}{Function arguments.}

\item{.require}{Whether at least one argument must be supplied.}

\item{.frame}{Environment where the arguments in \code{...} are defined.}

\item{.error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[=abort]{abort()}} for more information.}
}
\description{
Check that arguments are mutually exclusive
}
