% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg.R
\name{arg_expr}
\alias{arg_expr}
\alias{arg_label}
\alias{arg_text}
\title{Find the expression associated with an argument}
\usage{
arg_expr(x)

arg_label(x)

arg_text(x, width = 60L, nlines = Inf)
}
\arguments{
\item{x}{Argument to capture.}

\item{width}{Width of each line.}

\item{nlines}{Maximum number of lines to extract.}
}
\description{
\code{arg_expr()} returns the full expression ans is equivalent to
the base function \code{\link[base]{substitute}()};
\code{arg_text()} turns the expression into a single string;
\code{arg_label()} formats it nicely for use in messages.
}
\examples{
arg_label(10)

# Names a quoted with ``
arg_label(x)

# Strings are encoded
arg_label("a\\nb")

# Expressions are captured
arg_label(a + b + c)

# Long expressions are collapsed
arg_label(foo({
  1 + 2
  print(x)
}))
}

