% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr.R
\name{arg_missing}
\alias{arg_missing}
\title{Generate a missing argument.}
\usage{
arg_missing()
}
\description{
Generate a missing argument.
}
\examples{
f_interp(~f(x = uq(arg_missing())))
f_interp(~f(x = uq(NULL)))
}

