% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg.R
\name{arg_require}
\alias{arg_require}
\title{Check that argument is supplied
Throws an informative error if \code{arg} is missing.}
\usage{
arg_require(arg, error_arg = caller_arg(arg), error_call = caller_env())
}
\arguments{
\item{arg}{A function argument. Must be a symbol.}

\item{error_arg}{An argument name as a string. This argument
will be mentioned in error messages as the input that is at the
origin of a problem.}

\item{error_call}{The execution environment of a currently
running function. When an error occurs, the corresponding
function call (see \code{\link[=sys.call]{sys.call()}}) is retrieved and stored as
\code{call} field in the error object to provide users with contextual
information about the error.

Can also be \code{NULL} or a function call to respectively disable the
contextual call or hard-code it.

See also \code{\link[=local_error_call]{local_error_call()}} for an alternative way of
providing this information.}
}
\description{
Check that argument is supplied
Throws an informative error if \code{arg} is missing.
}
\examples{
f <- function(x)  {
  arg_require(x)
}

# Fails because `x` is not supplied
try(f())

# Succeeds
f(NULL)
}
\seealso{
\code{\link[=arg_match]{arg_match()}}
}
