% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dots-ellipsis.R
\name{args_dots_empty}
\alias{args_dots_empty}
\title{Helper for consistent documentation of empty dots}
\arguments{
\item{...}{These dots are for future extensions and must be empty.}
}
\description{
Use \verb{@inheritParams rlang::args_dots_empty} in your package
to consistently document \code{...} that must be empty.
}
\keyword{internal}
