% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg.R
\name{args_error_call}
\alias{args_error_call}
\title{Documentation anchor for \code{error_call} arguments}
\arguments{
\item{error_call}{An expression representing the context in which
the error occurred. If non-null, \code{abort()} displays the call
(stripped from its arguments to keep it simple) before \code{message}.

If you wrap an input-checking helper in another similar helper,
you should generally forward \code{error_call} from the calling frame
so that the relevant user expression is displayed when an error
occurs.}
}
\description{
Use \verb{@inheritParams rlang::args_error_call} in your package
to document an \code{error_call} argument.
}
