% Generated by roxygen2 (4.0.1): do not edit by hand
\name{as.lazy}
\alias{as.lazy}
\title{Convert an object to a lazy object.}
\usage{
as.lazy(x, env)
}
\arguments{
\item{x}{An R object. Current methods convert formulas, character
vectors, calls and names.}

\item{env}{Environment to use for objects that don't already have
associated environment.}
}
\description{
Convert an object to a lazy object.
}
\examples{
as.lazy(~ x + 1)
as.lazy(quote(x + 1), globalenv())
as.lazy("x + 1", globalenv())
}

