% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn.R
\name{as_closure}
\alias{as_closure}
\title{Convert to closure.}
\usage{
as_closure(x)
}
\arguments{
\item{x}{A function or a string. In the latter case, the function
is looked up in the calling environment.}
}
\description{
Convert an object to a closure. This is especially useful to
normalise primitive functions to a proper closure (see
\code{\link[=is_function]{is_function()}} about primitive functions).
}
\examples{
as_closure(list)
}
