% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle-retired.R
\name{as_env}
\alias{as_env}
\title{Coerce to an environment}
\usage{
as_env(x, parent = NULL)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated as it was renamed to \code{\link[=as_environment]{as_environment()}}
in rlang 0.2.0.
}
\keyword{internal}
