% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deparse.R
\name{as_label}
\alias{as_label}
\title{Create a default name for an R object}
\usage{
as_label(x)
}
\arguments{
\item{x}{Any R object.}
}
\description{
\code{as_label()} transforms any R object into a single string that is
suitable as default name:
\itemize{
\item Quosures are \link[=quo_squash]{squashed} before being labelled.
\item Symbols are transformed to string with \code{as_string()}.
\item Calls are abbreviated.
\item Numbers are represented as such.
\item Other constants are represented by their type, such as \code{<dbl>}
or \code{<data.frame>}.
}

Please note that simple symbols should generally be transformed to
strings with \code{as_string()}. Labelisation is not a well defined
operation and no assumption should be made about how the label is
created. On the other hand, \code{as_string()} only works with symbols
and is a well defined, deterministic operation.

In other words, if you don't know for sure what kind of object
you're dealing with, use \code{as_label()} and make no assumption about the
resulting string. If you know the object is a symbol, use
\code{\link[=as_string]{as_string()}}.
}
