% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-sym.R
\name{as_name}
\alias{as_name}
\title{Coerce a symbol to a name}
\usage{
as_name(x)
}
\arguments{
\item{x}{A symbol.}
}
\value{
A string.
}
\description{
Coerce a symbol to a name
}
\examples{
sym <- quote(sym)
as_name(sym)

quo <- quo(sym)
as_name(get_expr(quo))
}
\seealso{
\code{\link[=sym]{sym()}} for the inverse operation.
}
