% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle-retired.R
\name{as_overscope}
\alias{as_overscope}
\alias{new_overscope}
\alias{overscope_clean}
\title{Create an overscope}
\usage{
as_overscope(quo, data = NULL)

new_overscope(bottom, top = NULL, enclosure = base_env())

overscope_clean(overscope)
}
\arguments{
\item{quo}{A \link{quosure}.}

\item{data}{A data frame or named vector of masking data.}

\item{bottom}{The environment containing masking objects if the
data mask is one environment deep. The bottom environment if the
data mask comprises multiple environment.}

\item{top}{The last environment of the data mask. If the data mask
is only one environment deep, \code{top} should be the same as
\code{bottom}.}

\item{enclosure}{The \code{parent} argument of \code{\link[=new_data_mask]{new_data_mask()}}.}

\item{overscope}{The \code{mask} argument of \code{\link[=data_mask_clean]{data_mask_clean()}}}
}
\description{
These functions have been soft-deprecated in rlang 0.2.0. Please use
\code{\link[=as_data_mask]{as_data_mask()}}, \code{\link[=new_data_mask]{new_data_mask()}} and \code{\link[=data_mask_clean]{data_mask_clean()}} instead.
}
\keyword{internal}
