% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{as_pairlist}
\alias{as_pairlist}
\title{Coerce to pairlist}
\usage{
as_pairlist(x)
}
\arguments{
\item{x}{An object to coerce.}
}
\description{
This transforms vector objects to a linked pairlist of nodes. See
the \link[=node]{pairlist} type help page.
}
\section{Life cycle}{


\code{as_pairlist()} is experimental because we are still figuring out
the naming scheme for pairlists and node-like objects.
}

\keyword{internal}
