% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lang.R
\name{as_symbol}
\alias{as_symbol}
\alias{as_name}
\alias{as_call}
\title{Coerce an object to a name or call.}
\usage{
as_symbol(x)

as_name(x)

as_call(x)
}
\arguments{
\item{x}{An object to coerce}
}
\value{
\code{as_symbol()} and \code{as_call()} return a symbol or
  a call. \code{as_name()} returns a string.
}
\description{
These coercing functions can transform names, calls, formulas, and
strings. The distinction between a name and a call is particularly
important when coercing from a string. Coercing to a call will
parse the string, coercing to a name will create a (potentially)
non-syntactic name.
}
\examples{
as_symbol("x + y")
as_call("x + y")

as_call(~ f)
as_symbol(~ f())
}
