% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{auto_name}
\alias{auto_name}
\title{Automatically name all components of a lazy dots.}
\usage{
auto_name(x, max_width = 40)
}
\arguments{
\item{x}{A \code{\link{lazy_dots}}}

\item{max_width}{Maximum number of characters to use}
}
\description{
Any components missing a name will automatically get a name added by
looking at the first \code{max_width} characters of the deparsed expression.
}
\examples{
x <- lazy_dots(1 + 2, mean(mpg))
auto_name(x)

auto_name(list(~f, quote(x)))
}
\keyword{internal}

