% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{box}
\alias{box}
\alias{is_box}
\alias{unbox}
\title{Box a value}
\usage{
box(x, class = NULL)

is_box(x, class = NULL)

unbox(box)
}
\arguments{
\item{x}{An R object.}

\item{class}{For \code{box()}, an additional class for the boxed value
(which comes in addition to \code{box}). For \code{is_box()}, a class (or
vector of classes) to be passed to \code{\link[=inherits_all]{inherits_all()}}.}

\item{box}{A boxed value to unbox.}
}
\description{
\code{box()} is similar to \code{\link[base:I]{base::I()}} but it protects a value by
wrapping it in a scalar list rather than by adding an attribute.
\code{unbox()} retrieves the boxed value. \code{is_box()} tests whether an
object is boxed with optional class.
}
\examples{
boxed <- box(letters, "mybox")
is_box(boxed)
is_box(boxed, "mybox")
is_box(boxed, "otherbox")

unbox(boxed)
}
