% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types-raw.R
\name{bytes}
\alias{bytes}
\title{Create a raw vector.}
\usage{
bytes(x)
}
\arguments{
\item{x}{An integerish vector of bytes.}
}
\description{
This is a convenient raw vector constructor. It takes integerish
input (see \code{\link[=is_integerish]{is_integerish()}}).
}
\examples{
bytes(1:10)
bytes(c(0x01, 0xff))
}
