% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call.R
\name{call_fn}
\alias{call_fn}
\title{Extract function from a call}
\usage{
call_fn(call, env = caller_env())
}
\arguments{
\item{call}{A defused function call.}

\item{env}{The environment where to find the definition of the
function quoted in \code{call} in case \code{call} is not wrapped in a
quosure.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

If \code{call} is a quosure or formula, the function will be retrieved
from the associated environment. Otherwise, it is looked up in the
calling frame.
}
\examples{
# Extract from a quoted call:
call_fn(quote(matrix()))
}
\seealso{
\code{\link[=call_name]{call_name()}}
}
\keyword{internal}
