% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack.R
\name{call_inspect}
\alias{call_inspect}
\title{Inspect a call}
\usage{
call_inspect(...)
}
\arguments{
\item{...}{Arguments to display in the returned call.}
}
\description{
This function is a wrapper around \code{\link[base:match.call]{base::match.call()}}. It returns
its own function call.
}
\examples{
# When you call it directly, it simply returns what you typed
call_inspect(foo(bar), "" \%>\% identity())

# Pass `call_inspect` to functionals like `lapply()` or `map()` to
# inspect the calls they create around the supplied function
lapply(1:3, call_inspect)
}
