% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call.R
\name{call_parse_type}
\alias{call_parse_type}
\title{What is the parser type of a call?}
\usage{
call_parse_type(call)
}
\value{
A string, one of:
\itemize{
\item \code{""}
\item \code{"break"}
\item \code{"next"}
\item \code{"while"}
\item \code{"for"}
\item \code{"repeat"}
\item \code{"if"}
\item \code{"function"}
\item \code{"?"}
\item \code{"?unary"}
\item \code{"<-"}
\item \code{"<<-"}
\item \code{"="}
\item \code{":="}
\item \code{"~"}
\item \code{"~unary"}
\item \code{"|"}
\item \code{"||"}
\item \code{"&"}
\item \code{"&&"}
\item \code{"!"}
\item \code{"!!!"}
\item \code{">"}
\item \code{">="}
\item \code{"<"}
\item \code{"<="}
\item \code{"=="}
\item \code{"!="}
\item \code{"+"}
\item \code{"-"}
\item \code{"*"}
\item \code{"/"}
\item \code{"\%\%"}
\item \code{"special"}
\item \code{":"}
\item \code{"!!"}
\item \code{"+unary"}
\item \code{"-unary"}
\item \code{"^"}
\item \code{"$"}
\item \code{"@"}
\item \code{"::"}
\item \code{":::"}
\item \code{"("}
\item \code{"["}
\item \code{"[["}
\item \code{"{"}
}
}
\description{
What is the parser type of a call?
}
\keyword{internal}
