% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle-retired.R
\name{caller_frame}
\alias{caller_frame}
\title{Get caller frame}
\usage{
caller_frame(n = 1)
}
\arguments{
\item{n}{Number of frames to go back.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\keyword{internal}
