% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-check-tabular.R
\name{check_data_frame}
\alias{check_data_frame}
\title{Check argument type (data frame)}
\usage{
check_data_frame(
  x,
  ...,
  allow_null = FALSE,
  arg = caller_arg(x),
  call = caller_env()
)
}
\arguments{
\item{x}{The argument to check.}

\item{...}{Additional arguments passed to \code{\link[=abort]{abort()}}.}

\item{allow_null}{Whether \code{NULL} is allowed.}

\item{arg}{An argument name as a string. This argument
will be mentioned in error messages as the input that is at the
origin of a problem.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[=abort]{abort()}} for more information.}
}
\value{
\code{NULL} invisibly if the check passes, throws an error otherwise.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Checks that an argument is a data frame, producing a friendly error
message on failure.
}
\examples{
check_data_frame(mtcars)

try(check_data_frame(1:5))
}
\seealso{
Other input checkers: 
\code{\link{check_type_number}},
\code{\link{check_type_scalar}},
\code{\link{stop_input_type}()}
}
\concept{input checkers}
