% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dots-ellipsis.R
\name{check_dots_empty}
\alias{check_dots_empty}
\title{Check that dots are empty}
\usage{
check_dots_empty(env = caller_env(), action = abort)
}
\arguments{
\item{env}{Environment in which to look for \code{...}.}

\item{action}{The action to take when the dots have not been used. One of
\code{\link[=abort]{abort()}}, \code{\link[=warn]{warn()}}, \code{\link[=inform]{inform()}} or \code{\link[=signal]{signal()}}.}
}
\description{
Sometimes you just want to use \code{...} to force your users to fully name
the details arguments. This function warns if \code{...} is not empty.
}
\examples{
f <- function(x, ..., foofy = 8) {
  check_dots_empty()
  x + foofy
}

try(f(1, foof = 4))
f(1, foofy = 4)
}
