% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dots-ellipsis.R
\name{check_dots_empty0}
\alias{check_dots_empty0}
\title{Check that dots are empty (low level variant)}
\usage{
check_dots_empty0(..., call = caller_env())
}
\arguments{
\item{...}{Dots which should be empty.}
}
\description{
\code{check_dots_empty0()} is a more efficient version of
\code{\link[=check_dots_empty]{check_dots_empty()}} with a slightly different interface. Instead
of inspecting the current environment for dots, it directly takes
\code{...}. It is only meant for very low level functions where a
couple microseconds make a difference.
}
\keyword{internal}
