% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-check-language.R
\name{check_type_lang}
\alias{check_type_lang}
\alias{check_symbol}
\alias{check_call}
\alias{check_environment}
\alias{check_function}
\alias{check_closure}
\alias{check_formula}
\title{Check argument type (language objects)}
\usage{
check_symbol(
  x,
  ...,
  allow_null = FALSE,
  arg = caller_arg(x),
  call = caller_env()
)

check_call(
  x,
  ...,
  allow_null = FALSE,
  arg = caller_arg(x),
  call = caller_env()
)

check_environment(
  x,
  ...,
  allow_null = FALSE,
  arg = caller_arg(x),
  call = caller_env()
)

check_function(
  x,
  ...,
  allow_null = FALSE,
  arg = caller_arg(x),
  call = caller_env()
)

check_closure(
  x,
  ...,
  allow_null = FALSE,
  arg = caller_arg(x),
  call = caller_env()
)

check_formula(
  x,
  ...,
  allow_null = FALSE,
  arg = caller_arg(x),
  call = caller_env()
)
}
\arguments{
\item{x}{The argument to check.}

\item{...}{Additional arguments passed to \code{\link[=abort]{abort()}}.}

\item{allow_null}{Whether \code{NULL} is allowed.}

\item{arg}{An argument name as a string. This argument
will be mentioned in error messages as the input that is at the
origin of a problem.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[=abort]{abort()}} for more information.}
}
\value{
\code{NULL} invisibly if the check passes, throws an error otherwise.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

These functions check that an argument is an R language object
(symbol, call, formula, environment, function, or closure) and
produce friendly error messages on failure.
}
\examples{
# Succeeds
check_symbol(quote(foo))

# Fails
try(check_symbol("foo"))

}
\seealso{
\link{check_type_scalar}, \link{check_data_frame}
}
