% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type-check-scalar.R
\name{check_type_scalar}
\alias{check_type_scalar}
\alias{check_bool}
\alias{check_string}
\alias{check_name}
\alias{check_number_decimal}
\alias{check_number_whole}
\title{Check argument type (scalar)}
\usage{
check_bool(
  x,
  ...,
  allow_na = FALSE,
  allow_null = FALSE,
  arg = caller_arg(x),
  call = caller_env()
)

check_string(
  x,
  ...,
  allow_empty = TRUE,
  allow_na = FALSE,
  allow_null = FALSE,
  arg = caller_arg(x),
  call = caller_env()
)

check_name(
  x,
  ...,
  allow_null = FALSE,
  arg = caller_arg(x),
  call = caller_env()
)

check_number_decimal(
  x,
  ...,
  min = NULL,
  max = NULL,
  allow_infinite = TRUE,
  allow_na = FALSE,
  allow_null = FALSE,
  arg = caller_arg(x),
  call = caller_env()
)

check_number_whole(
  x,
  ...,
  min = NULL,
  max = NULL,
  allow_infinite = FALSE,
  allow_na = FALSE,
  allow_null = FALSE,
  arg = caller_arg(x),
  call = caller_env()
)
}
\arguments{
\item{x}{The argument to check.}

\item{...}{Additional arguments passed to \code{\link[=abort]{abort()}}.}

\item{allow_na}{Whether \code{NA} values are acceptable.}

\item{allow_null}{Whether \code{NULL} is acceptable.}

\item{arg}{An argument name as a string. This argument
will be mentioned in error messages as the input that is at the
origin of a problem.}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[=abort]{abort()}} for more information.}

\item{allow_empty}{Whether the empty string \code{""} is acceptable
(\code{check_string()} only).}

\item{min}{Minimum value (inclusive). If \code{NULL}, no lower bound is
checked.}

\item{max}{Maximum value (inclusive). If \code{NULL}, no upper bound is
checked.}

\item{allow_infinite}{Whether infinite values are acceptable.}
}
\value{
\code{x} invisibly if the check passes, otherwise an error is
thrown.
}
\description{
These functions check that an argument is of the expected scalar type
and produce friendly error messages otherwise.
}
\examples{
check_bool(TRUE)
try(check_bool(1))

check_string("hello")
try(check_string(42))

check_number_decimal(3.14)
try(check_number_decimal("x"))

check_number_whole(42)
try(check_number_whole(3.5))

}
\seealso{
\link{check_type_lang}, \link{check_data_frame}
}
