% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle-deprecated.R
\name{child_env}
\alias{child_env}
\title{Create a child environment}
\usage{
child_env(.parent, ...)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{\link[=env]{env()}} now supports creating child environments, please use it
instead.
}
\keyword{internal}
