% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnd-message.R
\name{cli-format}
\alias{cli-format}
\title{Formatting messages with the cli package:}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\details{
You can use \href{https://cli.r-lib.org/}{cli} and
\href{https://glue.tidyverse.org/}{glue} formatting syntax by adding
this flag to your namespace:\preformatted{.rlang_use_cli_format <- TRUE
}
\itemize{
\item If \code{TRUE}, an internal error is thrown if cli is not
installed. Make sure it is installed by adding it to your
Imports.
\item If \code{FALSE}, the fallback formatting is used.
}

You can opt out of cli formatting by supplying an escaped input:\preformatted{abort(I("Not interpolated"))
}
}
