% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\name{cnd_new}
\alias{cnd_error}
\alias{cnd_message}
\alias{cnd_new}
\alias{cnd_warning}
\title{Create a condition object.}
\usage{
cnd_new(.type = NULL, ..., .msg = NULL, .call = NULL)

cnd_error(.type = NULL, ..., .msg = NULL, .call = NULL)

cnd_warning(.type = NULL, ..., .msg = NULL, .call = NULL)

cnd_message(.type = NULL, ..., .msg = NULL, .call = NULL)
}
\arguments{
\item{.type}{The condition subclass.}

\item{...}{Named data fields stored inside the condition object.}

\item{.msg}{A default message to inform the user about the
condition when it is signalled.}

\item{.call}{The default call within which the condition occurred.}
}
\description{
These constructors make it easy to create subclassed conditions.
Conditions are objects that power the error system in R. They can
also be used for passing messages to pre-established handlers. See
\code{vignette("conditions")} for more information on how to use
the condition system.
}
\details{
\code{cnd_new()} creates objects inheriting from
\code{condition}. Conditions created with \code{cnd_error()},
\code{cnd_warning()} and \code{cnd_message()} inherit from
\code{error}, \code{warning} or \code{message}.
}
\examples{
# Create a condition inheriting from the s3 type "foo":
cnd <- cnd_new("foo")

# Signal the condition to potential handlers. This has no effect if no
# handler is registered to deal with conditions of type "foo":
cnd_signal(cnd)

# If a relevant handler is on the current evaluation stack, it will be
# called by cnd_signal():
with_handlers(cnd_signal(cnd), foo = thrown(function(c) "caught!"))

# Handlers can be thrown or executed inplace. See with_handlers()
# documentation for more on this.


# Note that merely signalling a condition inheriting of "error" is
# not sufficient to stop a program:
cnd_signal(cnd_error("my_error"))

# you need to use stop() to signal a critical condition that should
# terminate the program if not handled:
# stop(cnd_error("my_error"))
}
\seealso{
\code{\link{cnd_signal}()}, \code{\link{with_handlers}()}.
}

