% Generated by roxygen2 (4.0.2): do not edit by hand
\name{common_env}
\alias{common_env}
\title{Find common environment in list of lazy objects.}
\usage{
common_env(dots)
}
\arguments{
\item{dots}{A list of lazy objects}
}
\description{
Find common environment in list of lazy objects.
}
\keyword{internal}

