% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnd.R
\name{deprecated-cnd}
\alias{deprecated-cnd}
\alias{new_cnd}
\alias{cnd_error}
\alias{cnd_warning}
\alias{cnd_message}
\title{Deprecated condition constructors}
\usage{
new_cnd(.type = NULL, ..., .msg = NULL)

cnd_error(.type = NULL, ..., .msg = NULL)

cnd_warning(.type = NULL, ..., .msg = NULL)

cnd_message(.type = NULL, ..., .msg = NULL)
}
\arguments{
\item{.type}{The condition subclass.}

\item{...}{Named data fields stored inside the condition
object. These dots are evaluated with \link[=dots_list]{explicit
splicing}.}

\item{.msg}{A default message to inform the user about the
condition when it is signalled.}
}
\description{
Deprecated in 0.1.2. See \code{\link[=cnd]{cnd()}}.
}
