% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnd-abort.R
\name{dev-notes-cnd-abort}
\alias{dev-notes-cnd-abort}
\title{Development notes - \code{cnd-abort.R}}
\description{
Development notes - \code{cnd-abort.R}
}
\section{Interaction between error calls and NSE}{


Functions that forward their error context to their caller
shouldn't generally be called via NSE but there are expections,
such as testthat snapshots.
\itemize{
\item \code{do.call()} or \code{eval_bare()} shouldn't generally cause issues. If
the environment exists on the stack, we find its \code{sys.call()}. If
it doesn't exist, taking its \code{sys.call()} returns \code{NULL} which
disables the error context.
\item On the other hand, \code{eval()} always creates a specific frame for
all environments and the \code{sys.call()} for that frame is \code{eval()}.
It wouldn't be useful to display this as the context so calls to
\code{eval()} and \code{evalq()} are replaced by \code{NULL}.
}
}

\keyword{internal}
