% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dots.R
\name{dev-notes-dots}
\alias{dev-notes-dots}
\title{Development notes - \code{dots.R}}
\description{
Development notes - \code{dots.R}
}
\section{\code{.__error_call__.} flag in dots collectors}{


Dots collectors like \code{\link[=dots_list]{dots_list()}} are a little tricky because they
may error out in different situations. Do we want to forward the
context, i.e. set the call flag to the calling environment?
Collectors throw errors in these cases:
\enumerate{
\item While checking their own parameters, in which case the relevant
context is the collector itself and we don't forward.
\item While collecting the dots, during evaluation of the supplied
arguments. In this case forwarding or not is irrelevant because
expressions in \code{...} are evaluated in their own environment
which is not connected to the collector's context.
\item While collecting the dots, during argument constraints checks
such as determined by the \code{.homonyms} argument. In this case we
want to forward the context because the caller of the dots
collector is the one who determines the constraints for its
users.
}
}

\keyword{internal}
