% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle-retired.R
\name{dictionary}
\alias{dictionary}
\alias{as_dictionary}
\alias{is_dictionary}
\title{Create a dictionary}
\usage{
as_dictionary(x, lookup_msg = NULL, read_only = FALSE)

is_dictionary(x)
}
\arguments{
\item{x}{An object for which you want to find associated data.}

\item{lookup_msg}{An error message when your data source is
accessed inappropriately (by position rather than name).}

\item{read_only}{Whether users can replace elements of the
dictionary.}
}
\description{
\Sexpr[results=rd, stage=render]{rlang:::lifecycle("defunct")}

The dictionary class is defunct as of rlang 0.2.0. It was
trying to be too general and did not prove useful. Please use
\code{\link[=as_data_pronoun]{as_data_pronoun()}} or your own pronoun class instead.
}
\keyword{internal}
