% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnd-redispatch.R
\name{dispatch_miss}
\alias{dispatch_miss}
\title{Signal a dispatch failure.}
\usage{
dispatch_miss(.generic, .object, ..., .msg = NULL)
}
\arguments{
\item{.generic}{A string giving the name of the generic which
failed to dispatch a method.}

\item{.object}{The object that was passed to \code{.generic}.}

\item{...}{The additional arguments that were passed to
\code{.generic}. These arguments may be passed on to another
generic if a redispatch strategy is established on the
stack.}

\item{.msg}{An error message for the user if dispatch fails (if no
relevant generic redispatch is active). An error message similar
to normal S3 failure is displayed by default.}
}
\description{
\code{dispatch_miss()} signals an aborting condition (see
\code{\link{cnd_abort}()} and \code{\link{abort}()}) of type
\code{dispatch_miss}. It also establishes a recovery restart named
\code{rst_redispatch}. This is intended to be used in tandem with
\code{\link{with_redispatch}()} with which a user can devise a
redispatch strategy in case method dispatch fails.
}
\examples{
# To allow redispatching, the default method should signal a
# dispatch failure:
my_generic <- function(x, ...) UseMethod("my_generic")
my_generic.bar <- function(x, ...) "bar"
my_generic.default <- function(x, ...) {
  msg <- "dispatch failure."
  dispatch_miss("my_generic", x, ..., .msg = msg)
}

foo_obj <- structure(NULL, class = "foo")

# If the following line is executed, redispatch fails.  See
# with_redispatch() examples to set up a recovery mechanism.
\dontrun{
my_generic(foo_obj)
}
}
\seealso{
\code{\link{with_redispatch}()} to establish a redispatch
  strategy. See \code{\link{hnd_redispatch}()} and
  \code{\link{rst_redispatch}()} to set up a redispatch handler or
  a redispatch restart manually.
}

