% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dots-ellipsis.R
\name{dots-empty}
\alias{dots-empty}
\title{Helper for consistent documentation of empty dots}
\arguments{
\item{...}{These dots are for future extensions and must be empty.}
}
\description{
Use \verb{@inheritParams rlang::`dots-empty`} in your package
to consistently document an unused \code{...} argument.
}
\keyword{internal}
