% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dots-ellipsis.R
\name{dots-used}
\alias{dots-used}
\title{Helper for consistent documentation of used dots}
\arguments{
\item{...}{Arguments passed to methods.}
}
\description{
Use \verb{@inheritParams rlang::dots-used} in your package
to consistently document \code{...} that must be used.
}
\keyword{internal}
