% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dots.R
\name{dots_exprs}
\alias{dots_exprs}
\title{Extract dots forwarded as arguments.}
\usage{
dots_exprs(..., .ignore_empty = "trailing")
}
\arguments{
\item{...}{Arguments to extract.}

\item{.ignore_empty}{Whether to ignore empty arguments. Can be one
of \code{"trailing"}, \code{"none"}, \code{"all"}. If \code{"trailing"}, only the
last argument is ignored if it is empty.}
}
\description{
These functions return the arguments forwarded through \code{...}.
Contrarily to \code{\link[=dots_list]{dots_list()}} and \code{\link[=dots_splice]{dots_splice()}}, \code{dots_exprs()} and
\code{dots_node()} do not evaluate the arguments. The former returns a
list of expressions while the latter returns a \link{pairlist}.
}
\details{
\code{dots_exprs()} performs call-splicing and is compatible with
\link[=UQ]{unquote operators}, including unquote-splicing. \code{dots_node()}
is more bare bones and returns the pairlist as is, without
unquoting.
}
