% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dots.R
\name{dots_info}
\alias{dots_info}
\alias{dots_info_}
\title{Inspect dots}
\usage{
dots_info(...)

dots_info_(dots, stack)
}
\arguments{
\item{...}{Dots to inspect.}

\item{dots}{Dots to inspect.}

\item{stack}{A \code{call_stack} object as returned by
\code{\link{call_stack}()}.}
}
\description{
Runs \code{\link{arg_info}()} for each dots element, and return the
results in a list.
}
\details{
\code{dots_info_()} is the standard evaluation version of
\code{dots_info()} and takes a list of dots as captured by
\code{\link{frame_dots}()} or \code{\link{arg_dots}()}, and a call
stack as returned by \code{\link{call_stack}()}.
}
\seealso{
\code{\link{arg_info}()}
}

