% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dots.R
\name{dots_list}
\alias{dots_list}
\title{Extract dots with explicit splicing.}
\usage{
dots_list(...)
}
\arguments{
\item{...}{Arguments with explicit splicing semantics. Arguments
marked with \code{\link[rlang:spliced]{rlang::spliced()}} are spliced in the returned list.}
}
\description{
\code{dots_list()} evaluates all arguments contained in \code{...} and returns
them as a list. It splices lists that are explicitly marked for
\link[=splice]{splicing} with the \code{\link[=spliced]{spliced()}} adjective. See
\code{\link[=dots_splice]{dots_splice()}} for a version that splices bare lists as well.
}
\details{
\code{dots_list()} and \code{dots_splice()} are simple aliases to
\code{\link[=splice]{splice()}}. Their main purpose is to provide documentation for the
\code{...} argument that developers can inherit from in their packages.
}
\examples{
# Compared to simply using list(...) to capture dots, dots_list()
# splices explicitly:
x <- list(1, 2)
dots_list(spliced(x), 3)

# Unlike dots_splice(), it doesn't splice bare lists:
dots_list(x, 3)
}
\seealso{
\code{\link[=dots_splice]{dots_splice()}} for full splicing semantics, and
\code{\link[=dots_exprs]{dots_exprs()}} for extracting dots without evaluation.
}
