% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dots.R
\name{dots_n}
\alias{dots_n}
\title{How many arguments are currently forwarded in dots?}
\usage{
dots_n(...)
}
\arguments{
\item{...}{Forwarded arguments.}
}
\description{
This returns the number of arguments currently forwarded in \code{...}
as an integer.
}
\examples{
fn <- function(...) dots_n(..., baz)
fn(foo, bar)
}
\keyword{internal}
