% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-quote.R
\name{dots_quosures}
\alias{dots_quosures}
\alias{dots_definitions}
\title{Tidy quotation of multiple expressions and dots.}
\usage{
dots_quosures(..., .named = FALSE)

dots_definitions(..., .named = FALSE)
}
\arguments{
\item{...}{Arguments to capture.}

\item{.named}{Whether to ensure all dots are named. Unnamed
elements are processed with \code{\link[=expr_text]{expr_text()}} to figure out a default
name. If an integer, it is passed to the \code{width} argument of
\code{expr_text()}, if \code{TRUE}, the default width is used. See
\code{\link[=exprs_auto_name]{exprs_auto_name()}}.}
}
\description{
\code{dots_quosures()} quotes its arguments and returns them as a list of
tidy quotes. It is especially useful to "capture" arguments
forwarded through \code{...}.
}
\details{
Both \code{dots_quosures} and \code{dots_definitions()} have specific support for
definition expressions of the type \code{var := expr}, with some
differences:

\describe{
\item{\code{dots_quosures()}}{
When \code{:=} definitions are supplied to \code{dots_quosures()},
they are treated as a synonym of argument assignment
\code{=}. On the other hand, they allow unquoting operators on
the left-hand side, which makes it easy to assign names
programmatically.}
\item{\code{dots_definitions()}}{
This dots capturing function returns definitions as is. Unquote
operators are processed on capture, in both the LHS and the
RHS. Unlike \code{dots_quosures()}, it allows named definitions.}
}
}
\examples{
# dots_quosures() is like the singular version but allows quoting
# several arguments:
dots_quosures(foo(), bar(baz), letters[1:2], !! letters[1:2])

# It is most useful when used with dots. This allows quoting
# expressions across different levels of function calls:
fn <- function(...) dots_quosures(...)
fn(foo(bar), baz)

# Note that dots_quosures() does not check for duplicate named
# arguments:
fn <- function(...) dots_quosures(x = x, ...)
fn(x = a + b)


# Dots can be spliced in:
args <- list(x = 1:3, y = ~var)
dots_quosures(!!! args, z = 10L)

# Raw expressions are turned to formulas:
args <- alist(x = foo, y = bar)
dots_quosures(!!! args)


# Definitions are treated similarly to named arguments:
dots_quosures(x := expr, y = expr)

# However, the LHS of definitions can be unquoted. The return value
# must be a symbol or a string:
var <- "foo"
dots_quosures(!!var := expr)

# If you need the full LHS expression, use dots_definitions():
dots <- dots_definitions(var = foo(baz) := bar(baz))
dots$defs
}
