% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obj.R
\name{duplicate}
\alias{duplicate}
\title{Duplicate an R object}
\usage{
duplicate(x, shallow = FALSE)
}
\arguments{
\item{x}{An R object. Uncopyable objects like symbols and
environments are returned as is (just like with \verb{<-}).}

\item{shallow}{Recursive data structures like lists, calls and
pairlists are duplicated in full by default. A shallow copy only
duplicates the top-level data structure.}
}
\description{
\code{duplicate()} is an interface to the C-level \code{duplicate()} and
\code{shallow_duplicate()} functions. It is mostly meant for users of
the C API of R, e.g. for debugging, experimenting, or prototyping C
code in R.
}
\seealso{
pairlist
}
\keyword{internal}
