% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quosure.R
\name{quo_is_missing}
\alias{quo_is_missing}
\title{Is a quosure empty?}
\usage{
quo_is_missing(quo)
}
\arguments{
\item{quo}{A quosure.}
}
\description{
When missing arguments are captured as quosures, either through
\code{\link[=enquo]{enquo()}} or \code{\link[=quos]{quos()}}, they are returned as an empty quosure. These
quosures contain the \link[=missing_arg]{missing argument} and typically
have the \link[=empty_env]{empty environment} as enclosure.
}
\examples{
quo <- quo()
quo_is_missing(quo)
is_missing(f_rhs(quo))
}
