% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{env<-}
\alias{env<-}
\title{Assignment operator for environments.}
\usage{
env(x) <- value
}
\arguments{
\item{x}{An object with an \code{env_set} method.}

\item{value}{The new environment.}
}
\description{
Assignment operator for environments.
}

