% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{env_assign_lazily}
\alias{env_assign_lazily}
\alias{env_assign_lazily_}
\title{Assign a promise to an environment.}
\usage{
env_assign_lazily(env, nm, expr, eval_env = NULL)

env_assign_lazily_(env, nm, expr, eval_env = NULL)
}
\arguments{
\item{env}{An environment or an object with a S3 method for
\code{env()}.}

\item{nm}{The name of the binding.}

\item{expr}{An expression to capture for
\code{env_assign_lazily()}, or a captured expression (either
quoted or a formula) for the standard evaluation version
\code{env_assign_lazily_()}. This expression is used to create a
promise in \code{env}.}

\item{eval_env}{The environment where the promise will be evaluated
when the promise gets forced. If \code{expr} is a formula, its
environment is used instead. If not a formula and \code{eval_env}
is not supplied, the promise is evaluated in the environment
where \code{env_assign_lazily()} (or the underscore version) was
called.}
}
\description{
These functions let you create a promise in an environment. Such
promises behave just like lazily evaluated arguments. They are
evaluated whenever they are touched by code, but not when they are
passed as arguments.
}
\examples{
env <- env_new()
env_assign_lazily(env, "name", cat("forced!\\n"))
env$name

# Use the standard evaluation version with quoted expressions:
f <- ~message("forced!")
env_assign_lazily_(env, "name2", f)
env$name2
}
\seealso{
\code{\link{env_assign}()}
}

