% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{env_browse}
\alias{env_browse}
\alias{env_is_browsed}
\title{Browse environments}
\usage{
env_browse(env, value = TRUE)

env_is_browsed(env)
}
\arguments{
\item{env}{An environment.}

\item{value}{Whether to browse \code{env}.}
}
\value{
\code{env_browse()} returns the previous value of
\code{env_is_browsed()} (a logical), invisibly.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}
\itemize{
\item \code{env_browse(env)} is equivalent to evaluating \code{browser()} in
\code{env}. It persistently sets the environment for step-debugging.
Supply \code{value = FALSE} to disable browsing.
\item \code{env_is_browsed()} is a predicate that inspects whether an
environment is being browsed.
}
}
