% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle-retired.R
\name{env_bury}
\alias{env_bury}
\title{Bury symbols in a new child environment}
\usage{
env_bury(.env, ...)
}
\arguments{
\item{.env}{An environment or an object bundling an environment,
e.g. a formula, \link[=quotation]{quosure} or \link[=is_closure]{closure}.
This argument is passed to \code{\link[=get_env]{get_env()}}.}

\item{...}{Pairs of names and expressions, values or
functions. These dots support splicing (with varying semantics,
see above) and name unquoting.}
}
\description{
This function was soft-deprecated and renamed to \code{\link[=env_mask]{env_mask()}} in
rlang 0.2.0. This is part of a larger change in terminology. We are
now referring to overscoping and symbol burying as "masking", a
friendlier and more familiar term for R users (cf masked objects in
the search path).
}
\keyword{internal}
