% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{env_caller}
\alias{env_caller}
\title{Get the environment of the caller frame.}
\usage{
env_caller()
}
\description{
This is a shortcut for \code{\link{call_frame}()$env}.
}
\seealso{
\code{\link{call_frame}()}
}

