% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{env_empty}
\alias{env_empty}
\title{Get the empty environment.}
\usage{
env_empty()
}
\description{
The empty environment is the only that does not have a parent. It
is always used as the tail of a scope chain such as the search
path.
}
\examples{
# Create environments with nothing in scope (the default):
env_new(parent = env_empty())
}

