% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{env_has}
\alias{env_has}
\alias{env_sees}
\title{Does an environment have or see bindings?}
\usage{
env_has(env = env_caller(), nms)

env_sees(env = env_caller(), nms)
}
\arguments{
\item{env}{An environment or an object with a S3 method for
\code{env()}. If missing, the environment of the current
evaluation frame is returned.}

\item{nms}{A character vector containing the names of the bindings
to remove.}
}
\value{
A logical vector as long as \code{nms}.
}
\description{
These functions are vectorised predicates. \code{env_has()} queries
whether an environment owns bindings personally, while
\code{env_sees()} will also return \code{TRUE} for the bindings
that are owned by one of the parent environments.
}
\examples{
parent <- env_new(env_empty(), list(foo = "foo"))
env <- env_new(parent, list(bar = "bar"))

# env does not own `foo` but sees it in its parent environment:
env_has(env, "foo")
env_sees(env, "foo")
}

