% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{env_inherits}
\alias{env_inherits}
\title{Does environment inherits from another environment?}
\usage{
env_inherits(x, ancestor)
}
\arguments{
\item{x}{An environment.}

\item{ancestor}{Another environment from which `x` might inherit.}
}
\description{
This returns `TRUE` if `x` has `ancestor` among its parents.
}
