% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{env_label}
\alias{env_label}
\title{Label of an environment}
\usage{
env_label(env)
}
\arguments{
\item{env}{An environment.}
}
\description{
Special environments like the global environment have their own
names. \code{env_label()} returns:
\itemize{
\item "global" for the global environment.
\item "empty" for the empty environment.
\item "base" for the base package environment (the last environment on
the search path).
\item "namespace:pkg" if \code{env} is the namespace of the package "pkg".
\item The \code{name} attribute of \code{env} if it exists. This is how the
\link[=scoped_env]{package environments} and the \link[=ns_imports_env]{imports
environments} store their names. The name of package
environments is typically "package:pkg".
\item The memory address of the environment is returned as fallback.
}
}
\examples{
# Some environments have specific labels:
env_label(global_env())
env_label(ns_env("rlang"))

# Anonymous environments are referenced by their address in memory:
env_label(env())
}
