% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{env_names}
\alias{env_names}
\title{List object names in an environment.}
\usage{
env_names(env)
}
\arguments{
\item{env}{An environment or an object bundling an environment,
e.g. a formula, \link{quosure} or \link[=is_closure]{closure}.}
}
\value{
A character vector of object names.
}
\description{
\code{env_names()} returns object names from an enviroment \code{env} as a character
vector.
All names are returned, even those starting with a dot.
}
\examples{
env <- env(a = 1, b = 2)
env_names(env)
}
