% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{env_unlock}
\alias{env_unlock}
\title{Unlock an environment}
\usage{
env_unlock(env)
}
\arguments{
\item{env}{An environment.}
}
\value{
Whether the environment has been unlocked.
}
\description{
This function should only be used in development tools or
interactively.
}
\keyword{internal}
