% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnd-abort.R
\name{error_arg}
\alias{error_arg}
\title{Format argument for input checking errors}
\usage{
error_arg(arg)
}
\arguments{
\item{arg, error_arg}{A string or symbol for an argument name. This
argument will be mentioned in error messages as the input that is
at the origin of a problem.}
}
\value{
A single string formatted for output.
}
\description{
Transforms an argument name into a formatted string. The string is
formatted by the cli package (if available) with the \code{{.arg }}
format.
}
\details{
\code{arg} can also be a defused call passed to \code{\link[=as_label]{as_label()}},
e.g. \code{quote(1:2)}. This is meant for specific situations where
there are no user-relevant argument names, such as in \code{[} methods.

Use \verb{@inheritParams rlang::error_arg} to document \code{error_arg}
arguments.
}
\seealso{
\code{\link[=error_call]{error_call()}}
}
