% Generated by roxygen2 (4.0.2): do not edit by hand
\name{eval_call}
\alias{eval_call}
\title{Evaluate a call with \code{lazy_dots} as argument.}
\usage{
eval_call(fun, dots, env = parent.frame())
}
\arguments{
\item{fun}{Function as symbol or quoted call.}

\item{dots}{Arguments to function; must be a \code{lazy_dots} object.}

\item{env}{Environment in which to evaluate call. Defaults to
\code{\link{parent.frame}()}.}
}
\description{
Evaluate a call with \code{lazy_dots} as argument.
}
\examples{
make_env <- function(...) list2env(list(...), parent = emptyenv())

f1 <- as.lazy(quote(a()), make_env(a = function() {message("!"); 1}))
f2 <- as.lazy(quote(a), make_env(a = 10))
args <- as.lazy_dots(list(f1, f2))

a <- 100
eval_call(quote(`+`), args)
}

