% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.R
\name{eval_tidy}
\alias{eval_tidy}
\title{Evaluate a quosure.}
\usage{
eval_tidy(f, data = NULL)
}
\arguments{
\item{f}{A formula. Any expressions wrapped in \code{UQ()} will will be
"unquoted", i.e. they will be evaluated, and the results inserted
back into the formula. See \code{\link[=quo]{quo()}} for more details. If a
list of formulas, \code{eval_tidy()} is applied to each of them in
turn and the list of results is returned.}

\item{data}{A list (or data frame). This is passed to the
\code{\link[=as_dictionary]{as_dictionary()}} coercer, a generic used to transform an object
to a proper data source. If you want to make \code{eval_tidy()} work
for your own objects, you can define a method for this generic.}
}
\description{
If \code{data} is specified, variables will be looked for first in this
object, and if not found in the environment of the formula.
}
\section{Pronouns}{


When used with \code{data}, \code{eval_tidy()} provides two pronouns to
make it possible to be explicit about where you want values to
come from: \code{.env} and \code{.data}. These are thin wrappers around
\code{.data} and \code{.env} that throw errors if you try to access
non-existent values.
}

\examples{
eval_tidy(~ 1 + 2 + 3)

# formulas automatically capture their enclosing environment
foo <- function(x) {
  y <- 10
  ~ x + y
}
f <- foo(1)
f
eval_tidy(f)

# If you supply data, eval_tidy() will look there first:
eval_tidy(~ cyl, mtcars)

# To avoid ambiguity, you can use .env and .data pronouns to be
# explicit:
cyl <- 10
eval_tidy(~ .data$cyl, mtcars)
eval_tidy(~ .env$cyl, mtcars)

# Imagine you are computing the mean of a variable:
eval_tidy(~ mean(cyl), mtcars)
# How can you change the variable that's being computed?
# The easiest way is "unquote" with !!
# See ?quosure for more details
var <- ~ cyl
eval_tidy(quo(mean( !!var )), mtcars)
}
