% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle-retired.R
\name{eval_tidy_}
\alias{eval_tidy_}
\title{Tidy evaluation in a custom environment}
\usage{
eval_tidy_(expr, bottom, top = NULL, env = caller_env())
}
\arguments{
\item{expr}{An expression or quosure to evaluate.}

\item{bottom}{The environment containing masking objects if the
data mask is one environment deep. The bottom environment if the
data mask comprises multiple environment.

If you haven't supplied \code{top}, this \strong{must} be an environment
that you own, i.e. that you have created yourself.}

\item{top}{The last environment of the data mask. If the data mask
is only one environment deep, \code{top} should be the same as
\code{bottom}.

This \strong{must} be an environment that you own, i.e. that you have
created yourself. The parent of \code{top} will be changed by the tidy
eval engine and should be considered undetermined. Never make
assumption about the parent of \code{top}.}

\item{env}{The environment in which to evaluate \code{expr}. This
environment is not applicable for quosures because they have
their own environments.}
}
\description{
\Sexpr[results=rd, stage=render]{rlang:::lifecycle("defunct")}

This function is defunct as of rlang 0.3.0.
}
\keyword{internal}
