% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle-retired.R
\name{exiting}
\alias{exiting}
\title{Exiting handler}
\usage{
exiting(handler)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#soft-deprecated}{\figure{lifecycle-soft-deprecated.svg}{options: alt='[Soft-deprecated]'}}}{\strong{[Soft-deprecated]}}

\code{exiting()} is no longer necessary as handlers are exiting by default.
}
\keyword{internal}
