% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explicit-promise.R
\name{explicit_promise}
\alias{explicit_dots}
\alias{explicit_promise}
\title{Make a promise explicit by converting into a formula.}
\usage{
explicit_promise(x)

explicit_dots(..., .ignore_empty = TRUE)
}
\arguments{
\item{x, ...}{An unevaluated promises}

\item{.ignore_empty}{If \code{TRUE}, empty arguments will be silently
dropped.}
}
\value{
\code{explicit_promise} returns a formula; \code{explicit_dots}
  returns a list of formulas.
}
\description{
This should be used sparingly if you want to implement true non-standard
evaluation with 100\% magic. I recommend avoiding this unless you have
strong reasons otherwise since requiring arguments to be formulas only
adds one extra character to the inputs, and otherwise makes life much much
simpler.
}
\examples{
explicit_promise(a + b)
explicit_dots(a + b, c + d, e + f)

# These functions will follow a chain of promises back to the
# original definition
f <- function(x) g(x)
g <- function(y) h(y)
h <- function(z) explicit_promise(z)
f(a + b + c)
}

