% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{f_list}
\alias{as_f_list}
\alias{f_list}
\title{Build a named list from the LHS of formulas}
\usage{
f_list(...)

as_f_list(x)
}
\arguments{
\item{...}{Named arguments.}

\item{x}{An existing list}
}
\value{
A named list.
}
\description{
\code{f_list} makes a new list; \code{as_f_list} takes an existing list.
Both take the LHS of any two-sided formulas and evaluate it, replacing the
current name with the result.
}
\examples{
f_list("y" ~ x)
f_list(a = "y" ~ a, ~ b, c = ~c)
}

