% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{f_new}
\alias{f_new}
\title{Create a formula object by "hand".}
\usage{
f_new(rhs, lhs = NULL, env = parent.frame())
}
\arguments{
\item{lhs, rhs}{A call, name, or atomic vector.}

\item{env}{An environment}
}
\value{
A formula object
}
\description{
Create a formula object by "hand".
}
\examples{
f_new(quote(a))
f_new(quote(a), quote(b))
}

