% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{f_text}
\alias{f_text}
\alias{f_name}
\alias{f_label}
\title{Turn RHS of formula into a string or label}
\usage{
f_text(x, width = 60L, nlines = Inf)

f_name(x)

f_label(x)
}
\arguments{
\item{x}{A formula.}

\item{width}{Width of each line.}

\item{nlines}{Maximum number of lines to extract.}
}
\description{
Equivalent of \code{\link[=expr_text]{expr_text()}} and \code{\link[=expr_label]{expr_label()}} for formulas.
}
\examples{
f <- ~ a + b + bc
f_text(f)
f_label(f)

# Names a quoted with ``
f_label(~ x)
# Strings are encoded
f_label(~ "a\\nb")
# Long expressions are collapsed
f_label(~ foo({
  1 + 2
  print(x)
}))
}
