% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/faq.R
\name{faq-options}
\alias{faq-options}
\title{Global options for rlang}
\description{
rlang has several options which may be set globally to control
behavior. A brief description of each is given here. If any functions
are referenced, refer to their documentation for additional details.
\itemize{
\item \code{rlang_interactive}: A logical value used by \code{\link[=is_interactive]{is_interactive()}}. This
can be set to \code{TRUE} to test interactive behavior in unit tests,
for example.
\item \code{rlang_backtrace_on_error}: A character string which controls whether
backtraces are displayed with error messages, and the level of
detail they print. See \link{rlang_backtrace_on_error} for the possible option values.
\item \code{rlang_trace_format_srcrefs}: A logical value used to control whether
srcrefs are printed as part of the backtrace.
\item \code{rlang_trace_top_env}: An environment which will be treated as the
top-level environment when printing traces. See \code{\link[=trace_back]{trace_back()}}
for examples.
}
}
