% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finterp.R
\name{finterp}
\alias{finterp}
\title{Interpolate a formula}
\usage{
finterp(f)
}
\arguments{
\item{f}{A one-sided formula.}
}
\description{
Interpolation replaces sub-expressions of the form \code{(( x ))} with
the evaluated value of \code{x}.
}
\examples{
finterp(~ 1 + ((1 + 2 + 3)) + 10)

foo <- function(n) {
  ~ 1 + ((n))
}
f <- foo(10)
f
finterp(f)
}

