% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{fn_args}
\alias{fn_args}
\title{Extract arguments from a function}
\usage{
fn_args(fn = NULL)
}
\description{
Contrarily to \code{formals()}, \code{fn_args()} also works with
primitive functions.
}
\examples{
# Extract from current call:
fn <- function(a = 1, b = 2) fn_args()
fn()

# Works with primitive functions:
fn_args(base::switch)
}

