% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn.R
\name{fn_args_defaults}
\alias{fn_args_defaults}
\alias{fn_args_names}
\title{Extract arguments from a function}
\usage{
fn_args_defaults(fn = NULL)

fn_args_names(fn = NULL)
}
\description{
\code{fn_defaults()} returns a list of default
arguments. \code{fn_args_names()} returns the names of the arguments.
}
\details{
Contrarily to \code{formals()}, these helpers also work with
primitive functions. If no argument is named,
\code{fn_args_names()} returns a vector of empty strings.
}
\examples{
# Extract from current call:
fn <- function(a = 1, b = 2) fn_args_defaults()
fn()

# Works with primitive functions:
fn_args_defaults(base::switch)
}

