% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn.R
\name{fn_body}
\alias{fn_body}
\alias{fn_body<-}
\title{Get or set function body}
\usage{
fn_body(fn = caller_fn())

fn_body(fn) <- value
}
\arguments{
\item{fn}{A function. It is lookep up in the calling frame if not
supplied.}

\item{value}{New formals or formals names for \code{fn}.}
}
\description{
\code{fn_body()} is a simple wrapper around \code{base::body()}. The setter
version preserves attributes, unlike \code{body<-}.
}
