% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn.R
\name{fn_fmls}
\alias{fn_fmls}
\alias{fn_fmls_names}
\title{Extract arguments from a function}
\usage{
fn_fmls(fn = NULL)

fn_fmls_names(fn = NULL)
}
\arguments{
\item{fn}{A function. It is lookep up in the calling frame if not
supplied.}
}
\description{
\code{fn_defaults()} returns a named list of default
arguments. \code{fn_fmls_names()} returns the names of the
arguments.
}
\details{
Contrarily to \code{formals()}, these helpers also work with
primitive functions.
}
\examples{
# Extract from current call:
fn <- function(a = 1, b = 2) fn_fmls()
fn()

# Works with primitive functions:
fn_fmls(base::switch)
}
\seealso{
\code{\link{call_args}()} and
  \code{\link{call_args_names}()}
}

