% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnd-message.R
\name{format_bullets}
\alias{format_bullets}
\title{Format bullets for error messages}
\usage{
format_bullets(x)
}
\arguments{
\item{x}{A named character vector of messages. Elements named as
\code{x} or \code{i} are prefixed with the corresponding bullet.}
}
\description{
\Sexpr[results=rd, stage=render]{rlang:::lifecycle("experimental")}

\code{format_bullets()} takes a character vector and returns a single
string (or an empty vector if the input is empty). The elements of
the input vector are assempled as a list of bullets, depending on
their names:
\itemize{
\item Elements named \code{"i"} are bulleted with a blue "info" symbol.
\item Elements named \code{"x"} are bulleted with a red "cross" symbol.
\item Unnamed elements are bulleted with a "*" symbol.
}

This experimental infrastructure is based on the idea that
sentences in error messages are best kept short and simple. From
this point of view, the best way to present the information is in
the \code{\link[=cnd_body]{cnd_body()}} method of an error conditon, as a bullet list of
simple sentences containing a single clause. The info and cross
symbols of the bullets provide hints on how to interpret the bullet
relative to the general error issue, which should be supplied as
\code{\link[=cnd_header]{cnd_header()}}.
}
\section{The \code{bullets} field in condition objects}{


The messages for rlang errors created with \code{\link[=abort]{abort()}} are
automatically formatted with \code{\link[=cnd_message]{cnd_message()}}, which assembles a
message from the \code{cnd_header()}, \code{cnd_body()}, and \code{cnd_footer()}
generics. You can override the body generic by storing a \code{body}
field in the condition object. Similarly, you can store a \code{bullets}
field to override \code{cnd_body()}.

Like \code{body}, this can be a static character vector, a
\link[=as_function]{lambda-formula}, or a function with the same
signature as \code{cnd_body()}. Unlike \code{body} which is included in the
error message literally, the \code{bullets} field (or the result of its
invokation) is formatted with \code{format_bullets()}.
}

