% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle-retired.R
\name{format_error_bullets}
\alias{format_error_bullets}
\title{Format error bullets}
\usage{
format_error_bullets(x)
}
\arguments{
\item{x}{A named character vector of messages. Elements named as
\code{"*"}, \code{"x"}, \code{"i"}, \code{"v"}, or \code{"!"} are prefixed with the
corresponding bullet. Elements named with a single space \code{" "}
trigger a line break from the previous bullet.}
}
\description{
\Sexpr[results=rd, stage=render]{rlang:::lifecycle("deprecated")}
\code{format_error_bullets()} has been renamed to \code{\link[=format_bullets]{format_bullets()}}.
}
