% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack.R
\name{frame_distance}
\alias{frame_distance}
\title{Compute the distance between a frame and the current context.}
\usage{
frame_distance(frame)
}
\arguments{
\item{frame}{Either a frame object whose \code{pos} field will be
extracted, an object with an \code{\link{env}()} method whose
environment will be looked up on the evaluation stack with
\code{\link{frame_position}()}, or an integer representing the
position of the frame on the stack.}
}
\description{
Whereas \code{\link{frame_position}()} returns the number of
evaluation contexts counting from the global frame,
\code{frame_distance()} counts from the currently active frame. The
returned distance can be directly supplied to
\code{\link{eval_frame}()} to get the corresponding frame object.
}
\examples{
fn <- function() g(environment())
g <- function(env) h(env)
h <- function(env) frame_distance(env)
fn()

# frame_distance() takes care of intervening contexts and can be
# safely nested in other calls:
h <- function(env) identity(identity(frame_distance(env)))
fn()
}
\seealso{
\code{\link{frame_position}()}
}

