% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dots.R
\name{frame_dots}
\alias{frame_dots}
\alias{frame_dots_lsp}
\title{Extract dots from a frame}
\usage{
frame_dots(frame = NULL)

frame_dots_lsp(frame = NULL)
}
\arguments{
\item{frame}{The environment from which the dots should be
retrieved. Can be a frame, an environment, or a formula from
which to retrieve an environment. If not supplied, the calling
frame is used.}
}
\description{
\code{frame_dots_lsp()} returns a pairlist that is ready to be
spliced into a call, while \code{frame_dots()} returns a regular
list that is usually easier to work with.
}

