% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dots.R
\name{frame_dots_exprs}
\alias{frame_dots_exprs}
\alias{frame_dots_node}
\title{Extract dots from a frame.}
\usage{
frame_dots_exprs(frame = NULL)

frame_dots_node(frame = NULL)
}
\arguments{
\item{frame}{The environment from which the dots should be
retrieved. Can be a frame, an environment, or a formula from
which to retrieve an environment. If not supplied, the calling
frame is used.}

\item{...}{Arguments to extract. Can be both forwarded dots and
direct arguments.}
}
\description{
These functions extract dots from a frame environment on the
\link[=eval_stack]{evaluation stack}. The \code{_node()} version return a
pairlist that is ready to be spliced into a call, while the regular
version return a regular list that is usually easier to work with.
}
\details{
\code{frame_dots_exprs()} and \code{frame_dots_node()} never fail, even if
the frame does not contain dots. Instead they return an empty list
or \code{NULL} respectively. \code{frame_dots_exprs()} processes
\link[=UQ]{unquote} and \link[=UQS]{unquote-splice} operators.
}
