% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack.R
\name{frame_position}
\alias{frame_position}
\title{Find the position of a frame on the evaluation stack.}
\usage{
frame_position(frame)
}
\arguments{
\item{frame}{The environment of a frame. Can be any object with a
\code{\link{env}()} method. Note that for frame objects, this
simply returns \code{frame$pos}.}
}
\description{
The position is the number of frames on the evaluation stack
(\code{\link{eval_stack}()}), counting from the global frame (whose
position is zero). See also \code{\link{frame_distance}()} for a
function that counts frames from the opposite side (the current
context).
}
\details{
While this function returns the position of the frame on the
evaluation stack, it can safely be called with intervening frames
as those will be discarded.
}
\examples{
fn <- function() g(environment())
g <- function(env) frame_position(env)

# frame_position() returns the position of the frame on the evaluation stack:
fn()
identity(identity(fn()))

# Note that it trims off intervening calls before counting so you
# can safely nest it within other calls:
g <- function(env) identity(identity(frame_position(env)))
fn()
}
\seealso{
\code{\link{frame_distance}()}
}

