% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{friendly_type}
\alias{friendly_type}
\title{Format a type for error messages}
\usage{
friendly_type(type, article = c("indefinite", "definite"))
}
\arguments{
\item{type}{A type as returned by \code{\link[=type_of]{type_of()}}.}

\item{article}{A string, either \code{"indefinite"} or
\code{"definite"}. Indefinite articles depend on the starting letter
of \code{type}: \code{"an"} if starts with a vowel, \code{"a"} otherwise.}
}
\description{
Format a type for error messages
}
\examples{
friendly_type("logical")
friendly_type("integer")
friendly_type("string")
}
