% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{funwrap}
\alias{funwrap}
\title{Unwrap a formula}
\usage{
funwrap(f)
}
\arguments{
\item{f}{A formula to unwrap.}
}
\description{
This interpolates values in the formula that are defined in its environment,
replacing the environment with its parent.
}
\examples{
n <- 100
f <- ~ x + n
funwrap(f)
}

