% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnd-handlers.R
\name{global_handle}
\alias{global_handle}
\title{Register default global handlers}
\usage{
global_handle(entrace = TRUE, prompt_install = TRUE)
}
\arguments{
\item{entrace}{Passed as \code{enable} argument to \code{\link[=global_entrace]{global_entrace()}}.}

\item{prompt_install}{Passed as \code{enable} argument to
\code{\link[=global_prompt_install]{global_prompt_install()}}.}
}
\description{
\code{global_handle()} sets up a default configuration for error,
warning, and message handling. It calls:
\itemize{
\item \code{\link[=global_entrace]{global_entrace()}} to enable rlang errors and warnings globally.
\item \code{\link[=global_prompt_install]{global_prompt_install()}} to recover from \code{packageNotFoundError}s
with a user prompt to install the missing package. Note that at
the time of writing (R 4.1), there are only very limited
situations where this handler works.
}
}
