% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnd-handlers.R
\name{global_prompt_install}
\alias{global_prompt_install}
\title{Prompt user to install missing packages}
\usage{
global_prompt_install(enable = TRUE)
}
\arguments{
\item{enable}{Whether to enable or disable global handling.}
}
\description{
When enabled, \code{packageNotFoundError} thrown by \code{\link[=loadNamespace]{loadNamespace()}}
cause a user prompt to install the missing package and continue
without interrupting the current program.

This is similar to how \code{\link[=check_installed]{check_installed()}} prompts users to install
required packages. It uses the same install strategy, using pak if
available and \code{\link[=install.packages]{install.packages()}} otherwise.
}
