% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnd-signal.R
\name{interrupt}
\alias{interrupt}
\title{Simulate interrupt condition}
\usage{
interrupt()
}
\description{
\code{interrupt()} simulates a user interrupt of the kind that is
signalled with \code{Ctrl-C}. It is currently not possible to create
custom interrupt condition objects.
}
\keyword{internal}
