% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call.R
\name{is_call_infix}
\alias{is_call_infix}
\title{Is object a call to infix operator?}
\usage{
is_call_infix(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{FALSE} if not a call or not a call to an infix
operator. \code{TRUE} otherwise.
}
\description{
Is object a call to infix operator?
}
\examples{
is_call_infix(quote(-1))
is_call_infix(quote(1 - 2))
is_call_infix(quote(1 \%-\% 2))
is_call_infix(quote(a@b))
}
