% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-lang.R
\name{is_callable}
\alias{is_callable}
\title{Is an object callable?}
\usage{
is_callable(x)
}
\arguments{
\item{x}{An object to test.}
}
\description{
A callable object is an object that can be set as the head of a
\link[=lang_head]{call node}. This includes \link[=is_symbolic]{symbolic
objects} that evaluate to a function or literal
functions.
}
\details{
Note that strings may look like callable objects because
expressions of the form \code{"list"()} are valid R code. However,
that's only because the R parser transforms strings to symbols. It
is not legal to manually set language heads to strings.
}
\examples{
# Symbolic objects and functions are callable:
is_callable(quote(foo))
is_callable(base::identity)

# You can safely set the head of a language node to a callable
# object:
lang <- lang("identity", 10L)
lang

lang_literal <- mut_node_car(lang, base::identity)
lang_literal
eval_bare(lang_literal)
}
