% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle-retired.R
\name{is_dictionaryish}
\alias{is_dictionaryish}
\title{Is a vector uniquely named?}
\usage{
is_dictionaryish(x)
}
\arguments{
\item{x}{A vector.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Like \code{\link[=is_named]{is_named()}} but also checks that names are unique.
}
\keyword{internal}
