% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dots.R
\name{is_dot_symbol}
\alias{is_dot_symbol}
\title{Is object a ..n symbol?}
\usage{
is_dot_symbol(x)
}
\arguments{
\item{x}{An object to test.}
}
\description{
Is object a ..n symbol?
}
\examples{
is_dot_symbol(quote(..2))
is_dot_symbol(quote(sym))
}

