% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{is_empty}
\alias{is_empty}
\title{Is object an empty vector or NULL?}
\usage{
is_empty(x)
}
\arguments{
\item{x}{object to test}
}
\description{
Is object an empty vector or NULL?
}
\examples{
is_empty(NULL)
is_empty(list())
is_empty(list(NULL))
}
