% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle-retired.R
\name{is_env}
\alias{is_env}
\alias{is_bare_env}
\title{Is an object an environment?}
\usage{
is_env(x)

is_bare_env(x)
}
\arguments{
\item{x}{object to test}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

These functions were deprecated and renamed to \code{\link[=is_environment]{is_environment()}}
and \code{\link[=is_bare_environment]{is_bare_environment()}} in rlang 0.2.0. This is for consistency
with other type predicates which are not abbreviated.
}
\keyword{internal}
