% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{is_environment}
\alias{is_environment}
\alias{is_bare_environment}
\title{Is object an environment?}
\usage{
is_environment(x)

is_bare_environment(x)
}
\arguments{
\item{x}{object to test}
}
\description{
\code{is_bare_environment()} tests whether \code{x} is an environment without a s3 or
s4 class.
}
