% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{is_formula}
\alias{is_formula}
\title{Is object a formula?}
\usage{
is_formula(x)
}
\arguments{
\item{x}{object to test}
}
\description{
Is object a formula?
}
\examples{
x <- disp ~ am
is_formula(x)

is_formula(~ 10)
is_formula(10)
}

