% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env-special.R
\name{is_installed}
\alias{is_installed}
\title{Are packages installed in any of the libraries?}
\usage{
is_installed(pkg)
}
\arguments{
\item{pkg}{The package names.}
}
\value{
A named logical vector of the same length as \code{pkg}. \code{TRUE} for each
installed package and \code{FALSE} for each package not found in any of the
libraries.
}
\description{
This checks that packages are installed with minimal side effects.
If installed, the packages will be loaded but not attached.
}
\examples{
is_installed(c("graphics", "utils"))
is_installed("ggplot5")
}
