% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is_language}
\alias{is_atomic}
\alias{is_call}
\alias{is_language}
\alias{is_name}
\title{Is an object a language object?}
\usage{
is_language(x)

is_name(x)

is_call(x)

is_atomic(x)
}
\arguments{
\item{x}{An object to test.}
}
\description{
These helpers are consistent wrappers around their base R equivalents.
A language object is either an atomic vector (typically a scalar), a
name (aka a symbol), or a call.
}
\examples{
q1 <- quote(1)
is_language(q1)
is_atomic(q1)

q2 <- quote(x)
is_language(q2)
is_name(q2)

q3 <- quote(x + 1)
is_language(q3)
is_call(q3)
}

