% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attr.R
\name{is_named}
\alias{is_named}
\alias{have_name}
\title{Is object named?}
\usage{
is_named(x)

have_name(x)
}
\arguments{
\item{x}{An object to test.}
}
\value{
\code{is_named()} is a scalar predicate that returns \code{TRUE} or
\code{FALSE}. \code{have_name()} is vectorised and returns a logical vector
as long as the input.
}
\description{
\code{is_named()} checks that \code{x} has names attributes, and that none of
the names are missing or empty (\code{NA} or \code{""}). \code{have_name()} is a
vectorised version of \code{is_named()}.
}
\examples{
# A data frame normally has valid, unique names
is_named(mtcars)
have_name(mtcars)

# Empty and missing names are invalid:
invalid <- dups
names(invalid)[2] <- ""
names(invalid)[5] <- NA

# is_named() performs a global check while have_name() can show you
# where the problem is:
is_named(invalid)
have_name(invalid)

# is_named() only returns TRUE if there is a names attribute (a
# zero-length character vector in this case):
x <- set_names(list(), character(0))
is_named(x)

# have_name() will work even with vectors that don't have a names
# attribute:
have_name(letters)
}
