% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env-special.R
\name{is_namespace}
\alias{is_namespace}
\title{Is an object a namespace environment?}
\usage{
is_namespace(x)
}
\arguments{
\item{x}{An object to test.}
}
\description{
Is an object a namespace environment?
}
