% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle-retired.R
\name{is_quosureish}
\alias{is_quosureish}
\alias{as_quosureish}
\title{Test for or coerce to quosureish objects}
\usage{
is_quosureish(x, scoped = NULL)

as_quosureish(x, env = caller_env())
}
\arguments{
\item{x}{An object to test.}

\item{scoped}{A boolean indicating whether the quosure is scoped,
that is, has a valid environment attribute. If \code{NULL}, the scope
is not inspected.}

\item{env}{An environment specifying the lexical enclosure of the
quosure.}
}
\description{
These functions were deprecated in rlang 0.2.0 because they assumed
that quosures are formulas which is currently true but might not be
in the future.
}
\keyword{internal}
