% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle-retired.R
\name{is_scoped}
\alias{is_scoped}
\title{Retired \code{scoped} functions}
\usage{
is_scoped(nm)
}
\description{
\Sexpr[results=rd, stage=render]{rlang:::lifecycle("soft-deprecated")}

These functions are soft-deprecated as of rlang 0.3.0. They are
replaced by \code{\link[=is_attached]{is_attached()}}, ...
}
\keyword{internal}
