% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle-retired.R
\name{is_stack}
\alias{is_stack}
\alias{is_eval_stack}
\alias{is_call_stack}
\title{Is object a stack?}
\usage{
is_stack(x)

is_eval_stack(x)

is_call_stack(x)
}
\arguments{
\item{x}{An object to test}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#soft-deprecated}{\figure{lifecycle-soft-deprecated.svg}{options: alt='[Soft-deprecated]'}}}{\strong{[Soft-deprecated]}}
}
\keyword{internal}
