% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sym.R
\name{is_symbol}
\alias{is_symbol}
\title{Is object a symbol?}
\usage{
is_symbol(x, name = NULL)
}
\arguments{
\item{x}{An object to test.}

\item{name}{An optional name or vector of names that the symbol
should match.}
}
\description{
Is object a symbol?
}
