% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{is_true}
\alias{is_true}
\alias{is_false}
\title{Is object identical to TRUE or FALSE?}
\usage{
is_true(x)

is_false(x)
}
\arguments{
\item{x}{object to test}
}
\description{
These functions bypass R's automatic conversion rules and check
that \code{x} is literally \code{TRUE} or \code{FALSE}.
}
\examples{
is_true(TRUE)
is_true(1)

is_false(FALSE)
is_false(0)
}
