% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lang-call.R
\name{lang_fn}
\alias{lang_fn}
\title{Extract function from a call}
\usage{
lang_fn(call = caller_frame())
}
\arguments{
\item{call}{Can be a call, a formula quoting a call in the
right-hand side, or a frame object from which to extract the call
expression. If not supplied, the calling frame is used.}
}
\description{
If a frame or formula, the function will be retrieved from their
environment. Otherwise, it is looked up in the calling frame.
}
\examples{
# Extract from a quoted call:
lang_fn(~matrix())
lang_fn(quote(matrix()))

# Extract the calling function
test <- function() lang_fn()
test()
}
\seealso{
\code{\link[=lang_name]{lang_name()}}
}
