% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle-retired.R
\name{lang_head}
\alias{lang_head}
\alias{lang_tail}
\title{Return the head or tail of a call}
\usage{
lang_head(lang)

lang_tail(lang)
}
\arguments{
\item{lang}{A call.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

As of rlang 0.2.0 these functions are retired (deprecated for now)
because they are low level accessors that are rarely needed for end
users.
}
\keyword{internal}
