% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle-retired.R
\name{lang_modify}
\alias{lang_modify}
\title{Modify the arguments of a call}
\usage{
lang_modify(.lang, ..., .standardise = FALSE)
}
\arguments{
\item{.lang}{The \code{.call} argument of \code{\link[=call_modify]{call_modify()}}.}

\item{...}{Named or unnamed expressions (constants, names or calls)
used to modify the call. Use \code{NULL} to remove arguments. Dots are
evaluated with \link[=dots_list]{explicit splicing}.}

\item{.standardise}{If \code{TRUE}, the call is standardised beforehand
to match existing unnamed arguments to their argument names. This
prevents new named arguments from accidentally replacing original
unnamed arguments.}
}
\description{
This function is soft-deprecated, please use \code{\link[=call_modify]{call_modify()}}
instead.
}
\keyword{internal}
