% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lang-call.R
\name{lang_standardise}
\alias{lang_standardise}
\title{Standardise a call.}
\usage{
lang_standardise(call = caller_frame())
}
\arguments{
\item{call}{Can be a call, a formula quoting a call in the
right-hand side, or a frame object from which to extract the call
expression. If not supplied, the calling frame is used.}
}
\value{
A tidy quote if \code{.call} is a tidy quote, a call otherwise.
}
\description{
This is essentially equivalent to \code{\link[base:match.call]{base::match.call()}}, but handles
primitive functions more gracefully.
}
\seealso{
\code{\link[=lang_homogenise]{lang_homogenise()}} for a version more suitable to
language analysis.
}
