% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace.R
\name{last_error}
\alias{last_error}
\title{Last \code{abort()} error}
\usage{
last_error()
}
\description{
This returns the last error thrown with \code{\link[=abort]{abort()}}. The error is
printed with a backtrace.
}
