% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnd-last.R
\name{last_warnings}
\alias{last_warnings}
\alias{last_messages}
\alias{last_warning}
\alias{last_message}
\title{Display last warnings}
\usage{
last_warnings()

last_messages()

last_warning()

last_message()
}
\description{
\itemize{
\item \code{last_warnings()} and \code{last_messages()} return a list of all
warnings and messages that occurred during the last top-level R
command.
\item \code{last_warning()} and \code{last_message()} return the very last ones.
}

If you call these in the console, these warnings are printed with a
backtrace. Use \code{print(last_warnings(), simplify = level)} to
control the verbosity of the backtrace. The \code{simplify} argument
supports one of \code{"branch"} (the default), \code{"collapse"}, and
\code{"none"} (in increasing order of verbosity).
}
