% Generated by roxygen2 (4.0.1): do not edit by hand
\name{lazy_}
\alias{lazy}
\alias{lazy_}
\title{Capture expression for later lazy evaluation}
\usage{
lazy_(expr, env)

lazy(x, env = parent.frame())
}
\description{
\code{lazy()} uses non-standard evaluation to turn promises into lazy
objects; \code{lazy_()} does standard evaluation and is suitable for
programming.
}
\examples{
lazy_(quote(a + x), globalenv())
lazy(a + b / c)

f <- function(x = b - a) {
  lazy(x)
}
f()
f(a + b / c)
}

