% Generated by roxygen2 (4.0.1): do not edit by hand
\name{lazy_eval}
\alias{lazy_eval}
\title{Evaluate a lazy expression.}
\usage{
lazy_eval(x, data = NULL)
}
\arguments{
\item{x}{A lazy object.}

\item{data}{Option, a data frame or list in which to preferentially look
for variables before using the environment associated with the lazy
object.}
}
\description{
Evaluate a lazy expression.
}
\examples{
f <- function(x) {
  z <- 100
  ~ x + z
}
z <- 10
lazy_eval(f(10))
lazy_eval(f(10), list(x = 100))
lazy_eval(f(10), list(x = 1, z = 1))
}

