% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attr.R
\name{length_}
\alias{length_}
\title{Get length of an object.}
\usage{
length_(x)
}
\arguments{
\item{x}{A R object.}
}
\description{
Unlike \code{\link[base]{length}()}, this function is not a s3
generic and will always return the actual length of an object.
}
\examples{
# s3 objects can have a length() method:
x <- structure(letters, class = "foo")
length.foo <- function(x) 100
length(x)

# length_() will always reveal the true length of the object:
length_(x)
}

