% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle.R
\name{lifecycle-deprecated}
\alias{lifecycle-deprecated}
\title{Deprecated functions and arguments}
\description{
Deprecated functions and arguments
}
\section{rlang 0.2.0}{

\itemize{
\item \code{\link[=UQE]{UQE()}}
}
}

