% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle.R
\name{lifecycle-soft-deprecated}
\alias{lifecycle-soft-deprecated}
\title{Soft-deprecated functions and arguments}
\description{
Soft-deprecated functions and arguments
}
\section{rlang 0.2.0}{

\itemize{
\item \code{\link[=UQ]{UQ()}}
\item \code{\link[=UQS]{UQS()}}
}
}

