% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle-retired.R
\name{ll}
\alias{ll}
\title{Deprecated list constructor}
\usage{
ll(...)
}
\arguments{
\item{...}{Components of the new vector. Bare lists and explicitly
spliced lists are spliced.}
}
\description{
\code{ll()} has been soft-deprecated and renamed to \code{\link[=list2]{list2()}} in rlang
0.2.0. This is consistent with the new \code{\link[=call2]{call2()}} constructor for
calls with explicit \code{!!!} support.
}
\keyword{internal}
