% Generated by roxygen2 (4.0.2): do not edit by hand
\name{make_call}
\alias{make_call}
\title{Generate a call.}
\usage{
make_call(fun, dots)
}
\arguments{
\item{fun}{Function as symbol or quoted call.}

\item{dots}{Arguments to function; must be a \code{lazy_dots} object.}
}
\description{
Generate a call.
}
\examples{
make_env <- function(...) list2env(list(...), parent = emptyenv())

f1 <- as.lazy(quote(a), make_env(a = 1))
f2 <- as.lazy(quote(a), make_env(a = 10))
args <- as.lazy_dots(list(f1, f2))
call <- make_call(quote(`+`), args)

a <- 100
# eval(call)
}

