% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_formula}
\alias{make_formula}
\title{Create a formula object by "hand".}
\usage{
make_formula(expr, env = parent.frame())
}
\arguments{
\item{expr}{A call, name, or atomic vector.}

\item{env}{An environment}
}
\value{
A formula object
}
\description{
Create a formula object by "hand".
}
\keyword{internal}

