% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conn.R
\name{maybe_close}
\alias{maybe_close}
\title{Close a temporary connection.}
\usage{
maybe_close(conn)
}
\arguments{
\item{conn}{A connection to close.}
}
\description{
\code{maybe_close()} closes connections but only if they have been
marked as temporary (see \code{\link{temporary}()}). It returns
\code{TRUE} if the connection was temporary and has been closed.
}
\examples{
path <- tempfile("my-file")
conn <- file(path)

maybe_close(conn)
maybe_close(temporary(conn))
maybe_close(temporary(file(path)))
}

