% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{meow}
\alias{meow}
\alias{conn_meow}
\title{Cat a line with a trailing newline}
\usage{
meow(...)

conn_meow(.conn, ...)
}
\arguments{
\item{...}{Output lines. These are passed to \code{\link[=chr]{chr()}} so lists of
character vectors are automatically flattened.}

\item{.conn}{A connection to output to.}
}
\description{
This little helper is like \code{cat()} with a couple of differences:
\itemize{
\item Instead of interleaving its inputs with spaces, it interleaves
new lines.
\item It automatically adds a new line at the end.
}

\code{meow()} always outputs to the console while \code{conn_meow()} takes a
connection to output to.
}
\examples{
# meow() supports lists of character vectors:
lines <- list(
  "<my-type>",
  "Metadata: foo"
)

# It automatically adds newlines to all its inputs:
meow(lines)
print(letters)
meow(
  "---",
  paste("n:", length(letters)),
  paste("type:", typeof(letters))
)
}
