% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{missing_arg}
\alias{missing_arg}
\title{Generate a missing argument.}
\usage{
missing_arg()
}
\description{
Generate a missing argument.
}
\examples{
f_interp(~f(x = uq(missing_arg())))
f_interp(~f(x = uq(NULL)))
}

