% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{modify}
\alias{modify}
\title{Modify a vector.}
\usage{
modify(.x, ..., .elts = list())
}
\arguments{
\item{.x}{A vector to modify.}

\item{..., .elts}{List of elements to merge into \code{.x}.}
}
\value{
A modified vector upcasted to a list.
}
\description{
This function merges a list of arguments into a vector. It always
returns a list.
}
\examples{
modify(c(1, b = 2, 3), 4, b = "foo")

x <- list(a = 1, b = 2)
y <- list(b = 3, c = 4)
modify(x, .elts = y)
}

