% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle-retired.R
\name{mut_node_car}
\alias{mut_node_car}
\alias{mut_node_cdr}
\alias{mut_node_caar}
\alias{mut_node_cadr}
\alias{mut_node_cdar}
\alias{mut_node_cddr}
\alias{mut_node_tag}
\title{Mutate node components}
\usage{
mut_node_car(x, newcar)

mut_node_cdr(x, newcdr)

mut_node_caar(x, newcar)

mut_node_cadr(x, newcar)

mut_node_cdar(x, newcdr)

mut_node_cddr(x, newcdr)

mut_node_tag(x, newtag)
}
\arguments{
\item{x}{A language or pairlist node. Note that these functions are
barebones and do not perform any type checking.}

\item{newcar}{The new CAR or CDR for the node. These
can be any R objects.}

\item{newcdr}{The new CAR or CDR for the node. These
can be any R objects.}

\item{newtag}{The new tag for the node. This should be a symbol.}
}
\description{
\Sexpr[results=rd, stage=render]{rlang:::lifecycle("soft-deprecated")}

These functions were soft-deprecated and renamed with \code{node_poke_}
prefix in rlang 0.2.0. This change follows a new naming convention
where mutation is referred to as "poking".
}
\keyword{internal}
