% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attr.R
\name{names2}
\alias{names2}
\title{Get names of a vector.}
\usage{
names2(x)
}
\arguments{
\item{x}{A vector.}
}
\description{
This names getter always returns a character vector, even when an
object does not have a \code{names} attribute. In this case, it
returns a vector of empty names \code{""}.
}
\examples{
names2(letters)
}

