% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{names_inform_repair}
\alias{names_inform_repair}
\title{Inform about names repair}
\usage{
names_inform_repair(old, new)
}
\arguments{
\item{old}{Original names vector.}

\item{new}{Repaired names vector.}
}
\description{
Inform about names repair
}
\keyword{internal}
