% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle-retired.R
\name{new-vector-along-retired}
\alias{new-vector-along-retired}
\alias{new_logical_along}
\alias{new_integer_along}
\alias{new_double_along}
\alias{new_character_along}
\alias{new_complex_along}
\alias{new_raw_along}
\alias{new_list_along}
\title{Create vectors matching the length of a given vector}
\usage{
new_logical_along(x, names = base::names(x))

new_integer_along(x, names = base::names(x))

new_double_along(x, names = base::names(x))

new_character_along(x, names = base::names(x))

new_complex_along(x, names = base::names(x))

new_raw_along(x, names = base::names(x))

new_list_along(x, names = base::names(x))
}
\arguments{
\item{x}{A vector.}

\item{names}{Names for the new vector.}
}
\description{
These functions are deprecated as of rlang 0.3.0 because they
are longer to type than the equivalent \code{\link[=rep_along]{rep_along()}} or
\code{\link[=rep_named]{rep_named()}} calls without added clarity.
}
\keyword{internal}
