% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec-new.R
\name{new-vector}
\alias{new-vector}
\alias{new_logical}
\alias{new_integer}
\alias{new_double}
\alias{new_character}
\alias{new_complex}
\alias{new_raw}
\alias{new_list}
\title{Create vectors matching a given length}
\usage{
new_logical(.n)

new_integer(.n)

new_double(.n)

new_character(.n)

new_complex(.n)

new_raw(.n)

new_list(.n)
}
\arguments{
\item{.n}{The vector length.}
}
\description{
These functions construct vectors of given length, with attributes
specified via dots. Except for \code{new_list()} and \code{new_bytes()}, the
empty vectors are filled with typed \link{missing} values. This is in
contrast to the base function \code{\link[base:vector]{base::vector()}} which creates
zero-filled vectors.
}
\examples{
new_list(10)
new_logical(10)
}
\seealso{
vector-along
}
\keyword{internal}
