% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{new-vectors}
\alias{chr_new}
\alias{cpl_new}
\alias{dbl_new}
\alias{int_new}
\alias{lgl_new}
\alias{lst_new}
\alias{new-vectors}
\alias{raw_new}
\title{Create new vectors.}
\usage{
lgl_new(.n = 0, ..., .attrs = list())

int_new(.n = 0, ..., .attrs = list())

dbl_new(.n = 0, ..., .attrs = list())

chr_new(.n = 0, ..., .attrs = list())

lst_new(.n = 0, ..., .attrs = list())

cpl_new(.n = 0, ..., .attrs = list())

raw_new(.n = 0, ..., .attrs = list())
}
\arguments{
\item{.n}{The vector length.}

\item{...}{A list of named attributes.}

\item{.attrs}{A list of named attributes.}
}
\description{
These functions construct vectors of given length, with attributes
specified via dots. Except for \code{lst_new()} and
\code{raw_new()}, the empty vectors are filled with typed
\link{missing} values. This is in contrast to the base function
\code{\link[base]{vector}()} which creates zero-filled vectors.
}
\examples{
lst_new(10)

# Add attributes, including the S3 class:
int_new(0, index = 1)
dbl_new(10, class = "my_class")
}
\seealso{
along
}

