% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{new_error_package_not_found}
\alias{new_error_package_not_found}
\title{Signal that required packages were not found}
\usage{
new_error_package_not_found(
  pkg,
  version = NULL,
  ...,
  reason = NULL,
  class = NULL
)
}
\arguments{
\item{pkg}{The package names.}

\item{version}{Minimum versions for \code{pkg}. If supplied, must be the
same length as \code{pkg}. \code{NA} elements stand for any versions.}

\item{reason}{Optional string indicating why is \code{pkg} needed.
Appears in error messages (if non-interactive) and user prompts
(if interactive).}

\item{class, ...}{Subclass and additional condition fields.}
}
\description{
\itemize{
\item \code{new_error_package_not_found()} constructs a condition of class
\code{rlib_error_package_not_found}. This condition includes the
character vectors \code{pkg} and \code{version}. They may contain more than
one package. \code{version} can be \code{NULL} and must be otherwise the
same length as \code{pkg}.
}
}
