% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn.R
\name{new_fn}
\alias{new_fn}
\title{Make an \code{fn} object}
\usage{
new_fn(fn)
}
\arguments{
\item{fn}{A closure.}
}
\value{
An object of class \code{c("fn", "function")}.
}
\description{
\code{new_fn()} takes a function and sets the class to \code{c("fn", function)}.
\itemize{
\item Inheriting from \code{"fn"} enables a print method that strips all
attributes (except \code{srcref}) before printing. This is currently
the only purpose of the \code{fn} class.
\item Inheriting from \code{"function"} makes sure your function still
dispatches on type methods.
}
}
\examples{
fn <- set_attrs(function() "foo", attribute = "foobar")
print(fn)

# The `fn` object doesn't print with attributes:
fn <- new_fn(fn)
print(fn)
}
