% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\name{new_options}
\alias{new_options}
\alias{opt_env}
\alias{opt_bind}
\alias{opt_peek}
\title{Options}
\usage{
new_options()

opt_env(pkg, opt)

opt_bind(pkg, ...)

opt_peek(pkg, opt)
}
\description{
Options
}
\section{Defining package options}{


By convention, package options are exported as \code{mypkg_options}.
This allows \code{opt_env()} to find it in the package environment.\preformatted{#' Global options in mypkg
#' @export
mypkg_options <- new_options()
}
}

