% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle-retired.R
\name{new_overscope}
\alias{new_overscope}
\title{Create a new data mask}
\usage{
new_overscope(bottom, top = NULL, enclosure = base_env())
}
\arguments{
\item{enclosure}{The \code{parent} argument of \code{\link[=new_data_mask]{new_data_mask()}}.}
}
\description{
\code{new_overscope()} is soft-deprecated as of rlang 0.2.0. Please use
\code{\link[=new_data_mask]{new_data_mask()}} instead.
}
