% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quosure.R
\name{new_tidy_quote}
\alias{new_tidy_quote}
\title{Create a quosure by hand.}
\usage{
new_tidy_quote(rhs, env = caller_env())
}
\arguments{
\item{rhs}{A call, name, or atomic vector.}

\item{env}{An environment.}
}
\description{
This is similar to \code{\link{new_formula}()} but for one-sided
formulas. See \code{\link{tidy_quote}()} for the role played by
such formulas in the tidy evaluation framework.
}
\examples{
f <- new_tidy_quote(quote(mtcars), env("datasets"))
f
tidy_eval(f)
}
