% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weakref.R
\name{new_weakref}
\alias{new_weakref}
\title{Create a weak reference}
\usage{
new_weakref(key, value = NULL, finalizer = NULL, onexit = FALSE)
}
\arguments{
\item{key}{The key for the weak reference. Must be a reference object -- that
is, an environment or external pointer.}

\item{value}{The value for the weak reference. This can be \code{NULL}, if you
want to use the weak reference like a weak pointer.}

\item{finalizer}{A function that is run after the key becomes unreachable.}

\item{onexit}{Should the finalizer be run when R exits?}
}
\description{
A weak reference is a special R object which makes it possible to keep a
reference to an object without preventing garbage collection of that object.
It can also be used to keep data about an object without preventing GC of the
object, similar JavaScript WeakMap.

As long as the key is reachable, the value will not be garbage collected.
This is true even if the weak reference object becomes unreachable.

When the key becomes unreachable, the key and value in the weak reference
object are replaced by \code{NULL}, and the finalizer is scheduled to execute.
}
\examples{
e <- new.env()

# Create a weak reference to e
w <- new_weakref(e, finalizer = function(e) message("finalized"))

# Get the key object from the weak reference
identical(weakref_get_key(w), e)

# When the regular reference (the `e` binding) is removed and a GC occurs,
# the weak reference will not keep the object alive.
rm(e)
gc()
identical(weakref_get_key(w), NULL)


# A weak reference with a key and value. The value contains data about the
# key.
k <- new.env()
v <- list(1, 2, 3)
w <- new_weakref(k, v)

identical(weakref_get_key(w), k)
identical(weakref_get_value(w), v)

# When v is removed, the weak ref keeps it alive because k is still reachable.
rm(v)
gc()
identical(weakref_get_value(w), list(1, 2, 3))

# When k is removed, the weak ref does not keep k or v alive.
rm(k)
gc()
identical(weakref_get_key(w), NULL)
identical(weakref_get_value(w), NULL)
}
\seealso{
\code{\link[=is_weakref]{is_weakref()}}, \code{\link[=weakref_get_key]{weakref_get_key()}} and \code{\link[=weakref_get_value]{weakref_get_value()}}.
}
\keyword{experimental}
