% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-node.R
\name{node_modify}
\alias{node_modify}
\title{Modify a call or pairlist.}
\usage{
node_modify(node, ...)
}
\arguments{
\item{node}{A call (language object) or pairlist.}

\item{...}{New named or unnamed elements. These dots have \link[=dots_list]{explicit
splicing semantics}: the contents of arguments marked
with \code{\link[=spliced]{spliced()}} are embedded in surrounding dots list. Unnamed
elements are appended to named elements.}
}
\description{
This merges a list of arguments into a pairlist or call.
}
\seealso{
\code{\link[=lang_modify]{lang_modify()}} for a version that optionally standardises
arguments and handles quosures transparently.
}
