% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env-special.R
\name{ns_registry_env}
\alias{ns_registry_env}
\title{Return the namespace registry env}
\usage{
ns_registry_env()
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}
}
\keyword{internal}
