% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{op-get-attr}
\alias{op-get-attr}
\alias{\%@\%}
\alias{\%@\%<-}
\title{Infix attribute accessor and setter}
\usage{
x \%@\% name

x %@% name <- value
}
\arguments{
\item{x}{Object}

\item{name}{Attribute name}
}
\description{
Infix attribute accessor and setter
}
\examples{
factor(1:3) \%@\% "levels"
mtcars \%@\% class

mtcars \%@\% class <- NULL
mtcars
}
