% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{op-null-continuation}
\alias{op-null-continuation}
\alias{\%&&\%}
\title{Default value for non-\code{NULL}}
\usage{
x \%&&\% y
}
\arguments{
\item{x, y}{If \code{x} is NULL, will return \code{x}; otherwise returns \code{y}.}
}
\description{
This infix operator is the conceptual opposite of \verb{\%||\%}, providing a fallback
only if \code{x} is defined.
}
\examples{
1 \%&&\% 2
NULL \%&&\% 2
}
\seealso{
\link{op-null-default}
}
