% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle-retired.R
\name{overscope_eval_next}
\alias{overscope_eval_next}
\title{Evaluate next quosure in a data mask}
\usage{
overscope_eval_next(overscope, quo, env = base_env())
}
\arguments{
\item{overscope}{A valid overscope containing bindings for \code{~},
\code{.top_env} and \code{_F} and whose parents contain overscoped bindings
for tidy evaluation.}

\item{quo}{A quosure.}

\item{env}{The lexical enclosure in case \code{quo} is not a validly
scoped quosure. This is the \link[=base_env]{base environment} by
default.}
}
\description{
\Sexpr[results=rd, stage=render]{rlang:::lifecycle("soft-deprecated")}

\code{overscope_eval_next()} is soft-deprecated as of rlang
0.2.0. Please use \code{eval_tidy()} to which you can now supply an
overscope.
}
\keyword{internal}
