% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_quo}
\alias{parse_quo}
\alias{parse_quos}
\title{Parsing variants that return quosures}
\usage{
parse_quo(x, env)

parse_quos(x, env)
}
\arguments{
\item{x}{Text containing expressions to parse_expr for
\code{parse_expr()} and \code{parse_exprs()}. Can also be an R connection,
for instance to a file. If the supplied connection is not open,
it will be automatically closed and destroyed.}

\item{env}{The environment for the quosures. Depending on the use
case, a good default might be the \link[=global_env]{global environment} but you might also want to evaluate the
R code in an isolated context (perhaps a child of the global
environment or of the \link[=base_env]{base environment}).}
}
\description{
\Sexpr[results=rd, stage=render]{rlang:::lifecycle("questioning")}

These functions are in the questioning stage because they don't add
value compared to using \code{\link[=parse_expr]{parse_expr()}} with \code{\link[=new_quosure]{new_quosure()}}.
}
\keyword{internal}
