% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle-retired.R
\name{parse_quosure}
\alias{parse_quosure}
\title{Parse text into a quosure}
\usage{
parse_quosure(x, env = caller_env())
}
\arguments{
\item{x}{Text containing expressions to parse_expr for
\code{parse_expr()} and \code{parse_exprs()}. Can also be an R connection,
for instance to a file. If the supplied connection is not open,
it will be automatically closed and destroyed.}

\item{env}{The environment for the quosures. Defaults to the
context in which the parse_expr function was called. Can be any
object with a \code{as_environment()} method.}
}
\description{
These functions were soft-deprecated and renamed to \code{\link[=parse_quo]{parse_quo()}}
and \code{\link[=parse_quos]{parse_quos()}} in rlang 0.2.0. This is for consistency with the
convention that suffixes indicating return types are not
abbreviated.
}
\keyword{internal}
