% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compat-oldrel.R
\name{port_back}
\alias{port_back}
\title{Assign a backport in a package namespace}
\usage{
port_back(fns_names, pkg_name)
}
\arguments{
\item{fns_names}{Character vector giving the names of the functions
to backport.}

\item{pkg_name}{The name of the package where the backports will be
assigned. This is passed to \code{\link[=ns_env]{ns_env()}}. Can also be an
environment (this is handy for testing when your namespace is
locked).}
}
\description{
This function should be called from \code{\link[=.onLoad]{.onLoad()}} hooks. It checks
for the current R version and assigns the requested backports if
they are not available for that version.

Currently, only a few functions that appeared in 3.2.0 are
available: \code{dir.exists()}, \code{names.environment()} and \code{trimws()}.
}
\examples{
# Add a backport in your .onLoad() hook:
.onLoad <- function(lib_name, pkg_name) {
  port_back(c("trimws", "dir.exists"), pkg_name)
}
}
