% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn.R
\name{prim_name}
\alias{prim_name}
\title{Name of a primitive function}
\usage{
prim_name(prim)
}
\arguments{
\item{prim}{A primitive function such as \code{\link[base:c]{base::c()}}.}
}
\description{
Name of a primitive function
}
\keyword{internal}
