% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quosure.R
\name{quo_expr}
\alias{quo_expr}
\alias{quo_label}
\alias{quo_text}
\title{Splice a quosure and format it into string or label.}
\usage{
quo_expr(quo)

quo_label(quo)

quo_text(quo, width = 60L, nlines = Inf)
}
\arguments{
\item{quo}{A quosure or expression.}

\item{width}{Width of each line.}

\item{nlines}{Maximum number of lines to extract.}
}
\description{
\code{quo_expr()} flattens all quosures within an expression. I.e., it
turns \code{~foo(~bar(), ~baz)} to \code{foo(bar(), baz)}. \code{quo_text()} and
\code{quo_label()} are equivalent to \code{\link[=f_text]{f_text()}}, \code{\link[=expr_label]{expr_label()}}, etc,
but they first splice their argument using \code{quo_expr()}.
}
\examples{
quo_expr(~foo(~bar))
quo_text(~foo(~bar))
}
\seealso{
\code{\link[=expr_label]{expr_label()}}, \code{\link[=f_label]{f_label()}}
}
