% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quo.R
\name{quo_get_expr}
\alias{quo_get_expr}
\alias{quo_get_env}
\alias{quo_set_expr}
\alias{quo_set_env}
\title{Get or set the components of a quosure}
\usage{
quo_get_expr(quo)

quo_get_env(quo)

quo_set_expr(quo, expr)

quo_set_env(quo, env)
}
\arguments{
\item{quo}{A quosure.}

\item{expr}{A new expression for the quosure.}

\item{env}{A new environment for the quosure.}
}
\description{
These functions are equivalent to \code{\link[=get_expr]{get_expr()}}, \code{\link[=get_env]{get_env()}},
\code{\link[=set_expr]{set_expr()}}, and \code{\link[=set_env]{set_env()}}. However they only work on quosures
and are a bit more efficient.
}
\section{Life cycle}{


All these functions are stable.
}

\examples{
quo <- quo(foo(bar))
quo

quo_set_expr(quo, quote(baz))
quo_set_env(quo, empty_env())
}
