% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle.R
\name{register_experimental}
\alias{register_experimental}
\alias{check_experimental}
\title{Enable experimental functions}
\usage{
register_experimental(names, env = caller_env())

check_experimental(name, env = caller_env(2L))
}
\arguments{
\item{names}{Names of functions to enable.}

\item{env}{Environment in which to register or check a
function. The \code{topenv()} of that environment is taken for both
registration and validation.}

\item{name}{Name of function to check.}
}
\description{
Call this at the top level of your package to enable experimental
functions. Please avoid depending on experimental functions as API
changes are likely.
}
\examples{
# A package might export an experimental function for
# testing purposes:
highly_experimental <- function() {
  check_experimental("thispkg::highly_experimental")
  "return_value"
}

# To use it in your package, register it explicitly in your
# namespace:
register_experimental("thispkg::highly_experimental")

# You can then use it without triggering an error:
myfunction <- function() {
  highly_experimental()
}
}
