% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rhs}
\alias{rhs}
\title{Extract the right-hand-side of a formula}
\usage{
rhs(f)
}
\arguments{
\item{f}{A formula}
}
\value{
An atomic vector of length 1, a name, or a call.
}
\description{
Throws an error if \code{f} is not a formula, or it's not one-sided.
}
\examples{
rhs(~ 1 + 2 + 3)
rhs(~ x)
rhs(~ "A")
}

