% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rlang-package.R
\docType{package}
\name{rlang-package}
\alias{rlang}
\alias{rlang-package}
\title{rlang: Functions for Base Types and Core R and 'Tidyverse' Features}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A toolbox for working with base types, core R features like the condition system, and core 'Tidyverse' features like tidy evaluation.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rlang.r-lib.org}
  \item \url{https://github.com/r-lib/rlang}
  \item Report bugs at \url{https://github.com/r-lib/rlang/issues}
}

}
\author{
\strong{Maintainer}: Lionel Henry \email{lionel@posit.co}

Authors:
\itemize{
  \item Hadley Wickham \email{hadley@posit.co}
}

Other contributors:
\itemize{
  \item mikefc \email{mikefc@coolbutuseless.com} (Hash implementation based on Mike's xxhashlite) [copyright holder]
  \item Yann Collet (Author of the embedded xxHash library) [copyright holder]
  \item Posit, PBC [copyright holder, funder]
}

}
\keyword{internal}
