% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnd-abort.R
\name{rlang_backtrace_on_warning}
\alias{rlang_backtrace_on_warning}
\alias{rlang_backtrace_on_warning_report}
\title{Display backtrace on warning}
\description{
Similarly to \code{\link{rlang_backtrace_on_error_report}}, you can set
\code{\link{rlang_backtrace_on_warning_report}} inside RMarkdown documents to
tweak the display of warnings. This is useful in conjunction with
\code{\link[=global_entrace]{global_entrace()}}.

\if{html}{\out{<div class="sourceCode">}}\preformatted{```\{r setup\}
rlang::global_entrace()
options(rlang_backtrace_on_warning_report = "full")
```
}\if{html}{\out{</div>}}

Because of technical limitations, there is currently no
corresponding \code{rlang_backtrace_on_warning} option for normal R
sessions.
}
\seealso{
rlang_backtrace_on_error
}
