% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{rlang_options}
\alias{rlang_options}
\title{Global options in rlang}
\format{
An object of class \code{rlib_options} of length 3.
}
\usage{
rlang_options
}
\description{
Full documentation.
}
\keyword{datasets}
