% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnd-redispatch.R
\name{rst_redispatch}
\alias{rst_redispatch}
\title{Restart dispatching.}
\usage{
rst_redispatch(generic, object, args, env)
}
\arguments{
\item{generic}{A generic function or a name of a such a function in
scope from \code{env}. It is invoked to redispatch \code{object}.}

\item{object}{The object whose dispatching failed.}

\item{args}{Additional arguments passed on to \code{generic}.}

\item{env}{The environment in which \code{generic} should be
evaluated.}
}
\description{
This restart function is established at signal site by
\code{\link{dispatch_miss}()}. It is normally called by
\code{\link{hnd_redispatch}()} but can be invoked manually with
\code{\link{rst_jump}("rst_redispatch")}.
}
\seealso{
\code{\link{dispatch_miss}()}, \code{\link{hnd_redispatch}()}
}

