% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle-retired.R
\name{scoped_interactive}
\alias{scoped_interactive}
\alias{scoped_options}
\alias{scoped_bindings}
\title{Questioning \code{scoped_} functions}
\usage{
scoped_interactive(value = TRUE, frame = caller_env())

scoped_options(..., .frame = caller_env())

scoped_bindings(..., .env = .frame, .frame = caller_env())
}
\arguments{
\item{value}{A single \code{TRUE} or \code{FALSE}. This overrides the return
value of \code{is_interactive()}.}

\item{frame}{The environment of a running function which defines
the scope of the temporary options. When the function returns,
the options are reset to their original values.}

\item{...}{For \code{local_options()} and \code{push_options()}, named
values defining new option values. For \code{peek_options()}, strings
or character vectors of option names.}

\item{.frame}{The environment of a stack frame which defines the
scope of the temporary options. When the frame returns, the
options are set back to their original values.}

\item{.env}{An environment.}
}
\description{
\Sexpr[results=rd, stage=render]{rlang:::lifecycle("questioning")}

These functions have been renamed to use the conventional \code{local_}
prefix. They will be deprecated in the next minor version of rlang.
}
