% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attr.R
\name{set_attrs}
\alias{set_attrs}
\title{Add attributes to an object.}
\usage{
set_attrs(.x, ...)
}
\arguments{
\item{.x}{An object to decorate with attributes.}

\item{...}{A list of named attributes. These have \link[=dots_list]{explicit
splicing semantics}.}
}
\description{
Add attributes to an object.
}
\examples{
set_attrs(letters, names = 1:26, class = "my_chr")

# Splice a list of attributes:
attrs <- list(attr = "attr", names = 1:26, class = "my_chr")
set_attrs(letters, splice(attrs))
}
