% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{set_env}
\alias{set_env}
\title{Set an environment.}
\usage{
set_env(env, new_env = caller_env())
}
\arguments{
\item{env}{An environment or an object with a S3 method for
\code{set_env()}.}

\item{new_env}{An environment to replace \code{env} with. Can be an
object with an S method for \code{get_env()}.}
}
\description{
\code{set_env()} does not work by side effect. The input is copied
before being assigned an environment, and left unchanged.
}
\examples{
# Create a function with a given environment:
env <- child_env(base_env())
fn <- with_env(env, function() NULL)
identical(get_env(fn), env)

# set_env() does not work by side effect. Setting a new environment
# for fn has no effect on the original function:
other_env <- child_env()
set_env(fn, other_env)
identical(get_env(fn), other_env)

# set_env() returns a new function with a different environment, so
# you need to assign the returned function to the `fn` name:
fn <- set_env(fn, other_env)
identical(get_env(fn), other_env)
}
