% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sexp.R
\name{sexp_address}
\alias{sexp_address}
\title{Address of an R object}
\usage{
sexp_address(x)
}
\arguments{
\item{x}{Any R object.}
}
\value{
Its address in memory in a string.
}
\description{
Address of an R object
}
\keyword{internal}
