% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnd.R
\name{signal_context_info}
\alias{signal_context_info}
\title{Return information about signalling context}
\usage{
signal_context_info(nframe)
}
\arguments{
\item{nframe}{The depth of the frame to inspect. In a condition
handler, this would typically be \code{sys.nframe() - 1L}.}
}
\value{
A named list of two elements \code{type} and \code{depth}. The depth
is the call frame number of the signalling context. The type is
one of:
\itemize{
\item \code{"unknown"}
\item \code{"stop_message"} for errors thrown with `base::stop("message")"
\item \code{"stop_condition"} for errors thrown with \code{base::stop(cnd_object)}
\item \code{"stop_native"} for errors thrown from C
\item \code{"stop_rlang"} for errors thrown with \code{rlang::abort()}
\item \code{"warning_message"} for warnings signalled with `base::warning("message")"
\item \code{"warning_condition"} for warnings signalled with \code{base::warning(cnd_object)}
\item \code{"warning_native"} for warnings signalled from C
\item \code{"warning_promoted"} for warnings promoted to errors with \code{getOption("warn")}
\item \code{"warning_rlang"} for warnings signalled with \code{rlang::warn()}
\item \code{"message"} for messages signalled with \code{base::message()}
\item \code{"message_rlang"} for messages signalled with \code{rlang::inform()}
\item \code{"condition"} for conditions signalled with \code{base::signalCondition()}
}
}
\description{
Return information about signalling context
}
\keyword{internal}
