% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnd-utils.R
\name{signal_deprecated}
\alias{signal_deprecated}
\alias{warn_obsolete}
\title{Signal a deprecated or obsolete function}
\usage{
signal_deprecated(name, replacement, start = NULL)

warn_obsolete(name, replacement, start = NULL)
}
\arguments{
\item{name}{The name of the depreciated function as a string.}

\item{replacement}{The name of the suggested replacement as a
string.}

\item{start}{The package version where the function was
depreciated.}
}
\description{
These functions signal a function as deprecated or obsolete.
\itemize{
\item A deprecated function does not issue a warning unless the
\code{rlang_verbose_deprecation} option is set to \code{TRUE}. This is
often the first stage towards obsolescence. It is recommended to
stop using deprecated functions and switch to suggested
replacements.
\item An obsolete function always issues a warning when called. Users
should stop using that function as it might get removed in the
future.
}
}
\examples{
# A deprecated function is signalled silently:
foo <- function() signal_deprecated("foo", "bar", "0.2.0")
foo()

# If the `rlang_verbose_deprecation` option is TRUE, a warning is
# issued. If run, the following would throw a warning:
#> options(rlang_verbose_deprecation = TRUE)
#> foo()
}
