% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/versioning.R
\name{signal_deprecation}
\alias{signal_deprecation}
\title{Signal deprecation of a function or arguments}
\usage{
signal_deprecation(.name, .cycle, ..., .msg = NULL)
}
\arguments{
\item{.name}{The name of the function that is deprecated or whose
arguments are deprecated.}

\item{.cycle}{A character vector defining the deprecation cycle.}

\item{...}{Replacements.}

\item{.msg}{A custom error message.}
}
\description{
Signal deprecation of a function or arguments
}
