% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector-ctor.R
\name{splice}
\alias{splice}
\alias{is_spliced}
\title{Splice a list within a vector.}
\usage{
splice(x)

is_spliced(x)
}
\arguments{
\item{x}{A list to splice.}
}
\description{
This adjective signals to functions taking dots that \code{x} should be
spliced in a surrounding vector. Examples of functions that support
such explicit splicing are \code{\link[=list_splice]{list_splice()}}, \code{\link[=chr]{chr()}}, etc.
}
\seealso{
\code{\link[=list_splice]{list_splice()}}, \link{vector-construction}
}
