% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector-ctor.R
\name{spliced}
\alias{spliced}
\alias{is_spliced}
\title{Splice a list within a vector.}
\usage{
spliced(x)

is_spliced(x)
}
\arguments{
\item{x}{A list to splice.}
}
\description{
This adjective signals to functions taking dots that \code{x} should be
spliced in a surrounding vector. Examples of functions that support
splicing are \code{\link[=spliced]{spliced()}}, \code{\link[=chr]{chr()}}, etc.
}
\seealso{
\code{\link[=splice]{splice()}}, \link{vector-construction}
}
