% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle-retired.R
\name{stack}
\alias{stack}
\alias{ctxt_frame}
\alias{global_frame}
\title{Call stack information}
\usage{
ctxt_frame(n = 1)

global_frame()
}
\arguments{
\item{n}{The number of frames to go back in the stack.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Deprecated as of rlang 0.3.0.
}
\keyword{internal}
