% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call.R
\name{standardise_call}
\alias{standardise_call}
\title{Standardise a function call}
\usage{
standardise_call(call, env = parent.frame())
}
\arguments{
\item{call}{A call}

\item{env}{Environment in which to look up call value.}
}
\description{
Standardise a function call
}

