% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types-chr.R
\name{str_bytes}
\alias{str_bytes}
\title{Return a string as a raw vector.}
\usage{
str_bytes(x)
}
\arguments{
\item{x}{A string or character vector.}
}
\value{
A raw vector of bytes.
}
\description{
These helpers return the hexadecimal representation of a string.
}
