% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attr.R
\name{struct}
\alias{struct}
\title{Add attributes to an object.}
\usage{
struct(.x, ...)
}
\arguments{
\item{.x}{An object to decorate with attributes.}

\item{...}{A list of named attributes. These have \link[=dots_list]{explicit
splicing semantics}.}
}
\description{
Add attributes to an object.
}
\examples{
struct(letters, names = 1:26, class = "my_chr")

# Splice a list of attributes:
attrs <- list(attr = "attr", names = 1:26, class = "my_chr")
struct(letters, spliced(attrs))
}
