% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{switchpatch}
\alias{switchpatch}
\title{Dispatch on base types.}
\usage{
switchpatch(.x, ..., .to)
}
\arguments{
\item{.x}{An object from which to dispatch.}

\item{...}{Named clauses. The names should be types as returned by
\code{\link[base]{typeof}()}.}

\item{.to}{This is useful when you switchpatch in a coercing
function. If supplied, this should be a string indicating the
target type. A catch-all clause is then added to signal an error
stating the conversion failure.}
}
\description{
This is equivalent to
\code{\link[base]{switch}(\link[base]{typeof}(x, ...))}.
}
\examples{
switchpatch(3L,
  double = "foo",
  integer = "bar",
  "default"
)
}
