% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types-sym.R
\name{symbol}
\alias{symbol}
\alias{symbols}
\title{Create a symbol or list of symbols.}
\usage{
symbol(x)

symbols(x)
}
\arguments{
\item{x}{A string or list of strings.}
}
\value{
A symbol for \code{symbol()} and a list of symbols for
  \code{symbols()}.
}
\description{
These functions take strings as input and turn them into symbols.
}
