% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attr.R
\name{temporary}
\alias{is_temporary}
\alias{temporary}
\title{Add temporary attribute}
\usage{
temporary(x)

is_temporary(x)
}
\arguments{
\item{x}{An object to add the temporary attribute or check for the
attribute.}
}
\description{
This is a simple wrapper that sets the \code{tmp} attribute of an
object to \code{TRUE}.
}
\details{
This is used for instance in \code{\link{maybe_close}()} to detect
a transient connection that should be automatically closed. This
verb could be used in a similar way in other contexts.
}
\examples{
is_temporary(letters)
is_temporary(temporary(letters))
}

