% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-eval.R
\name{tidy_eval_env}
\alias{tidy_eval_env}
\title{Create a tidy evaluation environment.}
\usage{
tidy_eval_env(env = env_empty(), data = NULL)
}
\arguments{
\item{env}{The original scope.}

\item{data}{Additional data to put in scope.}
}
\description{
This is useful when want to use the tidy evaluation framework in
your own evaluating functions. The returned environment has
elements of \code{data} in scope, the \code{.env} and \code{.data}
pronouns, and treats formulas as self-evaluating promises. See
\code{\link{tidy_eval}()} and \code{\link{tidy_quote}()} for more
information.
}

