% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-capture.R
\name{tidy_patterns}
\alias{tidy_patterns}
\title{Capture patterns.}
\usage{
tidy_patterns(...)
}
\arguments{
\item{...}{Dots that may contain patterns (see \link{op-pattern}).
The LHS and RHS of patterns are interpolated on capture.}
}
\description{
This is like \code{\link{tidy_dots}()} but patterns are returned as
is. Unquote operators are processed on capture, in both the LHS and
the RHS.
}
\examples{
fn <- function(...) {
  tidy_patterns(...)
}
fn(var = foo(baz) := bar(baz))
}
\seealso{
\code{\link{tidy_dots}()}, \link{op-pattern}
}

