% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic-nse.R
\name{topic-embracing-non-args}
\alias{topic-embracing-non-args}
\title{Does \verb{\{\{} work on regular objects?}
\description{
The \link[=defusing]{embracing} operator \verb{\{\{} is meant for function
arguments:\if{html}{\out{<div class="sourceCode r">}}\preformatted{fn <- function(arg) \{
  quo(foo(\{\{ arg \}\}))
\}

fn(1 + 1)
#> <quosure>
#> expr: ^foo(^1 + 1)
#> env:  0x7ffd89aac518
}\if{html}{\out{</div>}}

However you may have noticed that it also works on regular objects:\if{html}{\out{<div class="sourceCode r">}}\preformatted{fn <- function(arg) \{
  arg <- force(arg)
  quo(foo(\{\{ arg \}\}))
\}

fn(1 + 1)
#> <quosure>
#> expr: ^foo(^2)
#> env:  0x7ffd8a633398
}\if{html}{\out{</div>}}

In that case, \verb{\{\{} captures the \emph{value} of the expression instead
of a defused expression. That's because only function arguments can
be defused.
}
\section{Why is this not an error?}{


Ideally we would have made \verb{\{\{} on regular objects an error.
However this is not possible because in compiled R code it is not
always possible to distinguish a regular variable from a function
argument. See \link[=topic-embracing-constants]{Why are strings and other constants enquosed in the empty environment?} for more about this.
}

