% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topic-nse.R
\name{topic-injection-out-of-context}
\alias{topic-injection-out-of-context}
\title{What happens if I use injection operators out of context?}
\description{
TODO!
\itemize{
\item \verb{\{\{} link to \link{topic-embrace-non-args}.
}
}
