% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{type_of}
\alias{type_of}
\title{Base type of an object.}
\usage{
type_of(x)
}
\arguments{
\item{x}{An R object.}
}
\description{
This is equivalent to \code{\link[base]{typeof}()} but also returns
the rlang base types. Currently, only tidy quotes have been added
to \code{type_of()}. Their rlang type is "quote".
}
\examples{
type_of(10L)
type_of(~10L)
}
