% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle-deprecated.R
\name{UQ}
\alias{UQ}
\alias{UQS}
\title{Deprecated \code{UQ()} and \code{UQS()} operators}
\usage{
UQ(x)

UQS(x)
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
These operators are deprecated in favour of
\code{\link[=injection-operator]{!!}} and \code{\link[=splice-operator]{!!!}}.
}
\keyword{internal}
