% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle-retired.R
\name{UQE}
\alias{UQE}
\title{Unquote as a bare expression}
\usage{
UQE(x)
}
\arguments{
\item{x}{Object to unquote.}
}
\description{
\Sexpr[results=rd, stage=render]{rlang:::lifecycle("defunct")}

\code{UQE()} is defunct as of rlang 0.3.0 in order to simplify the
quasiquotation syntax. You can replace its use by a combination of
\code{!!} and \code{get_expr()}: \code{!!get_expr(x)} is equivalent to \code{UQE(x)}.
}
\keyword{internal}
