% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle-retired.R
\name{vector-len}
\alias{vector-len}
\alias{lgl_len}
\alias{int_len}
\alias{dbl_len}
\alias{chr_len}
\alias{cpl_len}
\alias{raw_len}
\alias{bytes_len}
\alias{list_len}
\title{Retired vector construction by length}
\usage{
lgl_len(.n)

int_len(.n)

dbl_len(.n)

chr_len(.n)

cpl_len(.n)

raw_len(.n)

bytes_len(.n)

list_len(.n)
}
\arguments{
\item{.n}{The vector length.}
}
\description{
These functions were soft-deprecated and renamed with \code{new_} prefix
in rlang 0.2.0. This is for consistency with other non-variadic
object constructors.
}
\keyword{internal}
