% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle-retired.R
\name{node}
\alias{node}
\alias{vector-old-ctors}
\alias{lgl_along}
\alias{int_along}
\title{Retired vector construction by length}
\usage{
node(car, cdr = NULL)

lgl_along(.x)

int_along(.x)
}
\arguments{
\item{.x}{A vector.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

These functions were deprecated and renamed with \code{new_} prefix in
rlang 0.2.0. This is for consistency with other non-variadic object
constructors.
}
\keyword{internal}
