% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attr.R
\name{with_attributes}
\alias{with_attributes}
\title{Add attributes to an object.}
\usage{
with_attributes(.x, ..., .attrs = list())
}
\arguments{
\item{.x}{An object to decorate with attributes.}

\item{..., .attrs}{A list of named attributes.}
}
\description{
Add attributes to an object.
}
\examples{
with_attributes(letters, class = "my_chr", names = 1:26)

attrs <- list(class = "my_chr", names = 1:26)
with_attributes(letters, .attrs = attrs)
}
