% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle-deprecated.R
\name{with_handlers}
\alias{with_handlers}
\alias{calling}
\alias{exiting}
\title{Establish handlers on the stack}
\usage{
with_handlers(.expr, ...)

calling(handler)

exiting(handler)
}
\arguments{
\item{.expr, ..., handler}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

As of rlang 1.0.0, \code{with_handlers()} is deprecated. Use the base
functions or the experimental \code{\link[=try_fetch]{try_fetch()}} function instead.
}
\keyword{internal}
