% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weakref.R
\name{wref_key}
\alias{wref_key}
\alias{wref_value}
\title{Get key/value from a weak reference object}
\usage{
wref_key(x)

wref_value(x)
}
\arguments{
\item{x}{A weak reference object.}
}
\description{
Get key/value from a weak reference object
}
\seealso{
\code{\link[=is_weakref]{is_weakref()}} and \code{\link[=new_weakref]{new_weakref()}}.
}
