% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3.R
\name{zap}
\alias{zap}
\alias{zaps}
\alias{is_zap}
\title{Create zap objects}
\usage{
zap()

zaps(n)

is_zap(x)
}
\arguments{
\item{n}{Either the length of the zap list, or a character vector
of names for the zaps. In the latter case, the length of the list
is deduced from the length of the character vector.}

\item{x}{An object to test.}
}
\description{
\code{zap()} creates a sentinel object that indicates that an object
should be removed. For instance, named zaps instruct \code{\link[=env_bind]{env_bind()}}
and \code{\link[=call_modify]{call_modify()}} to remove those objects from the environment or
the call.

The advantage of zap objects is that they unambiguously signal the
intent of removing an object. Sentinels like \code{NULL} or
\code{\link[=missing_arg]{missing_arg()}} are ambiguous because they represent valid R
objects.
}
\examples{
# Create one zap object:
zap()

# Create a list of zaps:
zaps(2)
zaps(letters[1:3])
}
