% Generated by roxygen2 (4.0.2): do not edit by hand
\name{all_dots}
\alias{all_dots}
\title{Combine explicit and implicit dots.}
\usage{
all_dots(.dots, ..., env = parent.frame())
}
\arguments{
\item{.dots}{A list of lazy objects}

\item{...}{Individual lazy objects}

\item{env}{Default environment to use for non-lazy objects}
}
\value{
A \code{\link{lazy_dots}}
}
\description{
Combine explicit and implicit dots.
}

