% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{along}
\alias{along}
\alias{chr_along}
\alias{cpl_along}
\alias{dbl_along}
\alias{int_along}
\alias{lgl_along}
\alias{lst_along}
\alias{raw_along}
\alias{rep_along}
\title{Helper to create vectors with matching length.}
\usage{
lgl_along(.x, ..., .attrs = list())

int_along(.x, ..., .attrs = list())

dbl_along(.x, ..., .attrs = list())

chr_along(.x, ..., .attrs = list())

lst_along(.x, ..., .attrs = list())

cpl_along(.x, ..., .attrs = list())

raw_along(.x, ..., .attrs = list())

rep_along(.x, .y, ..., .attrs = list())
}
\arguments{
\item{.x}{A vector.}

\item{...}{A list of named attributes.}

\item{.attrs}{A list of named attributes.}

\item{.y}{Values to repeat.}
}
\description{
These functions take the idea of \code{\link{seq_along}} and
generalise it to creating lists (\code{lst_along}) and repeating
values (\code{rep_along}). The dots and \code{.attrs} are forwarded
to \code{\link{with_attributes}()} to make it easy to add
attributes. Except for \code{lst_along()} and \code{raw_along()},
the empty vectors are filled with typed \link{missing} values.
}
\examples{
x <- 0:5
rep_along(x, 1:2)
rep_along(x, 1)
lst_along(x)

# You can also add attributes with additional arguments:
rep_along(x, 1, class = "my_class")
dbl_along(x, class = "my_class")
}
\seealso{
new-vectors
}

