% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg.R
\name{arg_require}
\alias{arg_require}
\title{Check that argument is supplied
Throws an informative error if \code{arg} is missing.}
\usage{
arg_require(arg)
}
\arguments{
\item{arg}{A function argument. Must be a symbol.}
}
\description{
Check that argument is supplied
Throws an informative error if \code{arg} is missing.
}
\examples{
f <- function(x)  {
  arg_require(x)
}

# Fails because `x` is not supplied
try(f())

# Succeeds
f(NULL)
}
\seealso{
\code{\link[=arg_match]{arg_match()}}
}
