% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg.R
\name{arg_require}
\alias{arg_require}
\title{Check that argument is supplied
Throws an informative error if \code{arg} is missing.}
\usage{
arg_require(arg, error_call = caller_env())
}
\arguments{
\item{arg}{A function argument. Must be a symbol.}

\item{error_call}{An expression (as returned by e.g.
\code{sys.call()}) representing the context in which the error
occurred. If non-null, the call is stripped from its arguments to
keep it simple.

Can also be an execution environment of a currently running
function (as returned by e.g. \code{parent.frame()}). The
corresponding call is then retrieved.

See also \code{\link[=local_error_call]{local_error_call()}}.}
}
\description{
Check that argument is supplied
Throws an informative error if \code{arg} is missing.
}
\examples{
f <- function(x)  {
  arg_require(x)
}

# Fails because `x` is not supplied
try(f())

# Succeeds
f(NULL)
}
\seealso{
\code{\link[=arg_match]{arg_match()}}
}
