% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg.R
\name{arg_require}
\alias{arg_require}
\title{Check that argument is supplied}
\usage{
arg_require(arg, error_arg = caller_arg(arg), error_call = caller_env())
}
\arguments{
\item{arg}{A function argument. Must be a symbol.}

\item{error_arg}{An argument name as a string. This argument
will be mentioned in error messages as the input that is at the
origin of a problem.}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[=abort]{abort()}} for more information.}
}
\description{
Throws an informative error if \code{arg} is missing.
}
\examples{
f <- function(x)  {
  arg_require(x)
}

# Fails because `x` is not supplied
try(f())

# Succeeds
f(NULL)
}
\seealso{
\code{\link[=arg_match]{arg_match()}}
}
