% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arg.R
\name{arg_require}
\alias{arg_require}
\title{Check that argument is supplied
Throws an informative error if \code{arg} is missing.}
\usage{
arg_require(arg, error_call = caller_env())
}
\arguments{
\item{arg}{A function argument. Must be a symbol.}

\item{error_call}{An expression representing the context in which
the error occurred. If non-null, \code{abort()} displays the call
(stripped from its arguments to keep it simple) before \code{message}.

If you wrap an input-checking helper in another similar helper,
you should generally forward \code{error_call} from the calling frame
so that the relevant user expression is displayed when an error
occurs.}
}
\description{
Check that argument is supplied
Throws an informative error if \code{arg} is missing.
}
\examples{
f <- function(x)  {
  arg_require(x)
}

# Fails because `x` is not supplied
try(f())

# Succeeds
f(NULL)
}
\seealso{
\code{\link[=arg_match]{arg_match()}}
}
