% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnd-abort.R
\name{args_error_context}
\alias{args_error_context}
\title{Documentation anchor for error arguments}
\arguments{
\item{arg, error_arg}{An argument name as a string. This argument
will be mentioned in error messages as the input that is at the
origin of a problem.}

\item{call, error_call}{The execution environment of a currently
running function. When an error occurs, the corresponding
function call (see \code{\link[=sys.call]{sys.call()}}) is retrieved and stored as
\code{call} field in the error object to provide users with contextual
information about the error.

When you are writing an input checker function and need to pass
along \code{call}, you should generally use \code{caller_env()} as default,
e.g. \code{call = caller_env()} or \code{error_call = caller_env()}.

Can also be \code{NULL} or a function call to respectively disable the
contextual call or hard-code it.

See also \code{\link[=local_error_call]{local_error_call()}} for an alternative way of
providing this information.}
}
\description{
Use \verb{@inheritParams rlang::args_error_context} in your package to
document \code{arg} and \code{call} arguments (or equivalently their prefixed
versions \code{error_arg} and \code{error_call}).
\itemize{
\item \code{arg} parameters should be formatted as argument (e.g. using
cli's \code{.arg} specifier) and included in error messages. See also
\code{\link[=caller_arg]{caller_arg()}}.
\item \code{call} parameters should be included in error conditions in a
field named \code{call}. An easy way to do this is by passing a \code{call}
argument to \code{\link[=abort]{abort()}}. See also \code{\link[=local_error_call]{local_error_call()}}.
}
}
