% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-encoding.R
\name{as_bytes}
\alias{as_bytes}
\title{Coerce to a raw vector}
\usage{
as_bytes(x)
}
\arguments{
\item{x}{A string.}
}
\value{
A raw vector of bytes.
}
\description{
\Sexpr[results=rd, stage=render]{rlang:::lifecycle("experimental")}

This currently only works with strings, and returns its hexadecimal
representation.
}
\section{Life cycle}{


Raw vector functions are experimental.
}

\keyword{internal}
