% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{as_env}
\alias{as_env}
\alias{as_env.NULL}
\alias{as_env.character}
\alias{as_env.default}
\alias{as_env.environment}
\title{Coerce to an environment.}
\usage{
as_env(x, parent = NULL)

\method{as_env}{NULL}(x, parent = NULL)

\method{as_env}{environment}(x, parent = NULL)

\method{as_env}{character}(x, parent = NULL)

\method{as_env}{default}(x, parent = NULL)
}
\arguments{
\item{x}{An object to coerce.}

\item{parent}{A parent environment, \code{\link{env_empty}()} by
default. Can be ignored with a warning for methods where it does
not make sense to change the parent.}
}
\description{
This is a s3 generic. The default method coerces named vectors
(including lists) to an environment. It first checks that \code{x}
is a dictionary (see \code{\link{is_dictionary}()}). The method for
strings returns the corresponding package environment (see
\code{\link{env_package}()}).
}
\examples{
# Coerce a named vector to an environment:
env <- as_env(mtcars)

# By default it gets the empty environment as parent:
identical(env_parent(env), env_empty())


# With strings it is a handy shortcut for env_package():
as_env("base")
as_env("rlang")

# With NULL it returns the empty environment:
as_env(NULL)
}

