% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle-retired.R
\name{as_env}
\alias{as_env}
\title{Coerce to an environment}
\usage{
as_env(x, parent = NULL)
}
\description{
This function is soft-deprecated as it was renamed to
\code{\link[=as_environment]{as_environment()}} in rlang 0.2.0.
}
\keyword{internal}
