% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn.R
\name{as_function}
\alias{as_function}
\title{Coerce to function.}
\usage{
as_function(.f, ...)
}
\arguments{
\item{.f}{A function or formula.

If a \strong{function}, it is used as is.

If a \strong{formula}, e.g. \code{~ .x + 2}, it is converted to a function
with two arguments, \code{.x} or \code{.} and \code{.y}. This allows you to
create very compact anonymous functions with up to two inputs.}

\item{...}{Additional arguments passed on to methods. Currently
unused in rlang.}
}
\description{
This generic transforms objects to functions. It is especially
useful with formulas to create lambdas on the fly.
}
\examples{
f <- as_function(~ . + 1)
f(10)
}
