% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval-tidy.R, R/lifecycle-retired.R
\name{new_overscope}
\alias{new_overscope}
\alias{overscope_eval_next}
\alias{overscope_clean}
\alias{as_overscope}
\title{Convert to a data mask}
\usage{
new_overscope(bottom, top = NULL, enclosure = base_env())

overscope_eval_next(overscope, quo, env = base_env())

overscope_clean(overscope)

as_overscope(quo, data = NULL)
}
\arguments{
\item{bottom}{This is the environment (or the bottom of a set of
environments) containing definitions for overscoped symbols. The
bottom environment typically contains pronouns (like \code{.data})
while its direct parents contain the overscoping bindings. The
last one of these parents is the \code{top}.}

\item{top}{The top environment of the overscope. During tidy
evaluation, this environment is chained and rechained to lexical
enclosures of self-evaluating formulas (or quosures). This is the
mechanism that ensures hygienic scoping: the bindings in the
overscope have precedence, but the bindings in the dynamic
environment where the tidy quotes were created in the first place
are in scope as well. If \code{NULL} (the default), \code{bottom} is also the top of
the overscope.}

\item{enclosure}{The default enclosure. After a quosure is done
self-evaluating, the overscope is rechained to the default
enclosure.}

\item{overscope}{A valid overscope containing bindings for \code{~},
\code{.top_env} and \code{_F} and whose parents contain overscoped bindings
for tidy evaluation.}

\item{quo}{A \link{quosure}.}

\item{env}{The lexical enclosure in case \code{quo} is not a validly
scoped quosure. This is the \link[=base_env]{base environment} by
default.}

\item{data}{A data frame or named vector of masking data.}
}
\value{
A valid overscope: a child environment of \code{bottom}
containing the definitions enabling tidy evaluation
(self-evaluating quosures, formula-unguarding, ...).
}
\description{
This function has been soft-deprecated in rlang 0.2.0. Please use
\code{\link[=as_data_mask]{as_data_mask()}} instead.
}
\keyword{internal}
