% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle-retired.R
\name{as_overscope}
\alias{as_overscope}
\title{Convert to a data mask}
\usage{
as_overscope(quo, data = NULL)
}
\arguments{
\item{quo}{A \link{quosure}.}

\item{data}{A data frame or named vector of masking data.}
}
\description{
This function has been soft-deprecated in rlang 0.2.0. Please use
\code{\link[=as_data_mask]{as_data_mask()}} instead.
}
\keyword{internal}
