% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-node.R
\name{as_pairlist}
\alias{as_pairlist}
\title{Coerce to pairlist}
\usage{
as_pairlist(x)
}
\arguments{
\item{x}{An object to coerce.}
}
\description{
This transforms vector objects to a linked pairlist of nodes. See
\link{pairlist} for information about the pairlist type.
}
\seealso{
\link{pairlist}
}
