% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call.R
\name{call_args}
\alias{call_args}
\alias{call_args_lsp}
\alias{call_args_names}
\title{Extract arguments from a call}
\usage{
call_args(call = NULL)

call_args_lsp(call = NULL)

call_args_names(call = NULL)
}
\arguments{
\item{call}{Can be a call, a formula quoting a call in the
right-hand side, or a frame object from which to extract the call
expression. If not supplied, the calling frame is used.}
}
\value{
A named list of arguments. The \code{_lsp} version returns
  a named pairlist.
}
\description{
Extract arguments from a call
}
\examples{
call <- quote(f(a, b))

# Subsetting a call returns the arguments in a language pairlist:
call[-1]

# Whereas call_args() returns a list:
call_args(call)

# When the call arguments are supplied without names, a vector of
# empty strings is supplied (rather than NULL):
call_args_names(call)
}
\seealso{
\code{\link{fn_fmls}()} and
  \code{\link{fn_fmls_names}()}
}

