% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call.R
\name{call_fn}
\alias{call_fn}
\title{Extract function from a call}
\usage{
call_fn(call = NULL, env = NULL)
}
\arguments{
\item{call}{Can be a call, a formula quoting a call in the
right-hand side, or a frame object from which to extract the call
expression. If not supplied, the calling frame is used.}

\item{env}{Environment in which to look up call the function
definition and the contents of \code{...}. If not supplied, it is
retrieved from \code{call} if the latter is a frame object or a
formula.}
}
\description{
Extract function from a call
}
\examples{
# Extract from a quoted call:
call_fn(~matrix())
call_fn(quote(matrix()))

# Extract the calling function
test <- function() call_fn()
test()
}
\seealso{
\code{\link{call_fn_name}}()
}

