% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call.R
\name{call_fn}
\alias{call_fn}
\title{Function name of a call}
\usage{
call_fn(call)
}
\arguments{
\item{call}{A call from which to extract the function name.}
}
\value{
A string with the function name.
}
\description{
Function name of a call
}

