% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lang-call.R
\name{call_homogenise}
\alias{call_homogenise}
\title{Standardise a call against formal arguments.}
\usage{
call_homogenise(call = caller_frame(), dots_env = NULL, enum_dots = FALSE,
  add_missings = FALSE)
}
\arguments{
\item{call}{Can be a call, a formula quoting a call in the
right-hand side, or a frame object from which to extract the call
expression. If not supplied, the calling frame is used. Note that
\code{call_homogenise} needs access to the actual function
corresponding to the call. It will retrieve it from the tidy
quote or frame environment, or in the calling context.}

\item{dots_env}{Calling frame in which to look up call the contents
of \code{...}. If not supplied and \code{call} is a frame object,
it is retrieved from \code{call}.}

\item{enum_dots}{Whether to standardise the names of dotted
arguments. If \code{TRUE}, this produces calls such as
\code{f(..1 = ..2)} instead of \code{f(..2)}. The first form is
mostly intended for code analysis tools as it makes it easier to
climb arguments through nested calls. The latter form (the
default) is more faithful to the actual call and is ready to be
evaluated.}

\item{add_missings}{Whether to standardise missing arguments.}
}
\description{
This is a slower but more thorough version of
\code{\link{call_standardise}()}. Compared to
\code{\link{match.call}()}, \code{call_homogenise()}:
\itemize{
  \item Always report dotted arguments as such: \code{call(..1, ..2)}.
        In comparison, \code{match.call()} inlines literals:
        \code{call("foo", 10)}.
  \item Provides an argument \code{enum_dots} to impose
        enumerated names on dotted arguments. This produces
        \code{call(..1 = x, ..2 = ..3)} instead of
        \code{call(foo = x, ..3)}.
  \item Does not sort arguments according to the order of
        appearance in the function definition.
  \item Standardises missing arguments as well if you specify
        \code{add_missings}: \code{call(x = , y = , )}.
}
}
