% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack.R
\name{call_inspect}
\alias{call_inspect}
\title{Inspect a call}
\usage{
call_inspect(...)
}
\arguments{
\item{...}{Arguments to display in the returned call.}
}
\description{
This function is useful for quick testing and debugging when you
manipulate expressions and calls. It lets you check that a function
is called with the right arguments. This can be useful in unit
tests for instance. Note that this is just a simple wrapper around
\code{\link[base:match.call]{base::match.call()}}.
}
\examples{
call_inspect(foo(bar), "" \%>\% identity())
}
