% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lang-call.R
\name{call_modify}
\alias{call_modify}
\title{Modify the arguments of a call.}
\usage{
call_modify(.call = caller_frame(), ..., .args = list())
}
\arguments{
\item{.call}{Can be a call, a formula quoting a call in the
right-hand side, or a frame object from which to extract the call
expression. If not supplied, the calling frame is used.}

\item{..., .args}{Named or unnamed expressions (constants, names or
calls) used to modify the call. Use \code{NULL} to remove
arguments.}
}
\value{
A tidy quote if \code{.call} is a tidy quote, a call
  otherwise.
}
\description{
Modify the arguments of a call.
}
\examples{
call <- quote(mean(x, na.rm = TRUE))

# Modify an existing argument
call_modify(call, na.rm = FALSE)
call_modify(call, x = quote(y))

# Remove an argument
call_modify(call, na.rm = NULL)

# Add a new argument
call_modify(call, trim = 0.1)

# Add an explicit missing argument
call_modify(call, na.rm = quote(expr = ))

# Supply a list of new arguments with .args
newargs <- list(na.rm = NULL, trim = 0.1)
call_modify(call, .args = newargs)

# If the call is missing, the parent frame is used instead.
f <- function(bool = TRUE) call_modify(.args = list(bool = FALSE))
f()


# You can also modify a tidy quote inplace:
f <- ~matrix()
call_modify(f, quote(foo))
}
\seealso{
new_call
}
