% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lang-call.R
\name{call_name}
\alias{call_name}
\title{Extract function name of a call}
\usage{
call_name(call = caller_frame())
}
\arguments{
\item{call}{Can be a call, a formula quoting a call in the
right-hand side, or a frame object from which to extract the call
expression. If not supplied, the calling frame is used.}
}
\value{
A string with the function name, or \code{NULL} if the
  function is anonymous.
}
\description{
Extract function name of a call
}
\examples{
# Extract the function name from quoted calls:
call_name(~foo(bar))
call_name(quote(foo(bar)))

# The calling expression is used as default:
foo <- function(bar) call_name()
foo(bar)

# Namespaced calls are correctly handled:
call_name(~base::matrix(baz))

# Anonymous and subsetted functions return NULL:
call_name(~foo$bar())
call_name(~foo[[bar]]())
call_name(~foo()())
}
\seealso{
\code{\link{call_fn}}()
}
