% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{call_node}
\alias{call_node}
\title{Create a new call from components}
\usage{
call_node(car, cdr = NULL)
}
\arguments{
\item{car}{The head of the call. It should be a
\link[=is_callable]{callable} object: a symbol, call, or literal
function.}

\item{cdr}{The tail of the call, i.e. a \link[=node]{node list} of
arguments.}
}
\description{
Create a new call from components
}
\keyword{internal}
