% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call.R
\name{call_standardise}
\alias{call_standardise}
\title{Standardise a function call}
\usage{
call_standardise(call, env = parent.frame())
}
\arguments{
\item{call}{A call}

\item{env}{Environment in which to look up call value.}
}
\description{
Standardise a function call
}

