% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lang-call.R
\name{call_standardise}
\alias{call_standardise}
\title{Standardise a call.}
\usage{
call_standardise(call = caller_frame())
}
\arguments{
\item{call}{Can be a call, a formula quoting a call in the
right-hand side, or a frame object from which to extract the call
expression. If not supplied, the calling frame is used.}
}
\description{
This is essentially equivalent to \code{\link[base]{match.call}()},
but handles primitive functions more gracefully.
}
\seealso{
\code{\link{call_homogenise}()} for a version more
  suitable to language analysis.
}
