% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack.R
\name{caller_fn}
\alias{caller_fn}
\alias{current_fn}
\title{Get properties of the current or caller frame}
\usage{
caller_fn(n = 1)

current_fn()
}
\arguments{
\item{n}{The number of callers to go back.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
\itemize{
\item The current frame is the execution context of the function that
is currently being evaluated.
\item The caller frame is the execution context of the function that
called the function currently being evaluated.
}
}
\section{Life cycle}{

These functions are experimental.
}

\seealso{
\code{\link[=caller_env]{caller_env()}} and \code{\link[=current_env]{current_env()}}
}
\keyword{internal}
