% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack.R
\name{caller_fn}
\alias{caller_fn}
\alias{current_fn}
\title{Get properties of the current or caller frame}
\usage{
caller_fn(n = 1)

current_fn()
}
\arguments{
\item{n}{The number of callers to go back.}
}
\description{
\itemize{
\item \code{current_fn()} returns the function of the current frame.
\item \code{caller_fn()} returns the function of the calling frame.
}
}
\seealso{
\code{\link[=caller_env]{caller_env()}} and \code{\link[=current_env]{current_env()}}
}
