% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack.R
\name{caller_fn}
\alias{caller_fn}
\alias{current_fn}
\title{Get properties of the current or caller frame}
\usage{
caller_fn(n = 1)

current_fn()
}
\arguments{
\item{n}{The number of generations to go back.}
}
\description{
\Sexpr[results=rd, stage=render]{rlang:::lifecycle("experimental")}
\itemize{
\item The current frame is the execution context of the function that
is currently being evaluated.
\item The caller frame is the execution context of the function that
called the function currently being evaluated.
}

See the \link[=stack]{call stack} topic for more information.
}
\section{Life cycle}{


These functions are experimental.
}

\seealso{
\code{\link[=caller_env]{caller_env()}} and \code{\link[=current_env]{current_env()}}
}
