% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle-retired.R
\name{caller_frame}
\alias{caller_frame}
\title{Get caller frame}
\usage{
caller_frame(n = 1)
}
\arguments{
\item{n}{Number of frames to go back.}
}
\description{
\Sexpr[results=rd, stage=render]{rlang:::lifecycle("soft-deprecated")}
}
\keyword{internal}
