% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnd-abort.R
\name{cnd_as_unhandled_error}
\alias{cnd_as_unhandled_error}
\title{Create unhandled condition}
\usage{
cnd_as_unhandled_error(cnd)
}
\arguments{
\item{cnd}{A condition object.}
}
\value{
A condition with an embedded message.
}
\description{
Transform a classed condition into a simple condition.
\itemize{
\item The \code{conditionMessage()} method is run and the result is stored
in the \code{message} field so that it becomes a constant.
\item The \code{backtrace_on_error} global option is consulted to include a
backtrace or a \code{last_error()} reminder along with the message.
}

The resulting condition can be rethrown with \code{\link[=stop]{stop()}} or
\code{\link[=cnd_signal]{cnd_signal()}}.

Note that this function is for internal purposes. The caller is
responsible for making sure that there is no error catching
handlers on the stack.

The condition \emph{currently} inherits from \code{"rlang_error"} (but not
\code{"error"}) to prevent \code{\link[=entrace]{entrace()}} from repeatedly embedding a
backtrace.
}
\keyword{internal}
