% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnd.R
\name{cnd_message}
\alias{cnd_message}
\alias{cnd_issue}
\alias{cnd_bullets}
\title{Build an error message from a main issue and bullet messages}
\usage{
cnd_message(cnd)

cnd_issue(cnd, ...)

cnd_bullets(cnd, ...)
}
\arguments{
\item{cnd}{A condition object.}
}
\description{
\Sexpr[results=rd, stage=render]{rlang:::lifecycle("experimental")}

\code{cnd_message()} is meant to be used in \code{conditionMessage()}
methods. The error class should implement:
\itemize{
\item A \code{cnd_issue()} method returning a single, bullet-less line.
\item A \code{cnd_bullets()} method returning a list of issues or
informative statements, prefixed with bullets.
}
}
\keyword{internal}
