% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnd.R
\name{cnd_message}
\alias{cnd_message}
\alias{cnd_issue}
\alias{cnd_bullets}
\title{Build an error message from a main issue and bullet messages}
\usage{
cnd_message(cnd)

cnd_issue(cnd, ...)

cnd_bullets(cnd, ...)
}
\arguments{
\item{cnd}{A condition object.}
}
\description{
\Sexpr[results=rd, stage=render]{rlang:::lifecycle("experimental")}

\code{cnd_message()} is meant to be used in \code{conditionMessage()}
methods. The error class should implement:
\itemize{
\item A \code{cnd_issue()} method returning a single line.
\item A \code{cnd_bullets()} method returning a named vector of lines. These
lines are automatically prefixed with a bullet by \code{cnd_message()}.
}

The bullets are formatted depending on the names of the character
vector:
\itemize{
\item Elements named \code{"i"} are prefixed with a blue "info" symbol.
\item Elements named \code{"x"} are prefixed with a red "cross" symbol.
\item Unnamed elements are prefixed with a "*" symbol.
}

\code{cnd_message()} is automatically called by the \code{conditionMessage()}
for rlang errors. Errors thrown with \code{\link[=abort]{abort()}} can implement
\code{cnd_issue()} and \code{cnd_bullets()} right away.
}
\section{Overriding \code{cnd_bullets()}}{


Sometimes the generation of an error message depends on the state
of the type checking. In that case, it can be tricky to lazily
generate error messages with \code{cnd_bullets()}: you can either
overspecify your error class hierarchies with one class per state,
or replicate the type-checking control flow within the
\code{cnd_bullets()} method. None of these options are ideal.

To work around this, you can define a \code{.bullets} field in your
error object. This should be a function (or a lambda-formula which
will be passed to \code{\link[=as_function]{as_function()}}) with the same signature as
\code{cnd_bullets()} methods. This function overrides the
\code{cnd_bullets()} generic and can generate an error message tailored
to the state in which the error was constructed.

Note that as a rule, \code{cnd_issue()} should be a general thematic
issues that does not depend on state, and so does not need to be
overridden.
}

\keyword{internal}
