% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnd.R
\name{cnd_type}
\alias{cnd_type}
\title{What type is a condition?}
\usage{
cnd_type(cnd)
}
\arguments{
\item{cnd}{A condition object.}
}
\value{
A string, either \code{"condition"}, \code{"message"}, \code{"warning"},
\code{"error"} or \code{"interrupt"}.
}
\description{
Use \code{cnd_type()} to check what type a condition is.
}
\examples{
cnd_type(catch_cnd(abort("Abort!")))
cnd_type(catch_cnd(interrupt()))
}
