% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make-call.R
\name{common_env}
\alias{common_env}
\title{Find common environment in list of lazy objects.}
\usage{
common_env(dots)
}
\arguments{
\item{dots}{A list of lazy objects}
}
\description{
If no common environment is found, will return \code{baseenv()}.
}
\examples{
common_env(lazy_dots(a, b, c))

f <- function(x) ~x
common_env(list(f(1)))
common_env(list(f(1), f(2)))
}
\keyword{internal}

