% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle-retired.R
\name{deprecated-cnd}
\alias{deprecated-cnd}
\alias{new_cnd}
\alias{cnd_error}
\alias{cnd_warning}
\alias{cnd_message}
\title{Deprecated condition constructors}
\usage{
new_cnd(.type = NULL, ..., .msg = NULL)

cnd_error(.type = NULL, ..., .msg = NULL)

cnd_warning(.type = NULL, ..., .msg = NULL)

cnd_message(.type = NULL, ..., .msg = NULL)
}
\arguments{
\item{.type}{The condition subclass.}

\item{...}{Named data fields stored inside the condition
object. These dots are evaluated with \link[=tidy-dots]{explicit
splicing}.}

\item{.msg}{A default message to inform the user about the
condition when it is signalled.}
}
\description{
These functions were deprecated in rlang 0.2.0 to follow the
convention that return types are indicated as suffixes. Please use
\code{\link[=cnd]{cnd()}}, \code{\link[=error_cnd]{error_cnd()}}, \code{\link[=warning_cnd]{warning_cnd()}} and \code{\link[=message_cnd]{message_cnd()}}
instead.
}
\keyword{internal}
