% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dots.R
\name{dots_definitions}
\alias{dots_definitions}
\title{Capture definition objects}
\usage{
dots_definitions(..., .named = FALSE, .ignore_empty = c("trailing", "none",
  "all"))
}
\arguments{
\item{...}{For \code{enexprs()}, \code{ensyms()} and \code{enquos()}, names of
arguments to capture without evaluation (including \code{...}). For
\code{exprs()} and \code{quos()}, the expressions to capture unevaluated
(including expressions contained in \code{...}).}

\item{.named}{Whether to ensure all dots are named. Unnamed
elements are processed with \code{\link[=expr_text]{expr_text()}} to figure out a default
name. If an integer, it is passed to the \code{width} argument of
\code{expr_text()}, if \code{TRUE}, the default width is used. See
\code{\link[=exprs_auto_name]{exprs_auto_name()}}.}

\item{.ignore_empty}{Whether to ignore empty arguments. Can be one
of \code{"trailing"}, \code{"none"}, \code{"all"}. If \code{"trailing"}, only the
last argument is ignored if it is empty.}
}
\description{
Capture definition objects
}
\section{Life cycle}{


\code{dots_definitions()} is experimental. Expect API changes.
}

\keyword{internal}
