% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dots.R
\name{dots_inspect}
\alias{dots_inspect}
\alias{dots_inspect_}
\title{Inspect dots}
\usage{
dots_inspect(...)

dots_inspect_(dots, stack)
}
\arguments{
\item{...}{Dots to inspect.}

\item{dots}{Dots to inspect.}

\item{stack}{A \code{call_stack} object as returned by
\code{\link{call_stack}()}.}
}
\description{
Runs \code{\link{arg_inspect}()} for each dots element, and return the
results in a list.
}
\details{
\code{dots_inspect_()} is the standard evaluation version of
\code{dots_inspect()} and takes a list of dots as captured by
\code{\link{frame_dots}()} or \code{\link{arg_dots}()}, and a call
stack as returned by \code{\link{call_stack}()}.
}
\seealso{
\code{\link{arg_inspect}()}
}

