% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dots.R
\name{dots_patterns}
\alias{dots_patterns}
\alias{dots_patterns}
\alias{dots_formulas}
\alias{dots_definitions}
\title{Quote patterns, definitions or formulas}
\usage{
dots_patterns(..., .named = FALSE, .ignore_empty = c("trailing", "none",
  "all"))

dots_formulas(..., .named = FALSE, .ignore_empty = c("trailing", "none",
  "all"))

dots_definitions(..., .named = FALSE, .ignore_empty = c("trailing", "none",
  "all"))
}
\arguments{
\item{...}{Expressions to capture unevaluated.}

\item{.named}{Whether to ensure all dots are named. Unnamed
elements are processed with \code{\link[=expr_text]{expr_text()}} to figure out a default
name. If an integer, it is passed to the \code{width} argument of
\code{expr_text()}, if \code{TRUE}, the default width is used. See
\code{\link[=exprs_auto_name]{exprs_auto_name()}}.}

\item{.ignore_empty}{Whether to ignore empty arguments. Can be one
of \code{"trailing"}, \code{"none"}, \code{"all"}. If \code{"trailing"}, only the
last argument is ignored if it is empty.}
}
\value{
A named list of list of extracted LHS and RHS
components. The components are extracted as quosures in a list
with \code{lhs} and \code{rhs} fields. The outer list is thus named after
the dots inputs while the inner lists contain the components for
each input.
}
\description{
These quoting operators take dots and expect two-sided patterns
like \code{a ~ b} or \code{a := b}. The LHS and RHS of each input are
extracted in a list of quosures. The operators differ in the type
of inputs they accept:
\itemize{
\item \code{dots_patterns()} accepts both two-sided formulas and \code{:=}
definitions. There is no distinction between either type of
inputs. This lets your users choose which syntax they prefer.
\item \code{dots_formulas()} and \code{dots_definitions()} accept only one type
of inputs.
}
}
\examples{
dots_formulas(
  a ~ b,
  c ~ d
)

dots_definitions(
  a := b,
  c := d
)

# dots_patterns() supports both kinds of inputs at the same time
# but this should be discouraged. Users should opt for one type and
# stick to it consistently:
dots_patterns(
  a ~ b,
  c := d
)
}
\seealso{
\code{\link[=quos]{quos()}}
}
