% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quosure.R
\name{is_empty_quosure}
\alias{is_empty_quosure}
\title{Is a quosure empty?}
\usage{
is_empty_quosure(quo)
}
\arguments{
\item{quo}{A quosure.}
}
\description{
When missing arguments are captured as quosures, either through
\code{\link[=catch_quosure]{catch_quosure()}} or \code{\link[=dots_quosures]{dots_quosures()}}, they are returned as the
empty quosure. These quosures contain the \link[=missing_arg]{missing
argument} and typically have the \link[=empty_env]{empty
environment} as enclosure.
}
\examples{
quo <- quosure()
is_empty_quosure(quo)
is_missing(f_rhs(quo))
}
