% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-sym.R
\name{ensym}
\alias{ensym}
\title{Capture a symbol}
\usage{
ensym(arg)
}
\arguments{
\item{arg}{A symbol referring to an argument. The expression
supplied to that argument will be captured unevaluated.}
}
\description{
Like \code{\link[=enexpr]{enexpr()}} and \code{\link[=enquo]{enquo()}}, \code{ensym()} makes an argument
auto-quoting. It always returns a symbol and issues an error if the
input is not a string or a symbol. It supports \link{quasiquotation}.
}
\examples{
quote_sym <- function(arg) ensym(arg)
quote_sym(foo)
quote_sym("foo")

# The ensymed argument is quasiquoted:
var <- "foo"
quote_sym(var)
quote_sym(!!var)
}
