% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle-retired.R
\name{env_bind_exprs}
\alias{env_bind_exprs}
\alias{env_bind_fns}
\title{Bind a promise or active binding}
\usage{
env_bind_exprs(.env, ..., .eval_env = caller_env())

env_bind_fns(.env, ...)
}
\arguments{
\item{.env}{An environment.}

\item{...}{Pairs of names and expressions, values or functions.
These dots support \link[=tidy-dots]{tidy dots} features.}

\item{.eval_env}{The environment where the expressions will be
evaluated when the symbols are forced.}
}
\description{
\Sexpr[results=rd, stage=render]{rlang:::lifecycle("soft-deprecated")}

As of rlang 0.3.0, \code{env_bind_exprs()} and \code{env_bind_fns()} have
been renamed to \code{\link[=env_bind_lazy]{env_bind_lazy()}} and \code{\link[=env_bind_active]{env_bind_active()}} for
consistency.
}
\keyword{internal}
