% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env-binding.R
\name{env_binding_are_active}
\alias{env_binding_are_active}
\alias{env_binding_are_lazy}
\title{What kind of environment binding?}
\usage{
env_binding_are_active(env, nms = NULL)

env_binding_are_lazy(env, nms = NULL)
}
\arguments{
\item{env}{An environment.}

\item{nms}{Names of bindings. Defaults to all bindings in \code{env}.}
}
\value{
A logical vector as long as \code{nms} and named after it.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\keyword{internal}
