% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{env_caller}
\alias{env_caller}
\title{Get the environment of the caller frame.}
\usage{
env_caller(n = 1)
}
\arguments{
\item{n}{The number of generation to go back. Note that contrarily
to \code{\link{call_frame}()}, 1 represents the parent frame
rather than the current frame.}
}
\description{
This is a shortcut for \code{\link{call_frame}(2)$env}.
}
\seealso{
\code{\link{call_frame}()}
}

