% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{env_inherits}
\alias{env_inherits}
\title{Does environment inherit from another environment?}
\usage{
env_inherits(env, ancestor)
}
\arguments{
\item{env}{An environment.}

\item{ancestor}{Another environment from which \code{x} might inherit.}
}
\description{
This returns \code{TRUE} if \code{x} has \code{ancestor} among its parents.
}
