% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{env_names}
\alias{env_names}
\title{List object names in an environment.}
\usage{
env_names(env = caller_env())
}
\arguments{
\item{env}{An environment or an object bundling an environment:
either a \link{quosure} or a \link[=is_closure]{closure}. If
missing, the environment of the current evaluation frame is
returned.}
}
\value{
A character vector of object names.
}
\description{
\code{env_names()} returns object names from an enviroment \code{env} as a character
vector.
All names are returned, even those starting with a dot.
}
\examples{
env_names()
}
