% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{env_namespace}
\alias{env_imports}
\alias{env_namespace}
\title{Get the namespace of a package.}
\usage{
env_namespace(pkg = NULL)

env_imports(pkg = NULL)
}
\arguments{
\item{pkg}{The name of a package. If \code{NULL}, the surrounding
namespace is returned, or an error is issued if not called within
a namespace.}
}
\description{
Namespaces are the environment where all the functions of a package
live. The parent environments of namespaces are the \code{imports}
environments, which contain all the functions imported from other
packages.
}
\seealso{
\code{\link{env_package}()}
}

