% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env.R
\name{env_unlock}
\alias{env_unlock}
\title{Unlock an environment}
\usage{
env_unlock(env)
}
\arguments{
\item{env}{An environment.}
}
\value{
Whether the environment has been unlocked.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}. This function is now defunct
because recent versions of R no longer make it possible to
unlock an environment.
}
\keyword{internal}
