% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnd-abort.R
\name{error_call}
\alias{error_call}
\title{Validate a function call for use in error messages}
\usage{
error_call(call)
}
\arguments{
\item{call, error_call}{An expression (as returned by e.g.
\code{sys.call()}) representing the context in which the error
occurred. If non-null, the call is stripped from its arguments to
keep it simple.

Can also be an execution environment of a currently running
function (as returned by e.g. \code{parent.frame()}). The
corresponding call is then retrieved.}
}
\description{
Creates a function call ready to be used as \code{call} field of error
conditions. This field is displayed by \code{\link[=stop]{stop()}} and \code{\link[=abort]{abort()}} to
give context to an error message.
}
\seealso{
\code{\link[=error_arg]{error_arg()}}
}
