% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle-retired.R
\name{exiting}
\alias{exiting}
\title{Exiting handler}
\usage{
exiting(handler)
}
\description{
\Sexpr[results=rd, stage=render]{rlang:::lifecycle("soft-deprecated")}

\code{exiting()} is no longer necessary as handlers are exiting by default.
}
\keyword{internal}
