% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quotation.R
\name{exprs_auto_name}
\alias{exprs_auto_name}
\alias{quos_auto_name}
\title{Ensure that list of expressions are all named}
\usage{
exprs_auto_name(exprs, width = 60L, printer = expr_text)

quos_auto_name(quos, width = 60L)
}
\arguments{
\item{exprs}{A list of expressions.}

\item{width}{Maximum width of names.}

\item{printer}{A function that takes an expression and converts it
to a string. This function must take an expression as first
argument and \code{width} as second argument.}

\item{quos}{A list of quosures.}
}
\description{
This gives default names to unnamed elements of a list of
expressions (or expression wrappers such as formulas or
quosures). \code{exprs_auto_name()} deparses the expressions with
\code{\link[=expr_text]{expr_text()}} by default. \code{quos_auto_name()} deparses with
\code{\link[=quo_text]{quo_text()}}.
}
