% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nse-defuse.R
\name{exprs_auto_name}
\alias{exprs_auto_name}
\alias{quos_auto_name}
\title{Ensure that all elements of a list of expressions are named}
\usage{
exprs_auto_name(
  exprs,
  ...,
  repair_auto = c("minimal", "unique"),
  repair_quiet = FALSE
)

quos_auto_name(quos)
}
\arguments{
\item{exprs}{A list of expressions.}

\item{...}{These dots are for future extensions and must be empty.}

\item{repair_auto}{Whether to repair the automatic names. By
default, minimal names are returned. See \code{?vctrs::vec_as_names}
for information about name repairing.}

\item{repair_quiet}{Whether to inform user about repaired names.}

\item{quos}{A list of quosures.}
}
\description{
This gives default names to unnamed elements of a list of
expressions (or expression wrappers such as formulas or
quosures). \code{exprs_auto_name()} deparses the expressions with
\code{\link[=expr_name]{expr_name()}} by default. \code{quos_auto_name()} deparses with
\code{\link[=quo_name]{quo_name()}}.
}
