% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quotation.R
\name{exprs_auto_name}
\alias{exprs_auto_name}
\alias{quos_auto_name}
\title{Ensure that all elements of a list of expressions are named}
\usage{
exprs_auto_name(exprs, width = NULL, printer = NULL)

quos_auto_name(quos, width = NULL)
}
\arguments{
\item{exprs}{A list of expressions.}

\item{width}{Deprecated. Maximum width of names.}

\item{printer}{Deprecated. A function that takes an expression
and converts it to a string. This function must take an
expression as the first argument and \code{width} as the second
argument.}

\item{quos}{A list of quosures.}
}
\description{
This gives default names to unnamed elements of a list of
expressions (or expression wrappers such as formulas or
quosures). \code{exprs_auto_name()} deparses the expressions with
\code{\link[=expr_name]{expr_name()}} by default. \code{quos_auto_name()} deparses with
\code{\link[=quo_name]{quo_name()}}.
}
