% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{f_rhs}
\alias{f_rhs}
\title{Extract the right-hand-side of a formula}
\usage{
f_rhs(f)
}
\arguments{
\item{f}{A formula}
}
\value{
An atomic vector of length 1, a name, or a call.
}
\description{
Throws an error if \code{f} is not a formula.
}
\examples{
f_rhs(~ 1 + 2 + 3)
f_rhs(~ x)
f_rhs(~ "A")
f_rhs(1 ~ 2)
}

