% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dots.R
\name{frame_dots}
\alias{arg_dots}
\alias{arg_dots_lsp}
\alias{frame_dots}
\alias{frame_dots_lsp}
\title{Extract dots}
\usage{
frame_dots(frame = NULL)

frame_dots_lsp(frame = NULL)

arg_dots(...)

arg_dots_lsp(...)
}
\arguments{
\item{frame}{The environment from which the dots should be
retrieved. Can be a frame, an environment, or a formula from
which to retrieve an environment. If not supplied, the calling
frame is used.}

\item{...}{Arguments to extract. Can be both forwarded dots and
direct arguments.}
}
\description{
\code{frame_dots()} extracts dots from a frame and
\code{arg_dots()} extracts dots from its arguments. The
\code{_lsp()} versions return a pairlist that is ready to be
spliced into a call, while the regular versions return a regular
list that is usually easier to work with.
}
\details{
\code{frame_dots()} and \code{frame_dots_lsp()} never fail, even if
the frame does not contain dots. Instead they return an empty list
or \code{NULL} respectively.
}

