% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle-deprecated.R
\name{friendly_type}
\alias{friendly_type}
\title{Format a type for error messages}
\usage{
friendly_type(type)
}
\arguments{
\item{type}{A type as returned by \code{\link[=typeof]{typeof()}}.}
}
\value{
A string of the prettified type, qualified with an
indefinite article.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{friendly_type()} is deprecated. Please use the
\code{standalone-obj-type.R} file instead. You can import it
in your package with \code{usethis::use_standalone("r-lib/rlang", "obj-type")}.
}
\keyword{internal}
