% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{friendly_type}
\alias{friendly_type}
\title{Format a type for error messages}
\usage{
friendly_type(type)
}
\arguments{
\item{type}{A type as returned by \code{\link[=type_of]{type_of()}} or \code{\link[=lang_type_of]{lang_type_of()}}.}
}
\value{
A string of the prettified type, qualified with an
indefinite article.
}
\description{
Format a type for error messages
}
\examples{
friendly_type("logical")
friendly_type("integer")
friendly_type("string")
}
