% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{is_env}
\alias{is_env}
\alias{is_bare_env}
\title{Is object an environment?}
\usage{
is_env(x)

is_bare_env(x)
}
\arguments{
\item{x}{object to test}
}
\description{
\code{is_bare_env()} tests whether \code{x} is an environment without a s3 or
s4 class.
}
