% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle-retired.R
\name{is_expr}
\alias{is_expr}
\title{Is an object an expression?}
\usage{
is_expr(x)
}
\arguments{
\item{x}{An object to test.}
}
\description{
This function was soft-deprecated and renamed to \code{\link[=is_expression]{is_expression()}}
in rlang 0.2.0. This is for consistency with other type predicates
which are not abbreviated.
}
\keyword{internal}
