% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle-retired.R
\name{is_frame}
\alias{is_frame}
\title{Is object a frame?}
\usage{
is_frame(x)
}
\arguments{
\item{x}{Object to test}
}
\description{
\Sexpr[results=rd, stage=render]{rlang:::lifecycle("soft-deprecated")}
}
\keyword{internal}
