% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env-special.R
\name{is_installed}
\alias{is_installed}
\title{Is a package installed in the library?}
\usage{
is_installed(pkg)
}
\arguments{
\item{pkg}{The name of a package.}
}
\value{
\code{TRUE} if the package is installed, \code{FALSE} otherwise.
}
\description{
This checks that a package is installed with minimal side effects.
If installed, the package will be loaded but not attached.
}
\examples{
is_installed("utils")
is_installed("ggplot5")
}
