% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env-special.R
\name{is_installed}
\alias{is_installed}
\title{Are packages installed in any of the libraries?}
\usage{
is_installed(pkg)
}
\arguments{
\item{pkg}{The package names.}
}
\value{
\code{TRUE} if \emph{all} package names provided in \code{pkg} are installed,
\code{FALSE} otherwise.
}
\description{
This checks that packages are installed with minimal side effects.
If installed, the packages will be loaded but not attached.
}
\examples{
is_installed("utils")
is_installed(c("base", "ggplot5"))
}
