% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call.R, R/lifecycle-retired.R
\name{is_unary_call}
\alias{is_unary_call}
\alias{is_binary_call}
\alias{is_lang}
\alias{is_unary_lang}
\alias{is_binary_lang}
\title{Is object a call?}
\usage{
is_unary_call(x, name = NULL, ns = NULL)

is_binary_call(x, name = NULL, ns = NULL)

is_lang(x, name = NULL, n = NULL, ns = NULL)

is_unary_lang(x, name = NULL, ns = NULL)

is_binary_lang(x, name = NULL, ns = NULL)
}
\arguments{
\item{x}{An object to test. If a formula, the right-hand side is
extracted.}

\item{name}{An optional name that the call should match. It is
passed to \code{\link[=sym]{sym()}} before matching. This argument is vectorised
and you can supply a vector of names to match. In this case,
\code{is_call()} returns \code{TRUE} if at least one name matches.}

\item{ns}{The namespace of the call. If \code{NULL}, the namespace
doesn't participate in the pattern-matching. If an empty string
\code{""} and \code{x} is a namespaced call, \code{is_call()} returns
\code{FALSE}. If any other string, \code{is_call()} checks that \code{x} is
namespaced within \code{ns}.}

\item{n}{An optional number of arguments that the call should
match.}
}
\description{
These functions are soft-deprecated, please use \code{\link[=is_call]{is_call()}},
\code{\link[=is_unary_call]{is_unary_call()}}, and \code{\link[=is_binary_call]{is_binary_call()}} instead.
}
\keyword{internal}
