% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{is_pairlist}
\alias{is_pairlist}
\alias{is_node}
\alias{is_node_list}
\title{Is object a node or pairlist?}
\usage{
is_pairlist(x)

is_node(x)

is_node_list(x)
}
\arguments{
\item{x}{Object to test.}
}
\description{
\itemize{
\item \code{is_pairlist()} checks that \code{x} has type \code{pairlist}.
\item \code{is_node()} checks that \code{x} has type \code{pairlist} or \code{language}.
It tests whether \code{x} is a node that has a CAR and a CDR,
including callable nodes (language objects).
\item \code{is_node_list()} checks that \code{x} has type \code{pairlist} or \code{NULL}.
\code{NULL} is the empty node list.
}
}
\section{Life cycle}{


These functions are experimental. We are still figuring out a good
naming convention to refer to the different lisp-like lists in R.
}

\seealso{
\code{\link[=is_call]{is_call()}} tests for language nodes.
}
\keyword{internal}
