% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quosure.R
\name{is_quosure}
\alias{is_quosure}
\alias{is_quosureish}
\title{Is an object a quosure or quosure-like?}
\usage{
is_quosure(x, scoped = NULL)

is_quosureish(x, scoped = NULL)
}
\arguments{
\item{x}{An object to test.}

\item{scoped}{A boolean indicating whether the quosure or formula
is scoped, that is, has a valid environment attribute. If \code{NULL},
the scope is not inspected.}
}
\description{
These predicates test for \link[=new_quosure]{quosure} objects.
\itemize{
\item \code{is_quosure()} tests for the canonical R quosure: the one-sided
"formula".
\item \code{is_quosureish()} tests for general R quosure objects: quosures,
two-sided formulas, and \link[=op-definition]{definitions}.
}
}
\details{
The \code{scoped} argument patterns-match on whether the scoped bundled
with the quosure is valid or not. Invalid scopes may happen in
nested quotations like \code{~~expr}, where the outer quosure is validly
scoped but not the inner one. This is because \code{~} saves the
environment when it is evaluated, and quoted quosures are by
definition not evaluated. Note that in the \link[=tidy_eval]{tidy evaluation
framework}, unscoped quosures are automatically given
the environment of the outer quosure during the evaluation process.
}
\examples{
# Degenerate quosures are often created by quoting, since `~`
# records the environment when it is evaluated the first time:
f <- ~~expr

# The outer quosure has been evaluated and is scoped:
is_quosure(f, scoped = TRUE)

# But the inner formula is not:
inner_f <- f_rhs(f)
is_quosure(inner_f, scoped = TRUE)


# Formulas and definitions are not quosures:
is_quosure(a := b)
is_quosure(a ~ b)

# But they are quosureish objects:
is_quosureish(a := b)
is_quosureish(a ~ b)
}
\seealso{
\link[=new_quosure]{as_quosure()} and \code{\link[=new_quosure]{new_quosure()}} for creating
quosures, and \code{\link[=tidy_quote]{tidy_quote()}} or \code{\link[=tidy_eval]{tidy_eval()}} for information
about the role of quosures in the tidy evaluation framework.
}
