% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quo.R
\name{is_quosure}
\alias{is_quosure}
\title{Is an object a quosure or quosure-like?}
\usage{
is_quosure(x)
}
\arguments{
\item{x}{An object to test.}
}
\description{
This predicate tests that an object is a \link{quosure}.
}
\section{Life cycle}{

\itemize{
\item \code{is_quosure()} is stable.
\item \code{is_quosureish()} is deprecated as of rlang 0.2.0. This function
assumed that quosures are formulas which is currently true but
might not be in the future.
}
}

\examples{
is_quosure(quo(foo))
}
