% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle-retired.R, R/quo.R
\name{is_quosureish}
\alias{is_quosureish}
\alias{is_quosure}
\title{Is an object a quosure or quosure-like?}
\usage{
is_quosureish(x, scoped = NULL)

is_quosure(x)
}
\arguments{
\item{x}{An object to test.}

\item{scoped}{A boolean indicating whether the quosure is scoped,
that is, has a valid environment attribute. If \code{NULL}, the scope
is not inspected.}
}
\description{
This predicate tests that an object is a \link{quosure}.
}
\section{Life cycle}{

\itemize{
\item \code{is_quosure()} is stable.
\item \code{is_quosureish()} is deprecated as of rlang 0.2.0. This function
assumed that quosures are formulas which is currently true but
might not be in the future.
}
}

\examples{
is_quosure(quo(foo))
}
