% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle-retired.R
\name{is_quosureish}
\alias{is_quosureish}
\alias{as_quosureish}
\title{Test for or coerce to quosure-like objects}
\usage{
is_quosureish(x, scoped = NULL)

as_quosureish(x, env = caller_env())
}
\arguments{
\item{x}{An object to test.}

\item{scoped}{A boolean indicating whether the quosure is scoped,
that is, has a valid environment attribute. If \code{NULL}, the scope
is not inspected.}

\item{env}{The environment in which the expression should be
evaluated. Only used for symbols and calls. This should typically
be the environment in which the expression was created.}
}
\description{
\Sexpr[results=rd, stage=render]{rlang:::lifecycle("defunct")}

These functions are deprecated as of rlang 0.2.0 because they make
the assumption that quosures are a subtype of formula, which we are
now considering to be an implementation detail.
}
\keyword{internal}
