% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lang-call.R
\name{lang}
\alias{lang}
\title{Create a call by "hand"}
\usage{
lang(.fn, ..., .args = list())
}
\arguments{
\item{.fn}{Function to call. For \code{make_call}, either a string, a
symbol or a quoted call. For \code{do_call}, a bare function name or
call.}

\item{..., .args}{Arguments to the call either in or out of a list}
}
\description{
Create a call by "hand"
}
\examples{
# fn can either be a string, a symbol or a call
lang("f", a = 1)
lang(quote(f), a = 1)
lang(quote(f()), a = 1)

#' Can supply arguments individually or in a list
lang(quote(f), a = 1, b = 2)
lang(quote(f), .args = list(a = 1, b = 2))
}
\seealso{
lang_modify
}
