% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle-retired.R
\name{lang}
\alias{lang}
\alias{new_language}
\title{Create a call}
\usage{
lang(.fn, ..., .ns = NULL)

new_language(head, tail = NULL)
}
\arguments{
\item{.fn}{Function to call. Must be a callable object: a string,
symbol, call, or a function.}

\item{...}{<\link[=dyn-dots]{dynamic}> Arguments for the function
call. Empty arguments are preserved.}

\item{.ns}{Namespace with which to prefix \code{.fn}. Must be a string
or symbol.}
}
\description{
\Sexpr[results=rd, stage=render]{rlang:::lifecycle("deprecated")}

These functions are deprecated, please use \code{\link[=call2]{call2()}} and
\code{\link[=new_call]{new_call()}} instead.
}
\keyword{internal}
