% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lang-call.R
\name{lang_args}
\alias{lang_args}
\alias{lang_args_names}
\title{Extract arguments from a call}
\usage{
lang_args(call = caller_frame())

lang_args_names(call = caller_frame())
}
\arguments{
\item{call}{Can be a call, a formula quoting a call in the
right-hand side, or a frame object from which to extract the call
expression. If not supplied, the calling frame is used.}
}
\value{
A named list of arguments. The \code{lang_tail()} version
  returns a named pairlist (the CDR of the call).
}
\description{
Extract arguments from a call
}
\examples{
call <- quote(f(a, b))

# Subsetting a call returns the arguments in a language pairlist:
call[-1]

# Whereas lang_args() returns a list:
lang_args(call)

# When the call arguments are supplied without names, a vector of
# empty strings is supplied (rather than NULL):
lang_args_names(call)
}
\seealso{
\code{\link{fn_fmls}()} and
  \code{\link{fn_fmls_names}()}
}
