% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle-retired.R
\name{lang_head}
\alias{lang_head}
\alias{lang_tail}
\title{Return the head or tail of a call}
\usage{
lang_head(lang)

lang_tail(lang)
}
\arguments{
\item{lang}{A call.}
}
\description{
As of rlang 0.2.0 these functions are retired (soft-deprecated for
now) because they are low level accessors that are rarely needed
for end users.
}
