% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle-retired.R
\name{lang_standardise}
\alias{lang_standardise}
\title{Standardise a call}
\usage{
lang_standardise(lang)
}
\arguments{
\item{lang}{The \code{call} argument of \code{\link[=call_standardise]{call_standardise()}}.}
}
\description{
This function is soft-deprecated, please use \code{\link[=call_standardise]{call_standardise()}}
instead.
}
\keyword{internal}
