% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expr-lang.R
\name{lang_standardise}
\alias{lang_standardise}
\title{Standardise a call}
\usage{
lang_standardise(lang)
}
\arguments{
\item{lang}{Can be a call (language object), a formula quoting a
call in the right-hand side, or a frame object from which to
extract the call expression.}
}
\value{
A quosure if \code{.lang} is a quosure, a raw call otherwise.
}
\description{
This is essentially equivalent to \code{\link[base:match.call]{base::match.call()}}, but with
better handling of primitive functions.
}
