% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnd-last.R
\name{last_error}
\alias{last_error}
\alias{last_trace}
\title{Last \code{abort()} error}
\usage{
last_error()

last_trace()
}
\description{
\itemize{
\item \code{last_error()} returns the last error entraced by \code{\link[=abort]{abort()}} or
\code{\link[=global_entrace]{global_entrace()}}. The error is printed with a backtrace in
simplified form.
\item \code{last_trace()} is a shortcut to return the backtrace stored in
the last error. This backtrace is printed in full form.
}
}
\seealso{
\itemize{
\item \code{\link{rlang_backtrace_on_error}} to control what is displayed when an
error is thrown.
\item \code{\link[=global_entrace]{global_entrace()}} to enable \code{last_error()} logging for all errors.
\item \code{\link[=last_warnings]{last_warnings()}} and \code{\link[=last_messages]{last_messages()}}.
}
}
