% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnd-last.R
\name{last_warnings}
\alias{last_warnings}
\alias{last_messages}
\alias{last_warning}
\alias{last_message}
\title{Display last warnings}
\usage{
last_warnings()

last_messages()

last_warning()

last_message()
}
\description{
\itemize{
\item \code{last_warnings()} and \code{last_messages()} return a list of all
warnings and messages that occurred during the last top-level R
command.
\item \code{last_warning()} and \code{last_message()} return the very last ones.
}

By default the warnings and messages are printed with a simplified
backtrace, like \code{\link[=last_error]{last_error()}}. Use \code{summary()} to print the
conditions with a full backtrace.
}
\examples{
if (FALSE) {
  global_entrace()
  f <- function() { warning("foo"); g() }
  g <- function() { warning("bar", immediate. = TRUE); h() }
  h <- function() warning("baz")

  f()

  # Simplified backtraces
  last_warnings()

  # Full backtraces
  summary(last_warnings())
}
}
