% Generated by roxygen2 (4.0.1): do not edit by hand
\name{lazy_dots}
\alias{lazy_dots}
\title{Capture ... (dots) for later lazy evaluation.}
\usage{
lazy_dots(...)
}
\arguments{
\item{...}{Dots from another function}
}
\value{
A named list of \code{\link{lazy}} expressions.
}
\description{
Capture ... (dots) for later lazy evaluation.
}
\examples{
lazy_dots(x = 1)
lazy_dots(a, b, c * 4)

f <- function(x = a + b, ...) {
  lazy_dots(x = x, y = a + b, ...)
}
f(z = a + b)
}

