% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle.R
\name{lifecycle}
\alias{lifecycle}
\title{Life cycle of the rlang package}
\description{
The rlang package is currently maturing. Unless otherwise stated,
this applies to all its exported functions. Maturing functions are
susceptible to API changes. Only use these in packages if you're
prepared to make changes as the package evolves. See sections below
for a list of functions marked as stable.

The documentation pages of retired functions contain life cycle
sections that explain the reasons for their retirements.
}
\section{Stable functions}{


TODO: Review all functions in rlang and mark stable ones as such.
\itemize{
\item \code{\link[=eval_tidy]{eval_tidy()}}
\item \link{!!}, \link{!!!}
\item \code{\link[=enquo]{enquo()}}, \code{\link[=quo]{quo()}}, \code{\link[=quos]{quos()}}
\item \code{\link[=enexpr]{enexpr()}}, \code{\link[=expr]{expr()}}, \code{\link[=exprs]{exprs()}}
\item \code{\link[=sym]{sym()}}, \code{\link[=syms]{syms()}}
\item \code{\link[=new_quosure]{new_quosure()}}, \code{\link[=is_quosure]{is_quosure()}}
\item \code{\link[=quo_get_expr]{quo_get_expr()}}, \code{\link[=quo_set_expr]{quo_set_expr()}}
\item \code{\link[=quo_get_env]{quo_get_env()}}, \code{\link[=quo_set_env]{quo_set_env()}}
\item \code{\link[=set_names]{set_names()}}, \code{\link[=names2]{names2()}}
\item \code{\link[=as_function]{as_function()}}
}
}

\section{Experimental functions}{

\itemize{
\item \code{\link[=type_of]{type_of()}}, \code{\link[=switch_type]{switch_type()}}, \code{\link[=coerce_type]{coerce_type()}}
\item \code{\link[=switch_class]{switch_class()}}, \code{\link[=coerce_class]{coerce_class()}}
\item \code{\link[=lang_type_of]{lang_type_of()}}, \code{\link[=switch_lang]{switch_lang()}}, \code{\link[=coerce_lang]{coerce_lang()}}
\item \code{\link[=set_attrs]{set_attrs()}}, \code{\link[=mut_attrs]{mut_attrs()}}
}
}

\section{Soft-deprecated functions and arguments}{


\strong{rlang 0.2.0:}
\itemize{
\item \code{\link[=UQ]{UQ()}}
\item \code{\link[=UQS]{UQS()}}
\item \code{\link[=parse_quosure]{parse_quosure()}}
\item \code{\link[=parse_quosures]{parse_quosures()}}
\item \code{\link[=quo_expr]{quo_expr()}} => \code{\link[=quo_squash]{quo_squash()}}
\item \code{\link[=lang]{lang()}} => \code{\link[=call2]{call2()}}
\item \code{\link[=new_language]{new_language()}} => \code{\link[=new_call]{new_call()}}
\item \code{\link[=is_lang]{is_lang()}} => \code{\link[=is_call]{is_call()}}
\item \code{\link[=is_unary_lang]{is_unary_lang()}} => \code{\link[=is_unary_call]{is_unary_call()}}
\item \code{\link[=is_binary_lang]{is_binary_lang()}} => \code{\link[=is_binary_call]{is_binary_call()}}
}
}

\section{Deprecated functions and arguments}{


\strong{rlang 0.2.0:}
\itemize{
\item \code{\link[=UQE]{UQE()}}
\item \code{\link[=is_quosureish]{is_quosureish()}}, \code{\link[=as_quosureish]{as_quosureish()}}
}
}

\section{Defunct functions and arguments}{


\strong{rlang 0.2.0:}
\itemize{
\item \link[=quasiquotation]{:=}
}
}

