% Generated by roxygen2 (4.0.2): do not edit by hand
\name{make_call}
\alias{make_call}
\title{Make a call with \code{lazy_dots} as arguments.}
\usage{
make_call(fun, dots)
}
\arguments{
\item{fun}{Function as symbol or quoted call.}

\item{dots}{Arguments to function; must be a \code{lazy_dots} object.}
}
\value{
A list:
  \item{env}{The common environment for all elements}
  \item{expr}{The expression}
}
\description{
In order to make a valid call, the environment must be the same for
all of the dots.
}
\examples{
make_call(quote(f), lazy_dots(x = 1, 2))
}

