% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attr.R
\name{names2}
\alias{names2}
\title{Get names of a vector}
\usage{
names2(x)
}
\arguments{
\item{x}{A vector.}
}
\description{
\Sexpr[results=rd, stage=render]{rlang:::lifecycle("stable")}

This names getter always returns a character vector, even when an
object does not have a \code{names} attribute. In this case, it returns
a vector of empty names \code{""}. It also standardises missing names to
\code{""}.
}
\section{Life cycle}{


\code{names2()} is stable.
}

\examples{
names2(letters)

# It also takes care of standardising missing names:
x <- set_names(1:3, c("a", NA, "b"))
names2(x)
}
