% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec-new.R
\name{new-vector}
\alias{new-vector}
\alias{new_logical}
\alias{new_integer}
\alias{new_double}
\alias{new_character}
\alias{new_complex}
\alias{new_raw}
\alias{new_list}
\title{Create vectors matching a given length}
\usage{
new_logical(n, names = NULL)

new_integer(n, names = NULL)

new_double(n, names = NULL)

new_character(n, names = NULL)

new_complex(n, names = NULL)

new_raw(n, names = NULL)

new_list(n, names = NULL)
}
\arguments{
\item{n}{The vector length.}

\item{names}{Names for the new vector.}
}
\description{
\Sexpr[results=rd, stage=render]{rlang:::lifecycle("questioning")}

These functions construct vectors of a given length, with attributes
specified via dots. Except for \code{new_list()} and \code{new_raw()}, the
empty vectors are filled with typed \link{missing} values. This is in
contrast to the base function \code{\link[base:vector]{base::vector()}} which creates
zero-filled vectors.
}
\section{Lifecycle}{


These functions are likely to be replaced by a vctrs equivalent in
the future. They are in the questioning lifecycle stage.
}

\examples{
new_list(10)
new_logical(10)
}
\seealso{
rep_along
}
