% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{new-vectors}
\alias{character_until}
\alias{complex_until}
\alias{double_until}
\alias{integer_until}
\alias{list_until}
\alias{logical_until}
\alias{new-vectors}
\alias{raw_until}
\title{Create new vectors.}
\usage{
logical_until(.n = 0, ..., .attrs = list())

integer_until(.n = 0, ..., .attrs = list())

double_until(.n = 0, ..., .attrs = list())

character_until(.n = 0, ..., .attrs = list())

list_until(.n = 0, ..., .attrs = list())

complex_until(.n = 0, ..., .attrs = list())

raw_until(.n = 0, ..., .attrs = list())
}
\arguments{
\item{.n}{The vector length.}

\item{...}{A list of named attributes.}

\item{.attrs}{A list of named attributes.}
}
\description{
These functions construct vectors of given length, with attributes
specified via dots.
}
\examples{
list_until(10)

# Add attributes, including the S3 class:
integer_until(0, index = 1)
double_until(10, class = "my_class")
}
\seealso{
along
}

