% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.R
\name{new-vectors}
\alias{new-vectors}
\alias{new_lgl}
\alias{new_int}
\alias{new_dbl}
\alias{new_chr}
\alias{new_lst}
\alias{new_cpl}
\alias{new_raw}
\title{Create new vectors.}
\usage{
new_lgl(.n = 0, ..., .attrs = list())

new_int(.n = 0, ..., .attrs = list())

new_dbl(.n = 0, ..., .attrs = list())

new_chr(.n = 0, ..., .attrs = list())

new_lst(.n = 0, ..., .attrs = list())

new_cpl(.n = 0, ..., .attrs = list())

new_raw(.n = 0, ..., .attrs = list())
}
\arguments{
\item{.n}{The vector length.}

\item{...}{A list of named attributes.}

\item{.attrs}{A list of named attributes.}
}
\description{
These functions construct vectors of given length, with attributes
specified via dots. Except for \code{new_lst()} and \code{new_raw()}, the
empty vectors are filled with typed \link{missing} values. This is in
contrast to the base function \code{\link[base:vector]{base::vector()}} which creates
zero-filled vectors.
}
\examples{
new_lst(10)

# Add attributes, including the S3 class:
new_int(0, index = 1)
new_dbl(10, class = "my_class")
}
\seealso{
along
}
