% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new.R
\name{new-vectors}
\alias{new-vectors}
\alias{new_character}
\alias{new_complex}
\alias{new_double}
\alias{new_integer}
\alias{new_list}
\alias{new_logical}
\alias{new_raw}
\title{Create new vectors.}
\usage{
new_logical(.n = 0, ...)

new_integer(.n = 0, ...)

new_double(.n = 0, ...)

new_character(.n = 0, ...)

new_list(.n = 0, ...)

new_complex(.n = 0, ...)

new_raw(.n = 0, ...)
}
\arguments{
\item{.n}{The vector length.}

\item{...}{Optional attributes for the vector. These are passed to
\code{\link[base]{structure}()}.}
}
\description{
These functions construct vectors of given length, with attributes
specified via dots.
}
\examples{
new_list(10)

# Add attributes, including the S3 class:
new_integer(0, index = 1)
new_double(10, class = "my_class")
}
\seealso{
along
}

