% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector-ctor.R
\name{new-vectors}
\alias{new-vectors}
\alias{lgl_len}
\alias{int_len}
\alias{dbl_len}
\alias{chr_len}
\alias{cpl_len}
\alias{bytes_len}
\alias{list_len}
\title{Create new vectors.}
\usage{
lgl_len(.n = 0, ..., .attrs = list())

int_len(.n = 0, ..., .attrs = list())

dbl_len(.n = 0, ..., .attrs = list())

chr_len(.n = 0, ..., .attrs = list())

cpl_len(.n = 0, ..., .attrs = list())

bytes_len(.n = 0, ..., .attrs = list())

list_len(.n = 0, ..., .attrs = list())
}
\arguments{
\item{.n}{The vector length.}

\item{...}{A list of named attributes.}

\item{.attrs}{A list of named attributes.}
}
\description{
These functions construct vectors of given length, with attributes
specified via dots. Except for \code{list_len()} and \code{bytes_len()}, the
empty vectors are filled with typed \link{missing} values. This is in
contrast to the base function \code{\link[base:vector]{base::vector()}} which creates
zero-filled vectors.
}
\examples{
list_len(10)

# Add attributes, including the S3 class:
int_len(0, index = 1)
dbl_len(10, class = "my_class")
}
\seealso{
along
}
