% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lang-call.R
\name{new_call}
\alias{new_call}
\title{Create a call by "hand"}
\usage{
new_call(.fn, ..., .args = list())
}
\arguments{
\item{.fn}{Function to call. For \code{make_call}, either a string,
a symbol or a quoted call. For \code{do_call}, a bare function
name or call.}

\item{..., .args}{Arguments to the call either in or out of a list}
}
\description{
Create a call by "hand"
}
\examples{
# fn can either be a string, a symbol or a call
new_call("f", a = 1)
new_call(quote(f), a = 1)
new_call(quote(f()), a = 1)

#' Can supply arguments individually or in a list
new_call(quote(f), a = 1, b = 2)
new_call(quote(f), .args = list(a = 1, b = 2))
}
\seealso{
call_modify
}
