% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.R
\name{new_f}
\alias{new_f}
\title{Create a formula object by "hand".}
\usage{
new_f(rhs, lhs = NULL, env = parent.frame())
}
\arguments{
\item{lhs, rhs}{A call, name, or atomic vector.}

\item{env}{An environment}
}
\value{
A formula object
}
\description{
Create a formula object by "hand".
}
\examples{
new_f(quote(a))
new_f(quote(a), quote(b))
}

