% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.R
\name{new_formula}
\alias{new_formula}
\title{Create a formula}
\usage{
new_formula(lhs, rhs, env = caller_env())
}
\arguments{
\item{lhs, rhs}{A call, name, or atomic vector.}

\item{env}{An environment.}
}
\value{
A formula object.
}
\description{
Create a formula
}
\examples{
new_formula(quote(a), quote(b))
new_formula(NULL, quote(b))
}
\seealso{
\code{\link[=new_quosure]{new_quosure()}}
}
