% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quos.R
\name{new_quosures}
\alias{new_quosures}
\title{Create a list of quosures}
\usage{
new_quosures(x)
}
\arguments{
\item{x}{A list containing quosures}
}
\description{
This takes a list and adds the \code{quosures} class. This ensures that
quosures contained in the list are properly printed. This is mostly
meant as a workaround to a bug in the current R printer which
causes quosures wrapped in lists to evaluate early during printing.
}
