% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weakref.R
\name{new_weakref}
\alias{new_weakref}
\title{Create a weak reference}
\usage{
new_weakref(key, value = NULL, finalizer = NULL, on_quit = FALSE)
}
\arguments{
\item{key}{The key for the weak reference. Must be a reference object -- that
is, an environment or external pointer.}

\item{value}{The value for the weak reference. This can be \code{NULL}, if you
want to use the weak reference like a weak pointer.}

\item{finalizer}{A function that is run after the key becomes unreachable.}

\item{on_quit}{Should the finalizer be run when R exits?}
}
\description{
A weak reference is a special R object which makes it possible to keep a
reference to an object without preventing garbage collection of that object.
It can also be used to keep data about an object without preventing GC of the
object, similar to WeakMaps in JavaScript.

Objects in R are considered \emph{reachable} if they can be accessed by following
a chain of references, starting from a \emph{root node}; root nodes are
specially-designated R objects, and include the global environment and base
environment. As long as the key is reachable, the value will not be garbage
collected. This is true even if the weak reference object becomes
unreachable. The key effectively prevents the weak reference and its value
from being collected, according to the following chain of ownership:
\code{weakref <- key -> value}.

When the key becomes unreachable, the key and value in the weak reference
object are replaced by \code{NULL}, and the finalizer is scheduled to execute.
}
\examples{
e <- env()

# Create a weak reference to e
w <- new_weakref(e, finalizer = function(e) message("finalized"))

# Get the key object from the weak reference
identical(wref_key(w), e)

# When the regular reference (the `e` binding) is removed and a GC occurs,
# the weak reference will not keep the object alive.
rm(e)
gc()
identical(wref_key(w), NULL)


# A weak reference with a key and value. The value contains data about the
# key.
k <- env()
v <- list(1, 2, 3)
w <- new_weakref(k, v)

identical(wref_key(w), k)
identical(wref_value(w), v)

# When v is removed, the weak ref keeps it alive because k is still reachable.
rm(v)
gc()
identical(wref_value(w), list(1, 2, 3))

# When k is removed, the weak ref does not keep k or v alive.
rm(k)
gc()
identical(wref_key(w), NULL)
identical(wref_value(w), NULL)
}
\seealso{
\code{\link[=is_weakref]{is_weakref()}}, \code{\link[=wref_key]{wref_key()}} and \code{\link[=wref_value]{wref_value()}}.
}
\keyword{experimental}
