% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env-special.R
\name{ns_env}
\alias{ns_env}
\alias{ns_imports_env}
\alias{ns_env_name}
\title{Get the namespace of a package}
\usage{
ns_env(x = NULL)

ns_imports_env(x = NULL)

ns_env_name(x = NULL)
}
\arguments{
\item{x}{\itemize{
\item For \code{ns_env()}, the name of a package or an environment as a
string.
\itemize{
\item An environment.
\item A function. In that case, the environment of that function is
searched for a namespace with \code{\link[base:topenv]{base::topenv()}}.
}
}}

\item{env}{A namespace environment.}
}
\description{
Namespaces are the environment where all the functions of a package
live. The parent environments of namespaces are the \code{imports}
environments, which contain all the functions imported from other
packages.
}
\section{Life cycle}{


These functions are experimental and may not belong to the rlang
package. Expect API changes.
}

\seealso{
\code{\link[=pkg_env]{pkg_env()}}
}
\keyword{internal}
