% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env-special.R
\name{ns_registry_env}
\alias{ns_registry_env}
\title{Return the namespace registry env}
\usage{
ns_registry_env()
}
\description{
Note that the namespace registry does not behave like a normal
environment because the parent is \code{NULL} instead of the empty
environment. This is exported for expert usage in development tools
only.
}
\keyword{internal}
