% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle-retired.R
\name{parse_quosure}
\alias{parse_quosure}
\alias{parse_quosures}
\title{Parse text into a quosure}
\usage{
parse_quosure(x, env = caller_env())

parse_quosures(x, env = caller_env())
}
\arguments{
\item{x}{Text containing expressions to parse_expr for
\code{parse_expr()} and \code{parse_exprs()}. Can also be an R connection,
for instance to a file. If the supplied connection is not open,
it will be automatically closed and destroyed.}
}
\description{
\Sexpr[results=rd, stage=render]{rlang:::lifecycle("deprecated")}

These functions were deprecated and renamed to \code{\link[=parse_quo]{parse_quo()}}
and \code{\link[=parse_quos]{parse_quos()}} in rlang 0.2.0. This is for consistency with the
convention that suffixes indicating return types are not
abbreviated.
}
\keyword{internal}
