% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle-retired.R
\name{prepend}
\alias{prepend}
\alias{modify}
\title{Prepend a vector}
\usage{
prepend(x, values, before = 1)

modify(.x, ...)
}
\arguments{
\item{x}{the vector to be modified.}

\item{values}{to be included in the modified vector.}

\item{before}{a subscript, before which the values are to be appended.}

\item{.x}{A vector to modify.}

\item{...}{<\link[=dyn-dots]{dynamic}> List of elements to merge into
\code{.x}. Named elements already existing in \code{.x} are used as
replacements. Elements that have new or no names are inserted at
the end.}
}
\description{
\Sexpr[results=rd, stage=render]{rlang:::lifecycle("deprecated")}

Vector functions are now out of scope for rlang. They might be
revived in the vctrs or funs packages.
}
\keyword{internal}
