% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{scalar-type-predicates}
\alias{scalar-type-predicates}
\alias{is_scalar_list}
\alias{is_scalar_atomic}
\alias{is_scalar_vector}
\alias{is_scalar_integer}
\alias{is_scalar_double}
\alias{is_scalar_character}
\alias{is_scalar_logical}
\alias{is_scalar_raw}
\alias{is_string}
\alias{is_scalar_bytes}
\title{Scalar type predicates}
\usage{
is_scalar_list(x)

is_scalar_atomic(x)

is_scalar_vector(x)

is_scalar_integer(x)

is_scalar_double(x)

is_scalar_character(x, encoding = NULL)

is_scalar_logical(x)

is_scalar_raw(x)

is_string(x, encoding = NULL)

is_scalar_bytes(x)
}
\arguments{
\item{x}{object to be tested.}

\item{encoding}{Expected encoding of a string or character
vector. One of \code{UTF-8}, \code{latin1}, or \code{unknown}.}
}
\description{
These predicates check for a given type and whether the vector is
"scalar", that is, of length 1.
}
\seealso{
\link{type-predicates}, \link{bare-type-predicates}
}
