% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/env-special.R, R/lifecycle-retired.R
\name{pkg_env}
\alias{pkg_env}
\alias{pkg_env_name}
\alias{scoped_names}
\alias{scoped_envs}
\alias{is_attached}
\alias{base_env}
\alias{global_env}
\alias{scoped_env}
\alias{is_scoped}
\title{Retired \code{scoped} functions}
\usage{
pkg_env(pkg)

pkg_env_name(pkg)

scoped_names()

scoped_envs()

is_attached(x)

base_env()

global_env()

scoped_env(nm)

is_scoped(nm)
}
\arguments{
\item{pkg}{The name of a package.}

\item{nm}{The name of an environment attached to the search
path. Call \code{\link[base:search]{base::search()}} to see what is currently on the path.}
}
\description{
\Sexpr[results=rd, stage=render]{rlang:::lifecycle("soft-deprecated")}

These functions are soft-deprecated as of rlang 0.3.0. They are
replaced by \code{\link[=is_attached]{is_attached()}}, ...
}
\keyword{internal}
