% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle-deprecated.R
\name{scoped_env}
\alias{scoped_env}
\alias{is_scoped}
\title{Deprecated \code{scoped} functions}
\usage{
scoped_env(nm)

is_scoped(nm)
}
\arguments{
\item{nm}{The name of an environment attached to the search
path. Call \code{\link[base:search]{base::search()}} to see what is currently on the path.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

These functions are deprecated as of rlang 0.3.0. Please use
\code{\link[=is_attached]{is_attached()}} instead.
}
\keyword{internal}
