% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle-deprecated.R
\name{scoped_interactive}
\alias{scoped_interactive}
\alias{scoped_options}
\alias{scoped_bindings}
\title{Deprecated \code{scoped_} functions}
\usage{
scoped_interactive(value = TRUE, frame = caller_env())

scoped_options(..., .frame = caller_env())

scoped_bindings(..., .env = .frame, .frame = caller_env())
}
\arguments{
\item{value}{A single \code{TRUE} or \code{FALSE}. This overrides the return
value of \code{is_interactive()}.}

\item{frame, .frame}{The environment of a running function which defines
the scope of the temporary options. When the function returns,
the options are reset to their original values.}

\item{...}{For \code{local_options()} and \code{push_options()}, named
values defining new option values. For \code{peek_options()}, strings
or character vectors of option names.}

\item{.env}{An environment.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Deprecated as of rlang 0.4.2. Use \code{\link[=local_interactive]{local_interactive()}},
\code{\link[=local_options]{local_options()}}, or \code{\link[=local_bindings]{local_bindings()}} instead.
}
\keyword{internal}
