% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dots.R
\name{splice}
\alias{splice}
\alias{is_spliced}
\alias{is_spliced_bare}
\title{Splice a list within a vector}
\usage{
splice(x)

is_spliced(x)

is_spliced_bare(x)
}
\arguments{
\item{x}{A list to splice.}
}
\description{
This function marks an object to be spliced. It is equivalent to
using \code{!!!} in a function with \link[=dots_list]{tidy dots semantics}.
}
\keyword{internal}
