% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector-ctor.R
\name{splice}
\alias{splice}
\title{Construct a list with splicing.}
\usage{
splice(..., .bare = TRUE)
}
\arguments{
\item{...}{Components of the new vector. Explicitly spliced lists
are always spliced. Bare lists are spliced depending on the value
of \code{.bare}.}

\item{.bare}{Whether to splice bare lists. If \code{FALSE}, only lists
inheriting from \code{"spliced"} are spliced.}
}
\description{
Construct a list with splicing.
}
\examples{
# splice() is like the atomic vector constructors but for lists:
dbl(1, list(1, 2))
splice(1, list(1, 2))

# Only bare lists are spliced. Objects like data frames are not spliced:
splice(1, mtcars)

# Use the spliced() adjective to splice objects:
splice(1, spliced(mtcars))

# You can chose not to splice bare lists with `.bare`:
splice(list(1, 2), .bare = FALSE)

# Note that explicitly spliced lists are always spliced:
splice(spliced(list(1, 2)), .bare = FALSE)
}
