% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval-tidy.R
\docType{data}
\name{tidyeval-data}
\alias{tidyeval-data}
\alias{.data}
\title{Data pronoun for tidy evaluation}
\format{An object of class \code{dictionary} of length 0.}
\usage{
.data
}
\description{
This pronoun is installed by functions performing \link[=eval_tidy]{tidy
evaluation}. It allows you to refer to overscoped data
explicitly.
}
\details{
You can import this object in your package namespace to avoid \code{R CMD check} errors when referring to overscoped objects.
}
\examples{
quo <- quo(.data$foo)
eval_tidy(quo, list(foo = "bar"))
}
\keyword{datasets}
