% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector-ctor.R
\name{vector-ctors}
\alias{vector-ctors}
\alias{lgl}
\alias{int}
\alias{dbl}
\alias{cpl}
\alias{mk_chr}
\alias{mk_raw}
\alias{mk_list}
\title{Construct new vectors.}
\usage{
lgl(...)

int(...)

dbl(...)

cpl(...)

mk_chr(...)

mk_raw(...)

mk_list(...)
}
\arguments{
\item{...}{Components of the new vector. Bare lists and explicitly
spliced lists are spliced.}
}
\description{
Construct new vectors.
}
