% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector-ctor.R
\name{vector-len}
\alias{vector-len}
\alias{lgl_len}
\alias{int_len}
\alias{dbl_len}
\alias{chr_len}
\alias{cpl_len}
\alias{bytes_len}
\alias{list_len}
\title{Create new vectors.}
\usage{
lgl_len(.n = 0, ...)

int_len(.n = 0, ...)

dbl_len(.n = 0, ...)

chr_len(.n = 0, ...)

cpl_len(.n = 0, ...)

bytes_len(.n = 0, ...)

list_len(.n = 0, ...)
}
\arguments{
\item{.n}{The vector length.}

\item{...}{A list of named attributes. These have \link[=dots_list]{explicit
splicing semantics}. Pass a single unnamed \code{NULL} to
zap all attributes from \code{.x}.}
}
\description{
These functions construct vectors of given length, with attributes
specified via dots. Except for \code{list_len()} and \code{bytes_len()}, the
empty vectors are filled with typed \link{missing} values. This is in
contrast to the base function \code{\link[base:vector]{base::vector()}} which creates
zero-filled vectors.
}
\examples{
list_len(10)

# Add attributes, including the S3 class:
int_len(0, index = 1)
dbl_len(10, class = "my_class")
}
\seealso{
vector-along
}
