% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector-ctor.R
\name{vector-len}
\alias{vector-len}
\alias{lgl_len}
\alias{int_len}
\alias{dbl_len}
\alias{chr_len}
\alias{cpl_len}
\alias{raw_len}
\alias{bytes_len}
\alias{list_len}
\title{Create vectors matching a given length}
\usage{
lgl_len(.n)

int_len(.n)

dbl_len(.n)

chr_len(.n)

cpl_len(.n)

raw_len(.n)

bytes_len(.n)

list_len(.n)
}
\arguments{
\item{.n}{The vector length.}
}
\description{
These functions construct vectors of given length, with attributes
specified via dots. Except for \code{list_len()} and \code{bytes_len()}, the
empty vectors are filled with typed \link{missing} values. This is in
contrast to the base function \code{\link[base:vector]{base::vector()}} which creates
zero-filled vectors.
}
\examples{
list_len(10)
lgl_len(10)
}
\seealso{
vector-along
}
