% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lifecycle-retired.R
\name{vector-old-ctors}
\alias{vector-old-ctors}
\alias{lgl_len}
\alias{int_len}
\alias{dbl_len}
\alias{chr_len}
\alias{cpl_len}
\alias{raw_len}
\alias{bytes_len}
\alias{list_len}
\alias{lgl_along}
\alias{int_along}
\alias{dbl_along}
\alias{chr_along}
\alias{cpl_along}
\alias{raw_along}
\alias{bytes_along}
\alias{list_along}
\alias{node}
\title{Retired vector construction by length}
\usage{
lgl_len(.n)

int_len(.n)

dbl_len(.n)

chr_len(.n)

cpl_len(.n)

raw_len(.n)

bytes_len(.n)

list_len(.n)

lgl_along(.x)

int_along(.x)

dbl_along(.x)

chr_along(.x)

cpl_along(.x)

raw_along(.x)

bytes_along(.x)

list_along(.x)

node(car, cdr = NULL)
}
\arguments{
\item{.x}{A vector.}
}
\description{
These functions were soft-deprecated and renamed with \code{new_} prefix
in rlang 0.2.0. This is for consistency with other non-variadic
object constructors.
}
\keyword{internal}
